### 2025-06-25 #8443 @chrissoria
> **License info**:
>
> 🟡 License found: `Other` ([Check here](https://opensource.org/licenses) for OSI approval)

Resolved
