### 2025-07-15 #8443 @chrissoria
Hello @openjournals/joss-eics,

I hope this message finds you well. I submitted this paper for review on June 8, 2025, and wanted to politely follow up on the current status of my submission.

I understand that JOSS has been experiencing increased submission volumes and processing times, and I greatly appreciate the volunteer work that goes into managing the review process. I'm simply checking to see if there might be an estimated timeline for editor assignment or if there's anything I can do to help move the process forward.

Thank you for your time and consideration. I look forward to hearing from you when convenient.

Best regards,
Chrus
