### 2025-07-30 #8443 @samhforbes
Hi @chrissoria we do have a backlog at the moment and unfortunately that makes it tricky to assign. We are trying to onboard and recruit more editors but I will come back as soon as I have an update for you.
