### 2025-08-08 c9f4696
updating function so that image inputs that are file paths do not crash the function, cleaning up the cerad score calculation and prompts.
