+++ src/catllm/CERAD_functions.py
@@ -67,8 +67,8 @@ def cerad_drawn_score(
         categories = ["The image contains a drawing that clearly represents overlapping rectangles",
                     "The image does NOT contain any drawing that resembles overlapping rectangles",
                     "The image contains a drawing that resembles overlapping rectangles",
-                    "If rectangle 1 is present it has 4 sides",
-                    "If rectablge 2 is present it has 4 sides",
+                    "If rectangle 1 is present and it has 4 sides",
+                    "If rectangle 2 is present and it has 4 sides",
                     "The drawn rectangles are overlapping",
                     "The drawn rectangles overlap to form a longer vertical rectangle with top and bottom sticking out",
                     "None of the above descriptions apply"]
@@ -131,19 +131,30 @@ def cerad_drawn_score(
             continue  # Skip the rest of the loop iteration
         
     # Only open the file if path is valid
-        with open(img_path, "rb") as f:
-            encoded = base64.b64encode(f.read()).decode("utf-8")
-    
+        if os.path.isdir(img_path):
+            encoded = "Not a Valid Image, contains file path"
+        else:
+            try:
+                with open(img_path, "rb") as f:
+                    encoded = base64.b64encode(f.read()).decode("utf-8")
+            except Exception as e:
+                    encoded = f"Error: {str(e)}"
     # Handle extension safely
-        ext = Path(img_path).suffix.lstrip(".").lower()
-        encoded_image = f"data:image/{ext};base64,{encoded}"
+        if encoded.startswith("Error:") or encoded == "Not a Valid Image, contains file path":
+            encoded_image = encoded
+            valid_image = False
+            
+        else:
+            ext = Path(img_path).suffix.lstrip(".").lower()
+            encoded_image = f"data:image/{ext};base64,{encoded}"
+            valid_image = True
 
         if reference_in_image:
             reference_text = f"This image contains a perfect reference image of a {shape}. Next to is a drawing that is meant to be similar to the reference {shape}.\n\n"
         else:
             reference_text = f"Image is expected to show within it a drawing of a {shape}.\n\n"
 
-        if model_source == "OpenAI":
+        if model_source == "OpenAI" and valid_image:
             prompt = [
                 {
                     "type": "text",
@@ -154,8 +165,8 @@ def cerad_drawn_score(
                         f"{reference_text}"
                         f"Categories:\n{categories_str}\n\n"
                         f"Output format ► Respond with **only** a JSON object whose keys are the "
-                        f"quoted category numbers ('1', '2', …) and whose values are 1 or 0. "
-                        f"No additional keys, comments, or text.\n\n"
+                        f"quoted category numbers ('1', '2', …) and whose values are 1 if present or 0 if not present. "
+                        f"No additional keys, comments, numbers beyond 0 or 1, or text.\n\n"
                         f"Example:\n"
                         f"{example_JSON}"
                         )
@@ -173,7 +184,7 @@ def cerad_drawn_score(
                 "image_url": {"url": encoded_image, "detail": "high"}
             })
         
-        elif model_source == "Anthropic":
+        elif model_source == "Anthropic" and valid_image:
             prompt = [
                 {
                     "type": "text",
@@ -184,8 +195,8 @@ def cerad_drawn_score(
                         f"{reference_text}"
                         f"Categories:\n{categories_str}\n\n"
                         f"Output format ► Respond with **only** a JSON object whose keys are the "
-                        f"quoted category numbers ('1', '2', …) and whose values are 1 or 0. "
[…]
