### 2025-08-08 84e90a4
inputted error message if the model is not a valid model
