+++ src/catllm/image_functions.py
@@ -148,8 +148,11 @@ def image_multi_class(
                 reply = response_obj.choices[0].message.content
                 link1.append(reply)
             except Exception as e:
-                print(f"An error occurred: {e}")
-                link1.append(f"Error processing input: {e}")
+                if "model" in str(e).lower():
+                    raise ValueError(f"Invalid OpenAI model '{user_model}': {e}")
+                else:
+                    print("An error occurred: {e}")
+                    link1.append("Error processing input: {e}")
 
         elif model_source == "Anthropic":
             import anthropic
@@ -165,8 +168,11 @@ def image_multi_class(
                 reply = message.content[0].text
                 link1.append(reply)
             except Exception as e:
-                print(f"An error occurred: {e}")
-                link1.append(f"Error processing input: {e}")
+                if "model" in str(e).lower():
+                    raise ValueError(f"Invalid OpenAI model '{user_model}': {e}")
+                else:
+                    print("An error occurred: {e}")
+                    link1.append("Error processing input: {e}")
 
         elif model_source == "Mistral":
             from mistralai import Mistral
@@ -182,8 +188,11 @@ def image_multi_class(
                 reply = response.choices[0].message.content
                 link1.append(reply)
             except Exception as e:
-                print(f"An error occurred: {e}")
-                link1.append(f"Error processing input: {e}")
+                if "model" in str(e).lower():
+                    raise ValueError(f"Invalid OpenAI model '{user_model}': {e}")
+                else:
+                    print("An error occurred: {e}")
+                    link1.append("Error processing input: {e}")
         #if no valid image path is provided
         elif  valid_image == False:
             reply = "invalid image path"
@@ -436,8 +445,11 @@ def image_score_drawing(
                 reply = response_obj.choices[0].message.content
                 link1.append(reply)
             except Exception as e:
-                print(f"An error occurred: {e}")
-                link1.append(f"Error processing input: {e}")
+                if "model" in str(e).lower():
+                    raise ValueError(f"Invalid OpenAI model '{user_model}': {e}")
+                else:
+                    print("An error occurred: {e}")
+                    link1.append("Error processing input: {e}")
 
         elif model_source == "Anthropic":
             import anthropic
@@ -452,8 +464,11 @@ def image_score_drawing(
                 reply = message.content[0].text  # Anthropic returns content as list
                 link1.append(reply)
             except Exception as e:
-                print(f"An error occurred: {e}")
-                link1.append(f"Error processing input: {e}")
+                if "model" in str(e).lower():
+                    raise ValueError(f"Invalid OpenAI model '{user_model}': {e}")
+                else:
+                    print("An error occurred: {e}")
+                    link1.append("Error processing input: {e}")
 
         elif model_source == "Mistral":
             from mistralai import Mistral
@@ -469,8 +484,11 @@ def image_score_drawing(
                 reply = response.choices[0].message.content
                 link1.append(reply)
             except Exception as e:
-                print(f"An error occurred: {e}")
-                link1.append(f"Error processing input: {e}")
+                if "model" in str(e).lower():
+                    raise ValueError(f"Invalid OpenAI model '{user_model}': {e}")
+                else:
+                    print("An error occurred: {e}")
+                    link1.append("Error processing input: {e}")
         #if no valid image path is provided
         elif  valid_image == False:
             reply = "invalid image path"
@@ -567,10 +585,6 @@ def image_features(
     import base64
     from pathlib import Path
 
-    if save_directory is not None and not os.path.isdir(save_directory):
[…]
