### 2025-09-18 7301be2
adding a web search build dataset function
