### 2025-09-19 b23e799
updating readme to have latest function
