### 2025-09-19 743722d
updating to allow for google in multi class text
