### 2025-09-19 428cc3c
changing the column names so they are easier to understand
