+++ src/catllm/image_functions.py
@@ -225,7 +225,7 @@ def image_multi_class(
             # Save progress so far
             temp_df = pd.DataFrame({
                 'image_input': image_files[:i+1],
-                'link1': link1,
+                'model_response': link1,
                 'json': extracted_jsons
             })
             # Normalize processed jsons so far
@@ -522,7 +522,7 @@ def image_score_drawing(
             # Save progress so far
             temp_df = pd.DataFrame({
                 'image_input': image_files[:i+1],
-                'link1': link1,
+                'model_response': link1,
                 'json': extracted_jsons
             })
             # Normalize processed jsons so far
@@ -844,7 +844,7 @@ def image_features(
             image_files.reset_index(drop=True) if isinstance(image_files, (pd.DataFrame, pd.Series)) 
             else pd.Series(image_files)
         ),
-        'link1': pd.Series(link1).reset_index(drop=True),
+        'model_response': pd.Series(link1).reset_index(drop=True),
         'json': pd.Series(extracted_jsons).reset_index(drop=True)
     })
     categorized_data = pd.concat([categorized_data, normalized_data], axis=1)
