+++ pyproject.toml
@@ -28,7 +28,8 @@ dependencies = [
   "pandas",
   "tqdm",
   "requests",
-  "openai"
+  "openai",
+  "random"
 ]
 
 [project.urls]
