### 2025-09-19 f7fd0d9
added an example script for categorizing text data
