+++ examples/__init__.py
@@ -1,14 +1,0 @@
-# SPDX-FileCopyrightText: 2025-present Christopher Soria <chrissoria@berkeley.edu>
-#
-# SPDX-License-Identifier: MIT
-
-from .__about__ import (
-    __version__,
-    __author__,
-    __description__,
-    __title__,
-    __url__,
-    __license__,
-)
-
-from .catllm import *  # Import functions from cat_llm.py
