### 2025-09-19 89640da
added google search abilities
