### 2025-09-29 e0a6849
updated package so that inputting creativity is not necessary
