### 2025-10-08 77fd8e1
updating the web search method so it doesn't halt on rate limit, requiring anthropic package, removing case sensitivity for model source input
