+++ academic_examples/paper.md
@@ -31,13 +31,16 @@ new opportunities for automated text and image analysis in social
 science research [@schulze_buschoff_visual_2025; @yang_large_2024;
 @sachdeva_normative_2025]. Researchers increasingly use these tools to
 code open-ended survey responses, categorize qualitative data, and
-analyze visual content at scale. However, challenges remain due to
+analyze visual content at scale. Yet challenges persist due to
 inconsistent output formats, diverse API interfaces, and the lack of
-standardized workflows for integrating model outputs into traditional
-statistical analysis pipelines [@rossi_problems_2024]. CatLLM addresses
-these issues by providing a modular framework with specialized functions
-that enable researchers to work with consistent data structures across
-text and image analysis workflows while maintaining compatibility with
+standardized workflows for integrating both model outputs and external
+data sources into traditional statistical analysis pipelines
+[@rossi_problems_2024]. CatLLM addresses these issues by providing a
+modular framework with specialized functions that not only ensure
+consistent data structures across text and image analysis workflows, but
+also facilitate the automated retrieval of structured data from the web.
+This integration allows researchers to seamlessly combine large model
+outputs with real-world datasets, maintaining compatibility with
 standard statistical analysis tools.
 
 # Statement of need
@@ -68,18 +71,19 @@ begun using large language models (LLMs) directly through web
 interfaces, this approach lacks standardization, reproducibility, and
 systematic output formatting necessary for quantitative analysis.
 
-`CatLLM` addresses these gaps by providing a standardized, free-to-use,
+`CatLLM` addresses these gaps by providing a standardized, free-to-use
 interface for applying state-of-the-art language and vision models to
 common research tasks without requiring machine learning expertise. The
-package enables researchers to transform qualitative data into
-quantitative datasets suitable for statistical analysis, bridging the
-gap between traditional qualitative methods and computational
-approaches. Recent research demonstrates that LLMs from OpenAI and
-Anthropic, particularly GPT-4, can effectively replicate human analysis
-performance in content analysis tasks, with some studies showing LLMs
-achieving higher inter-rater reliability than human annotators in
-sentiment analysis and political leaning assessments
-[@bojic_comparing_2025]. Unlike existing tools, CatLLM improves
+package enables researchers to transform diverse data sources—from
+open-ended survey responses and qualitative interviews to unstructured
+web content—into quantitative datasets suitable for statistical
+analysis, bridging the gap between traditional research methods and
+computational approaches. Recent research demonstrates that LLMs from
+OpenAI and Anthropic, particularly GPT-4, can effectively replicate
+human analysis performance in content analysis tasks, with some studies
+showing LLMs achieving higher inter-rater reliability than human
+annotators in sentiment analysis and political leaning assessments
+[@bojic_comparing_2025]. Unlike existing tools, CatLLM provides
 reproducible, structured outputs while supporting multiple AI providers
 and maintaining cost efficiency through built-in optimization features.
 
@@ -87,12 +91,13 @@ and maintaining cost efficiency through built-in optimization features.
 Responses](move_reasons.png)
 
 The software has demonstrated practical impact across diverse research
-domains. It has been successfully applied in studies examining
-demographic differences in LLM performance using the UC Berkeley Social
-Networks Study [@soria_empirical_2025], categorizing occupational data
-according to Standard Occupational Classification codes, and
-implementing automated scoring for cognitive assessments in the
-Caribbean-American Dementia and Aging Study
+domains. It has been successfully applied by institutional researchers
+at UC Berkeley to track student experience and outcomes, in studies
+examining demographic differences in LLM performance using the UC
+Berkeley Social Networks Study [@soria_empirical_2025], categorizing
[…]
