### 2025-10-12 eb4444b
outputting formal urls in web search function
