### 2025-10-13 cfc2455
adding advanced search, which includes dates and confidence scores, more thorough search
