+++ src/catllm/build_web_research.py
@@ -5,14 +5,19 @@ def build_web_research_dataset(
     api_key,
     answer_format = "concise",
     additional_instructions = "",
-    categories = ['Answer','URL'],
+    categories = ['Answer'],
     user_model="claude-sonnet-4-20250514",
     creativity=None,
     safety=False,
     filename="categorized_data.csv",
     save_directory=None,
     model_source="Anthropic",
-    max_retries = 7, #API rate limit error handler retries
+    start_date=None,
+    end_date=None,
+    search_depth="", #enables Tavily searches
+    advanced_search_api_key=None,
+    output_urls = True,
+    max_retries = 6, #API rate limit error handler retries
     time_delay=5
 ):
     import os
@@ -22,22 +27,47 @@ def build_web_research_dataset(
     import regex
     from tqdm import tqdm
     import time
+    from datetime import datetime
+
+    #ensures proper date format
+    def _validate_date(date_str):
+        """Validates YYYY-MM-DD format"""
+        if date_str is None:
+            return True  # None is acceptable (means no date constraint)
+        
+        if not isinstance(date_str, str):
+            return False
+        
+        # Check pattern: YYYY_MM_DD
+        pattern = r'^\d{4}-\d{2}-\d{2}$'
+        if not re.match(pattern, date_str):
+            return False
+        
+        # Validate actual date
+        try:
+            year, month, day = date_str.split('-')
+            datetime(int(year), int(month), int(day))
+            return True
+        except (ValueError, OverflowError):
+            return False
+    
+    # Validate dates at the start of the function
+    if not _validate_date(start_date):
+        raise ValueError(f"start_date must be in YYYY-MM-DD format, got: {start_date}")
+    
+    if not _validate_date(end_date):
+        raise ValueError(f"end_date must be in YYYY-MM-DD format, got: {end_date}")
 
     model_source = model_source.lower() # eliminating case sensitivity 
 
+    # in case user switches to google but doesn't switch model
     if model_source == "google" and user_model == "claude-sonnet-4-20250514":
         user_model = "gemini-2.5-flash"
     
     categories_str = "\n".join(f"{i + 1}. {cat}" for i, cat in enumerate(categories))
-    print(categories_str)
     cat_num = len(categories)
     category_dict = {str(i+1): "0" for i in range(cat_num)}
     example_JSON = json.dumps(category_dict, indent=4)
-
-    # ensure number of categories is what user wants
-    #print("\nThe information to be extracted:")
-    #for i, cat in enumerate(categories, 1):
-        #print(f"{i}. {cat}")
     
     link1 = []
     extracted_jsons = []
@@ -53,7 +83,6 @@ def build_web_research_dataset(
             extracted_urls.append([])
             default_json = example_JSON 
             extracted_jsons.append(default_json)
-            #print(f"Skipped NaN input.")
         else:
             prompt = f"""<role>You are a research assistant specializing in finding current, factual information.</role>
 
@@ -71,14 +100,70 @@ def build_web_research_dataset(
             <format>
             Return your response as valid JSON with this exact structure:
             {{
-            "answer": "Your factual answer or 'Information not found'"
+            "answer": "Your factual answer or 'Information not found'",
+            "second_best_answer": "Your second best factual answer or 'Information not found'",
         }}
+
         </format>"""
-            #print(prompt)
-            if model_source == "anthropic":
+
+            if start_date is not None and end_date is not None:
+                append_text = f"\n- Focus on webpages with a page age between {start_date} and {end_date}."
+                prompt = prompt.replace("<rules>", "<rules>" + append_text)
+            elif start_date is not None:
+                append_text = f"\n- Focus on webpages published after {start_date}."
[…]
