### 2025-10-13 e531e99
added perplexity to web search
