### 2025-10-13 f146c8d
updated package with all latest search features for perplexity
