+++ src/catllm/__about__.py
@@ -1,10 +1,10 @@
 # SPDX-FileCopyrightText: 2025-present Christopher Soria <chrissoria@berkeley.edu>
 #
 # SPDX-License-Identifier: MIT
-__version__ = "0.0.63"
+__version__ = "0.0.65"
 __author__ = "Chris Soria"
 __email__ = "chrissoria@berkeley.edu"
 __title__ = "cat-llm"
-__description__ = "A tool for categorizing and exploring text data and images using LLMs and vision models"
+__description__ = "An AI tool for categorizing and exploring text data, images, and building high quality datasets from the web."
 __url__ = "https://github.com/chrissoria/cat-llm" 
 __license__ = "GPL-3.0"
