### 2025-10-20 5c1c229
added few shot option to mult_class function
