### 2025-10-20 fe1f5a1
changed function back to multi class so users arent pushed off, setting default model for text classification to gpt-5
