### 2025-10-20 a8578af
added huggingface capability to the text classification function
