### 2025-10-20 3d60d51
adding huggingface, auto-detecting model source based on user model
