### 2025-10-21 5db69ca
adding cove prompting
