### 2025-10-21 95a2050
added cove to google statement along with rate limit handling
