+++ src/catllm/text_functions.py
@@ -345,7 +345,7 @@ def multi_class(
             Categorize this survey response "{response}" into the following categories that apply: \
             {categories_str}
             {examples_text}
-            Provide your work in JSON format..."""
+            Provide your work in JSON format where the number belonging to each category is the key and a 1 if the category is present and a 0 if it is not present as key values."""
 
             if chain_of_verification:
                 step2_prompt = f"""You provided this initial categorization:
@@ -410,11 +410,11 @@ def multi_class(
                         try:
                             initial_reply = reply
                             #STEP 2: Generate verification questions
-                            step2_prompt = step2_prompt.replace('<<INITIAL_REPLY>>', initial_reply)
+                            step2_filled = step2_prompt.replace('<<INITIAL_REPLY>>', initial_reply)
 
                             verification_response = client.chat.completions.create(
                                 model=user_model,
-                                messages=[{'role': 'user', 'content': step2_prompt}],
+                                messages=[{'role': 'user', 'content': step2_filled}],
                                 **({"temperature": creativity} if creativity is not None else {})
                                 )
                             
@@ -456,14 +456,15 @@ def multi_class(
                                 **({"temperature": creativity} if creativity is not None else {})
                             )
 
-                            cove_reply = final_response.choices[0].message.content
+                            reply = final_response.choices[0].message.content
     
                             print("Chain of verification completed. Final response generated.\n")
-                            link1.append(cove_reply)
+                            link1.append(reply)
 
                         except Exception as e:
                             print(f"ERROR in Chain of Verification: {str(e)}")
                             print("Falling back to initial response.\n")
+                            link1.append(reply)
                     else:
                         #if chain of verification is not enabled, just append initial reply
                         link1.append(reply)
@@ -476,23 +477,112 @@ def multi_class(
                     link1.append(f"Error processing input: {e}")
 
             elif model_source == "anthropic":
+
                 import anthropic
                 client = anthropic.Anthropic(api_key=api_key)
+                
                 try:
-                    message = client.messages.create(
+                    response_obj = client.messages.create(
                     model=user_model,
-                    max_tokens=1024,
-                    **({"temperature": creativity} if creativity is not None else {}),
-                    messages=[{"role": "user", "content": prompt}]
-                )
-                    reply = message.content[0].text  # Anthropic returns content as list
-                    link1.append(reply)
+                    max_tokens=4096,
+                    messages=[{'role': 'user', 'content': prompt}],
+                    **({"temperature": creativity} if creativity is not None else {})
+                    )
+        
+                    reply = response_obj.content[0].text
+                    
+                    if chain_of_verification:
+                        try:
+                            initial_reply = reply
+                            #STEP 2: Generate verification questions
+                            step2_filled = step2_prompt.replace('<<INITIAL_REPLY>>', initial_reply)
+
+                            verification_response = client.messages.create(
+                                model=user_model,
+                                messages=[{'role': 'user', 'content': step2_filled}],
+                                max_tokens=4096,
+                                **({"temperature": creativity} if creativity is not None else {})
+                                )
+                            
+                            verification_questions = verification_response.content[0].text
+                            #STEP 3: Answer verification questions
+                            questions_list = [
+                                remove_numbering(q) 
+                                for q in verification_questions.split('\n') 
+                                if q.strip()
+                                ]
+                            print(f"Verification questions:\n{questions_list}\n")
+                            verification_qa = []
+
+                            #prompting each question individually
+                            for question in questions_list:
+
+                                step3_filled = step3_prompt.replace('<<QUESTION>>', question)
+
+                                answer_response = client.messages.create(
+                                    model=user_model,
+                                    messages=[{'role': 'user', 'content': step3_filled}],
[…]
