### 2025-10-21 69ae30c
added cove to all model sources
