### 2025-10-21 c82677a
cove, step back, context, seperate functions for text multi class
