+++ src/catllm/text_functions.py
@@ -1,3 +1,15 @@
+from .calls.all_calls import (
+    get_stepback_insight_openai,
+    get_stepback_insight_anthropic,
+    get_stepback_insight_google,
+    get_stepback_insight_mistral,
+    chain_of_verification_openai,
+    chain_of_verification_google,
+    chain_of_verification_anthropic,
+    chain_of_verification_mistral
+)
+
+
 #extract categories from corpus
 def explore_corpus(
     survey_question, 
@@ -244,13 +256,15 @@ def multi_class(
     example4 = None,
     example5 = None,
     example6 = None,
-    creativity=None,
-    safety=False,
-    to_csv=False,
-    chain_of_verification=False,
-    filename="categorized_data.csv",
-    save_directory=None,
-    model_source="auto"
+    creativity = None,
+    safety = False,
+    to_csv = False,
+    chain_of_verification = False,
+    step_back_prompt = False,
+    context_prompt = False,
+    filename = "categorized_data.csv",
+    save_directory = None,
+    model_source = "auto"
 ):
     import os
     import json
@@ -331,6 +345,49 @@ def multi_class(
     else:
         survey_question_context = ""
 
+    # step back insight initializationif step_back_prompt:
+    if step_back_prompt:
+        if survey_question == "": # step back requires the survey question to function well
+            raise TypeError("survey_question is required when using step_back_prompt. Please provide the survey question you are analyzing.")
+                
+        stepback = f"""What are the underlying factors or dimensions that explain how people typically answer "{survey_question}"?"""
+
+        if model_source in ["openai", "perplexity", "huggingface"]:
+            stepback_insight, step_back_added = get_stepback_insight_openai(
+                stepback=stepback,
+                api_key=api_key,
+                user_model=user_model,
+                model_source=model_source,
+                creativity=creativity
+            )
+        elif model_source == "anthropic":
+            stepback_insight, step_back_added = get_stepback_insight_anthropic(
+                stepback=stepback,
+                api_key=api_key,
+                user_model=user_model,
+                model_source=model_source,
+                creativity=creativity
+            )
+        elif model_source == "google":
+            stepback_insight, step_back_added = get_stepback_insight_google(
+                stepback=stepback,
+                api_key=api_key,
+                user_model=user_model,
+                model_source=model_source,
+                creativity=creativity
+            )
+        elif model_source == "mistral":
+            stepback_insight, step_back_added = get_stepback_insight_mistral(
+                stepback=stepback,
+                api_key=api_key,
+                user_model=user_model,
+                model_source=model_source,
+                creativity=creativity
+            )
+    else:
+        stepback_insight = None
+        step_back_added = False
+
     for idx, response in enumerate(tqdm(survey_input, desc="Categorizing responses")):
         reply = None  
 
@@ -346,6 +403,14 @@ def multi_class(
             {categories_str}
             {examples_text}
             Provide your work in JSON format where the number belonging to each category is the key and a 1 if the category is present and a 0 if it is not present as key values."""
+
+            if context_prompt:
+                context = """You are an expert researcher in survey data categorization. 
+                Apply multi-label classification and base decisions on explicit and implicit meanings. 
+                When uncertain, prioritize precision over recall."""
+
+                prompt = context + prompt
+                print(prompt)
 
             if chain_of_verification:
                 step2_prompt = f"""You provided this initial categorization:
@@ -384,7 +449,7 @@ def multi_class(
                 If no categories are present, assign "0" to all categories.
                 Provide the final corrected categorization in the same JSON format:"""
 
-
+            # Main model interaction
             if model_source in ["openai", "perplexity", "huggingface"]:
                 from openai import OpenAI
[…]
