### 2025-10-21 3d36f22
updating readme, removing print from step back, updating text classification example
