### 2025-10-21 9ebc088
added cot prompting
