### 2025-10-23 8ac1cee
fixed issue with exlore corpus, added option to auto create categories in the multi_class function
