### 2025-10-23 f142c6a
updated examples extracting categories with auto
