### 2025-10-24 2b96fed
enabled grok in multi_class
