### 2025-10-25 6622719
fixed rate limiting issues with openai and google
