+++ src/catllm/text_functions.py
@@ -354,11 +354,12 @@ def multi_class(
     chain_of_thought = True,
     step_back_prompt = False,
     context_prompt = False,
+    thinking_budget = 0,
     top_n = 12,
     cat_num = 10,
     divisions = 10,
     research_question = None,
-    filename = "categorized_data.csv",
+    filename = None,
     save_directory = None,
     model_source = "auto"
 ):
@@ -366,6 +367,7 @@ def multi_class(
     import json
     import pandas as pd
     import regex
+    import time
     from tqdm import tqdm
 
     #used in chain of verification 
@@ -441,6 +443,16 @@ def multi_class(
     category_dict = {str(i+1): "0" for i in range(cat_num)}
     example_JSON = json.dumps(category_dict, indent=4)
 
+    # for ensuring JSON format for anthropic api
+    properties = {
+            str(i+1): {
+                "type": "string",
+                "description": cat,
+                "enum": ["0", "1"]
+                } 
+            for i, cat in enumerate(categories)
+        }
+    
     print(f"\nThe categories you entered to be coded by {model_source} {user_model}:")
 
     if categories != "auto":
@@ -585,6 +597,9 @@ def multi_class(
             if model_source in ["openai", "perplexity", "huggingface", "xai"]:
                 from openai import OpenAI
                 from openai import OpenAI, BadRequestError, AuthenticationError
+
+                max_retries = 5
+                delay = 2
                 # conditional base_url setting based on model source
                 base_url = (
                     "https://api.perplexity.ai" if model_source == "perplexity" 
@@ -594,60 +609,93 @@ def multi_class(
                 )
     
                 client = OpenAI(api_key=api_key, base_url=base_url)
-                
-                try:
-                    messages = [
-                        *([{'role': 'user', 'content': stepback}] if step_back_prompt and step_back_added else []), # only if step back is enabled and successful
-                        *([{'role': 'assistant', 'content': stepback_insight}] if step_back_added else {}), # include insight if step back succeeded
-                        {'role': 'user', 'content': prompt}
-                    ]
 
-                    response_obj = client.chat.completions.create(
-                    model=user_model,
-                    messages=messages,
-                    **({"temperature": creativity} if creativity is not None else {})
-                    )
+                max_retries = 5
+                delay = 2
+
+                for attempt in range(max_retries):
+                    try:
         
-                    reply = response_obj.choices[0].message.content
-                    
-                    if chain_of_verification:
-                        reply = chain_of_verification_openai(
-                            initial_reply=reply,
-                            step2_prompt=step2_prompt,
-                            step3_prompt=step3_prompt,
-                            step4_prompt=step4_prompt,
-                            client=client,
-                            user_model=user_model,
-                            creativity=creativity,
-                            remove_numbering=remove_numbering
-                        )
+                        messages = [
+                            *([{'role': 'user', 'content': stepback}] if step_back_prompt and step_back_added else []), # only if step back is enabled and successful
+                            *([{'role': 'assistant', 'content': stepback_insight}] if step_back_added else {}), # include insight if step back succeeded
+                            {'role': 'user', 'content': prompt}
+                        ]
 
-                        link1.append(reply)
-                    else:
-                        #if chain of verification is not enabled, just append initial reply
-                        link1.append(reply)
+                        response_obj = client.chat.completions.create(
+                        model=user_model,
+                        messages=messages,
+                        response_format={"type": "json_object"},
+                        **({"temperature": creativity} if creativity is not None else {})
+                        )
+        
+                        reply = response_obj.choices[0].message.content
                     
-                except BadRequestError as e:
-                    # Model doesn't exist - halt immediately
-                    raise ValueError(f"❌ Model '{user_model}' on {model_source} not found. Please check the model name and try again.") from e
[…]

[…]
