### 2025-10-25 199209c
fixing issue where whole row was converted to missing if one of the categories wasn't output
