### 2025-10-25 a03c801
added fix so that if huggingface outputs incorrect JSON format, retries
