### 2025-10-26 bedf63a
fixing column converting to 0s for valid rows
