### 2025-10-28 8e3358e
fixing glitch that caused explore corpus to fail if a non string value was in one of the rows
