+++ src/catllm/text_functions.py
@@ -126,211 +126,226 @@ def explore_common_categories(
     creativity=None,
     specificity="broad",
     research_question=None,
-    filename=None,
-    model_source="OpenAI"
+    filename=None,          # (kept, unused here)
+    model_source="openai",  # case-insensitive
+    iterations=1,
+    random_state=None
 ):
-    import os
+    import re
     import pandas as pd
-    import random
-    from openai import OpenAI
-    from openai import OpenAI, BadRequestError
+    import numpy as np
     from tqdm import tqdm
 
-    print(f"Exploring class for question: '{survey_question}'.\n          {cat_num * divisions} unique categories to be extracted and {top_n} to be identified as the most common.")
-    print()
+    model_source = (model_source or "openai").lower()
 
-    model_source = model_source.lower() # eliminating case sensitivity 
+    # --- input normalization ---
+    if not isinstance(survey_input, pd.Series):
+        survey_input = pd.Series(survey_input)
+    survey_input = survey_input.dropna().astype("string")
+    n = len(survey_input)
+    if n == 0:
+        raise ValueError("survey_input is empty after dropping NA.")
 
-    chunk_size = round(max(1, len(survey_input) / divisions),0)
-    chunk_size = int(chunk_size)
+    # --- chunk sizing ---
+    chunk_size = int(round(max(1, n / divisions), 0))
+    if chunk_size < (cat_num / 2):
+        raise ValueError(
+            f"Cannot extract {cat_num} categories from chunks of only {chunk_size} responses.\n"
+            f"Solutions:\n"
+            f"  (1) Reduce 'divisions' (currently {divisions}) to make larger chunks, or\n"
+            f"  (2) Reduce 'cat_num' (currently {cat_num})."
+        )
 
-    if chunk_size < (cat_num/2):
-        raise ValueError(f"Cannot extract {cat_num} categories from chunks of only {chunk_size} responses. \n" 
-                    f"Choose one solution: \n"
-                    f"(1) Reduce 'divisions' parameter (currently {divisions}) to create larger chunks, or \n"
-                    f"(2) Reduce 'cat_num' parameter (currently {cat_num}) to extract fewer categories per chunk.")
+    print(
+        f"Exploring class for question: '{survey_question}'.\n"
+        f"          {cat_num * divisions} unique categories to be extracted and {top_n} to be identified as the most common.\n"
+    )
 
-    random_chunks = []
-    for i in range(divisions):
-        chunk = survey_input.sample(n=chunk_size).tolist()
-        random_chunks.append(chunk)
-    
-    responses = []
-    responses_list = []
-    
-    for i in tqdm(range(divisions), desc="Processing chunks"):
-        survey_participant_chunks = '; '.join(random_chunks[i])
-        prompt = f"""Identify {cat_num} {specificity} categories of responses to the question "{survey_question}" in the following list of responses. \
-Responses are each separated by a semicolon. \
-Responses are contained within triple backticks here: ```{survey_participant_chunks}``` \
-Number your categories from 1 through {cat_num} and be concise with the category labels and provide no description of the categories."""
-        
-        if model_source == "openai":
-            try:
-                from openai import OpenAI
+    # --- RNG for reproducible re-sampling across passes ---
+    rng = np.random.default_rng(random_state)
 
-                client = OpenAI(api_key=api_key)
-
-                response_obj = client.chat.completions.create(
-                    model=user_model,
-                    messages=[
-                        {'role': 'system', 'content': f"""You are a helpful assistant that extracts categories from survey responses. \
-                                        The specific task is to identify {specificity} categories of responses to a survey question. \
-                                        The research question is: {research_question}""" if research_question else "You are a helpful assistant."},
-                        {'role': 'user', 'content': prompt}
-                        ],
[…]
