### 2025-11-04 20f79c7
incorporated more error handling in google, openai, and mistral, fixed issue with CoT prompt wasn't producing a structured output
