### 2025-11-04 1a15f35
updated paper to reflect that packge allows switching prompt types and added citations
