+++ academic_examples/catllm.bib
@@ -50,7 +50,7 @@
 	month = jun,
 	year = {2024},
 	note = {arXiv:2309.02726 [cs]},
-	keywords = {Computer Science - Artificial Intelligence, Computer Science - Computation and Language},
+	keywords = {Computer Science - Computation and Language, Computer Science - Artificial Intelligence},
 	annote = {Comment: Accepted by ACL 2024 (findings)},
 	file = {Preprint PDF:/Users/chrissoria/Zotero/storage/N6AR8YTY/Yang et al. - 2024 - Large Language Models for Automated Open-domain Sc.pdf:application/pdf;Snapshot:/Users/chrissoria/Zotero/storage/ANMAFX7I/2309.html:text/html},
 }
@@ -104,7 +104,7 @@ The findings imply that social science researchers who want to use LLMs should d
 	author = {Soria, Chris},
 	month = jun,
 	year = {2025},
-	keywords = {AI, Automated Coding, Coding, Large Language Models, Open Ended Survey Data},
+	keywords = {Large Language Models, AI, Automated Coding, Coding, Open Ended Survey Data},
 	file = {OSF Preprint:/Users/chrissoria/Zotero/storage/B8WQNS5U/Soria - 2025 - An Empirical Investigation into the Utility of Lar.pdf:application/pdf},
 }
 
@@ -119,7 +119,7 @@ The findings imply that social science researchers who want to use LLMs should d
 	month = nov,
 	year = {2024},
 	note = {arXiv:2406.00515 [cs]},
-	keywords = {Computer Science - Artificial Intelligence, Computer Science - Computation and Language, Computer Science - Software Engineering},
+	keywords = {Computer Science - Computation and Language, Computer Science - Artificial Intelligence, Computer Science - Software Engineering},
 	file = {Preprint PDF:/Users/chrissoria/Zotero/storage/7QTN7H6U/Jiang et al. - 2024 - A Survey on Large Language Models for Code Generat.pdf:application/pdf;Snapshot:/Users/chrissoria/Zotero/storage/4YWQBZFT/2406.html:text/html},
 }
 
@@ -162,4 +162,51 @@ The findings imply that social science researchers who want to use LLMs should d
 	keywords = {Human behaviour, Information theory and computation},
 	pages = {11477},
 	file = {Full Text PDF:/Users/chrissoria/Zotero/storage/YKX29DXN/Bojić et al. - 2025 - Comparing large Language models and human annotato.pdf:application/pdf},
+}
+
+@misc{wei_chain_thought_2023,
+	title = {Chain-of-{Thought} {Prompting} {Elicits} {Reasoning} in {Large} {Language} {Models}},
+	url = {http://arxiv.org/abs/2201.11903},
+	doi = {10.48550/arXiv.2201.11903},
+	abstract = {We explore how generating a chain of thought -- a series of intermediate reasoning steps -- significantly improves the ability of large language models to perform complex reasoning. In particular, we show how such reasoning abilities emerge naturally in sufficiently large language models via a simple method called chain of thought prompting, where a few chain of thought demonstrations are provided as exemplars in prompting. Experiments on three large language models show that chain of thought prompting improves performance on a range of arithmetic, commonsense, and symbolic reasoning tasks. The empirical gains can be striking. For instance, prompting a 540B-parameter language model with just eight chain of thought exemplars achieves state of the art accuracy on the GSM8K benchmark of math word problems, surpassing even finetuned GPT-3 with a verifier.},
+	urldate = {2025-11-04},
+	publisher = {arXiv},
[…]
