### 2025-11-23 1ddcb69
cleaned up prompt code, made extract categories function work with other model providers
