+++ src/catllm/text_functions.py
@@ -1,14 +1,20 @@
-from .calls.all_calls import (
+from .calls.stepback import (
     get_stepback_insight_openai,
     get_stepback_insight_anthropic,
     get_stepback_insight_google,
-    get_stepback_insight_mistral,
+    get_stepback_insight_mistral
+)
+from .calls.CoVe import (
     chain_of_verification_openai,
     chain_of_verification_google,
     chain_of_verification_anthropic,
-    chain_of_verification_mistral,
+    chain_of_verification_mistral
+)
+from .calls.top_n import (
     get_openai_top_n,
-    get_anthropic_top_n
+    get_anthropic_top_n,
+    get_google_top_n,
+    get_mistral_top_n
 )
 
 
@@ -19,7 +25,7 @@ def explore_corpus(
     api_key,
     research_question=None,
     specificity="broad",
-    cat_num=10,
+    categories_per_chunk=10,
     divisions=5,
     user_model="gpt-5",
     creativity=None,
@@ -33,7 +39,7 @@ def explore_corpus(
     from openai import OpenAI, BadRequestError
     from tqdm import tqdm
 
-    print(f"Exploring class for question: '{survey_question}'.\n          {cat_num * divisions} unique categories to be extracted.")
+    print(f"Exploring class for question: '{survey_question}'.\n          {categories_per_chunk * divisions} unique categories to be extracted.")
     print()
 
     model_source = model_source.lower() # eliminating case sensitivity 
@@ -41,11 +47,11 @@ def explore_corpus(
     chunk_size = round(max(1, len(survey_input) / divisions),0)
     chunk_size = int(chunk_size)
 
-    if chunk_size < (cat_num/2):
-        raise ValueError(f"Cannot extract {cat_num} {specificity} categories from chunks of only {chunk_size} responses. \n" 
+    if chunk_size < (categories_per_chunk/2):
+        raise ValueError(f"Cannot extract {categories_per_chunk} {specificity} categories from chunks of only {chunk_size} responses. \n" 
                     f"Choose one solution: \n"
                     f"(1) Reduce 'divisions' parameter (currently {divisions}) to create larger chunks, or \n"
-                    f"(2) Reduce 'cat_num' parameter (currently {cat_num}) to extract fewer categories per chunk.")
+                    f"(2) Reduce 'categories_per_chunk' parameter (currently {categories_per_chunk}) to extract fewer categories per chunk.")
 
     random_chunks = []
     for i in range(divisions):
@@ -57,10 +63,10 @@ def explore_corpus(
     
     for i in tqdm(range(divisions), desc="Processing chunks"):
         survey_participant_chunks = '; '.join(random_chunks[i])
-        prompt = f"""Identify {cat_num} {specificity} categories of responses to the question "{survey_question}" in the following list of responses. \
+        prompt = f"""Identify {categories_per_chunk} {specificity} categories of responses to the question "{survey_question}" in the following list of responses. \
 Responses are each separated by a semicolon. \
 Responses are contained within triple backticks here: ```{survey_participant_chunks}``` \
-Number your categories from 1 through {cat_num} and be concise with the category labels and provide no description of the categories."""
+Number your categories from 1 through {categories_per_chunk} and be concise with the category labels and provide no description of the categories."""
         
         if model_source == "openai":
             client = OpenAI(api_key=api_key)
@@ -116,18 +122,18 @@ Number your categories from 1 through {cat_num} and be concise with the category
 
 #extract top categories from corpus
 def explore_common_categories(
-    survey_question, #make optional later
     survey_input,
     api_key,
-    top_n=12, #change paaram name to num_top_categories
-    cat_num=10, #need to make more clear what this is
+    survey_question="",
+    max_categories=12,
[…]
