### 2025-12-10 5c2b07b
refactored and tested multi_class
