+++ src/catllm/calls/image_CoVe.py
@@ -1,6 +1,8 @@
-# openai chain of verification calls
+# Image-aware Chain of Verification (CoVe) functions for various LLM providers
+# These functions include the image in verification steps for accurate image-based categorization
 
-def chain_of_verification_openai(
+
+def image_chain_of_verification_openai(
     initial_reply,
     step2_prompt,
     step3_prompt,
@@ -8,68 +10,87 @@ def chain_of_verification_openai(
     client,
     user_model,
     creativity,
-    remove_numbering
+    remove_numbering,
+    image_content
 ):
     """
-    Execute Chain of Verification (CoVe) process.
+    Execute Chain of Verification (CoVe) process for images with OpenAI.
+    The image is included in verification steps for accurate assessment.
     Returns the verified reply or initial reply if error occurs.
+
+    Args:
+        image_content: The image content in OpenAI format (list with image_url dict)
     """
     try:
-        # STEP 2: Generate verification questions
+        # STEP 2: Generate verification questions (text only - questions about the categorization)
         step2_filled = step2_prompt.replace('<<INITIAL_REPLY>>', initial_reply)
-        
+
         verification_response = client.chat.completions.create(
             model=user_model,
             messages=[{'role': 'user', 'content': step2_filled}],
             **({"temperature": creativity} if creativity is not None else {})
         )
-        
+
         verification_questions = verification_response.choices[0].message.content
-        
-        # STEP 3: Answer verification questions
+
+        # STEP 3: Answer verification questions WITH the image
         questions_list = [
-            remove_numbering(q) 
-            for q in verification_questions.split('\n') 
+            remove_numbering(q)
+            for q in verification_questions.split('\n')
             if q.strip()
         ]
         verification_qa = []
-        
-        # Prompting each question individually
+
         for question in questions_list:
             step3_filled = step3_prompt.replace('<<QUESTION>>', question)
-            
+
+            # Include image in the verification question
+            message_content = [
+                {"type": "text", "text": step3_filled},
+                image_content
+            ]
+
             answer_response = client.chat.completions.create(
                 model=user_model,
-                messages=[{'role': 'user', 'content': step3_filled}],
+                messages=[{'role': 'user', 'content': message_content}],
                 **({"temperature": creativity} if creativity is not None else {})
             )
-            
+
             answer = answer_response.choices[0].message.content
             verification_qa.append(f"Q: {question}\nA: {answer}")
-        
-        # STEP 4: Final corrected categorization
+
+        # STEP 4: Final corrected categorization WITH the image
         verification_qa_text = "\n\n".join(verification_qa)
-        
+
         step4_filled = (step4_prompt
             .replace('<<INITIAL_REPLY>>', initial_reply)
             .replace('<<VERIFICATION_QA>>', verification_qa_text))
-        
+
+        # Include image in final categorization
+        final_message_content = [
+            {"type": "text", "text": step4_filled},
+            image_content
+        ]
+
         final_response = client.chat.completions.create(
             model=user_model,
-            messages=[{'role': 'user', 'content': step4_filled}],
+            messages=[{'role': 'user', 'content': final_message_content}],
+            response_format={"type": "json_object"},
             **({"temperature": creativity} if creativity is not None else {})
         )
 
         verified_reply = final_response.choices[0].message.content
+        print("Image chain of verification completed. Final response generated.\n")
 
         return verified_reply
 
     except Exception as e:
+        print(f"ERROR in Image Chain of Verification: {str(e)}")
+        print("Falling back to initial response.\n")
         return initial_reply
 
-# anthropic chain of verification calls
 
-def chain_of_verification_anthropic(
+def image_chain_of_verification_anthropic(
     initial_reply,
     step2_prompt,
     step3_prompt,
[…]
