### 2025-12-10 64c4b5c
adding warning to cove users
