### 2025-12-10 a3da319
fixed up image_multi_class to not print statement in cove prompting
