+++ src/catllm/calls/stepback.py
@@ -33,7 +33,6 @@ def get_stepback_insight_openai(
         return stepback_insight, True
         
     except Exception as e:
-        print(f"An error occurred during step-back prompting: {e}")
         return None, False
 
 
@@ -63,7 +62,6 @@ def get_stepback_insight_anthropic(
         return stepback_insight, True
         
     except Exception as e:
-        print(f"An error occurred during step-back prompting: {e}")
         return None, False
 
 
@@ -103,7 +101,6 @@ def get_stepback_insight_google(
         return stepback_insight, True
         
     except Exception as e:
-        print(f"An error occurred during step-back prompting: {e}")
         return None, False
 
 
@@ -132,5 +129,4 @@ def get_stepback_insight_mistral(
         return stepback_insight, True
         
     except Exception as e:
-        print(f"An error occurred during step-back prompting: {e}")
         return None, False
