### 2025-12-10 fc49ec1
revising paper to communicate corpus discovery a bit clearer and pyshing the cog use case to background
