+++ src/catllm/build_web_research.py
@@ -1,6 +1,6 @@
 #build dataset classification
 def build_web_research_dataset(
-    search_question, 
+    search_question,
     search_input,
     api_key,
     answer_format = "concise",
@@ -14,8 +14,6 @@ def build_web_research_dataset(
     model_source="Anthropic",
     start_date=None,
     end_date=None,
-    search_depth="", #enables Tavily searches
-    tavily_api=None,
     output_urls = True,
     max_retries = 6, #API rate limit error handler retries
     time_delay=5
@@ -34,15 +32,15 @@ def build_web_research_dataset(
         """Validates YYYY-MM-DD format"""
         if date_str is None:
             return True  # None is acceptable (means no date constraint)
-        
+
         if not isinstance(date_str, str):
             return False
-        
+
         # Check pattern: YYYY_MM_DD
         pattern = r'^\d{4}-\d{2}-\d{2}$'
         if not re.match(pattern, date_str):
             return False
-        
+
         # Validate actual date
         try:
             year, month, day = date_str.split('-')
@@ -50,30 +48,30 @@ def build_web_research_dataset(
             return True
         except (ValueError, OverflowError):
             return False
-    
+
     # Validate dates at the start of the function
     if not _validate_date(start_date):
         raise ValueError(f"start_date must be in YYYY-MM-DD format, got: {start_date}")
-    
+
     if not _validate_date(end_date):
         raise ValueError(f"end_date must be in YYYY-MM-DD format, got: {end_date}")
 
-    model_source = model_source.lower() # eliminating case sensitivity 
+    model_source = model_source.lower() # eliminating case sensitivity
 
     if model_source == "perplexity" and start_date is not None:
         start_date = datetime.strptime(start_date, "%Y-%m-%d").strftime("%m/%d/%Y")
-    if model_source == "perplexity" and end_date is not None:  
+    if model_source == "perplexity" and end_date is not None:
         end_date = datetime.strptime(end_date, "%Y-%m-%d").strftime("%m/%d/%Y")
 
     # in case user switches to google but doesn't switch model
     if model_source == "google" and user_model == "claude-sonnet-4-20250514":
         user_model = "gemini-2.5-flash"
-    
+
     categories_str = "\n".join(f"{i + 1}. {cat}" for i, cat in enumerate(categories))
     cat_num = len(categories)
     category_dict = {str(i+1): "0" for i in range(cat_num)}
     example_JSON = json.dumps(category_dict, indent=4)
-    
+
     link1 = []
     extracted_jsons = []
     extracted_urls = []
@@ -81,12 +79,12 @@ def build_web_research_dataset(
     for idx, item in enumerate(tqdm(search_input, desc="Building dataset")):
         if idx > 0:  # Skip delay for first item only
             time.sleep(time_delay)
-        reply = None  
+        reply = None
 
-        if pd.isna(item): 
+        if pd.isna(item):
             link1.append("Skipped NaN input")
             extracted_urls.append([])
-            default_json = example_JSON 
+            default_json = example_JSON
             extracted_jsons.append(default_json)
         else:
             prompt = f"""<role>You are a research assistant specializing in finding current, factual information.</role>
@@ -122,58 +120,9 @@ def build_web_research_dataset(
                 append_text = f"\n- Focus on webpages published before {end_date}."
                 prompt = prompt.replace("<rules>", "<rules>" + append_text)
 
-            if search_depth == "advanced" and model_source != "perplexity":
-                try:
-                    from tavily import TavilyClient
-                    tavily_client = TavilyClient(tavily_api)
-                    tavily_response = tavily_client.search(
-                        query=f"{item}'s {search_question}",
-                        include_answer=True,
-                        max_results=15,
-                        search_depth="advanced",
[…]
