### 2025-12-10 56c3d67
cleaning up the cerad function to be more clearly an offshoot of image_multi_class
