+++ src/catllm/CERAD_functions.py
@@ -1,479 +1,238 @@
-# a function specifically for CERAD Constructional Praxis test
-# specifically for pictures of drawings of shapes like circles, diamonds, rectangles, and cubes
+# CERAD Constructional Praxis scoring function
+# Scores drawings of shapes (circles, diamonds, rectangles, cubes) using image_multi_class
 
 """
-Ket features:
-1. Shape-specific scoring: The function can handle different shapes (circle, diamond, rectangles, cube) and provides tailored categories for each shape.
-2. Image input handling: It accepts image inputs either as file paths or a list of images.
-3. Model flexibility: The function allows users to specify different models (OpenAI, Anthropic, Perplexity, Mistral) for image analysis.
-4. Safety and progress saving: It can save progress to a CSV file, which is useful for long-running tasks or when processing many images.
+This function wraps image_multi_class to implement CERAD Constructional Praxis test scoring.
+It defines shape-specific categories, calls image_multi_class for the LLM classification,
+then applies algorithmic scoring rules to convert binary classifications into CERAD scores.
 
-Areas for improvement:
-1. Prompt refinement: adjusting the prompt so that it produces a more accurate score.
-2. Image preprocessing: adjusting the images so that they are easier to be analyzed by the models.
-3. Model selection: using a different model that is better suited for image analysis.
-4. Model Ensembling: using multiple models and combining their scores to produce a more accurate score.
-5. Prompt ensembling: using multiple prompts and combining their scores to produce a more accurate score.
-6. Post-processing: adjusting the way scores are calculated after the model has output its assessment.
-7. Efficiency: optimizing the code to run faster, cheaper, and more efficiently.
-8. Drawn-format versatility: making the function more versatile to handle different scenarios, such as shapes drawn on tablets.
-9. Image input flexibility: allowing the function to accept images in various formats, such as URLs or raw image data.
-10. Test variety: expanding or adding functions to handle score more tests relevant for cogntive assesment, such as the MMSE.
-11. Error handling: improving error handling to better manage unexpected inputs or model failures.
+Supported shapes:
+- circle: max score 2
+- diamond: max score 3
+- rectangles (overlapping): max score 2
+- cube: max score 4
 """
+
+from .image_functions import image_multi_class
+
 
 def cerad_drawn_score(
-    shape, 
+    shape,
     image_input,
     api_key,
     user_model="gpt-4o",
     creativity=None,
-    reference_in_image=False,
-    provide_reference=False,
     safety=False,
-    filename="categorized_data.csv",
-    model_source="OpenAI"
+    chain_of_thought=True,
+    filename=None,
+    save_directory=None,
+    model_source="auto"
 ):
-    import os
-    import json
-    import pandas as pd
-    import regex
-    from tqdm import tqdm
-    import glob
-    import base64
-    from pathlib import Path
-    import pkg_resources
+    """
+    Score CERAD Constructional Praxis drawings using image_multi_class.
 
-    model_source = model_source.lower() # eliminating case sensitivity 
+    Args:
+        shape (str): Shape to score - 'circle', 'diamond', 'rectangles', or 'cube'
+        image_input: Directory path or list of image file paths
+        api_key (str): API key for the model provider
+        user_model (str): Model to use (default: gpt-4o)
+        creativity (float): Temperature setting (None uses model default)
+        safety (bool): If True, saves progress after each image
+        chain_of_thought (bool): If True, uses step-by-step reasoning
+        filename (str): Output filename for CSV
+        save_directory (str): Directory to save results
+        model_source (str): Provider - 'auto', 'openai', 'anthropic', 'google', 'mistral'
+
+    Returns:
[…]
