### 2025-12-10 1608956
cleaning up paper to be more JOSS friendly
