+++ academic_examples/paper.md
@@ -8,188 +8,67 @@ tags:
 - content coding
 - image classification
 date: "29 May 2025"
-authors:
-- name: Chris Soria
-  orcid: "0009-0007-3558-5896"
-  affiliation: 1
+output: pdf_document
 bibliography: catllm.bib
 affiliations:
 - name: University of California, Berkeley, United States
   index: 1
+authors:
+- name: Chris Soria
+  orcid: "0009-0007-3558-5896"
+  affiliation: 1
 ---
 
 # Summary
 
-The rapid advancement of large language and vision models has created
-new opportunities for automated text and image analysis in social
-science research [@schulze_buschoff_visual_2025; @yang_large_2024;
-@sachdeva_normative_2025]. Researchers increasingly use these tools to
-code open-ended survey responses, categorize qualitative data, and
-analyze visual content at scale. Yet challenges persist due to
-inconsistent output formats, diverse API interfaces, and the lack of
-standardized workflows for integrating both model outputs and external
-data sources into traditional statistical analysis pipelines
-[@rossi_problems_2024]. CatLLM addresses these issues by providing a
-modular framework with specialized functions that not only ensure
-consistent data structures across text and image analysis workflows, but
-also facilitate the automated retrieval of structured data from the web.
-The package handles different prompting strategies reactively through
-configurable parameters that allow users to switch between techniques
-such as Chain-of-Thought (CoT) [@wei_chain_thought_2023],
-Chain-of-Verification (CoVe) [@dhuliawala_chain_verification_2023], and
-step-back prompting [@zheng_take_2024], enabling researchers to optimize
-model reasoning based on task complexity without requiring expertise in
-prompt engineering. This integration allows researchers to seamlessly
-combine large model outputs with real-world datasets, maintaining
-compatibility with standard statistical analysis tools.
+The rapid advancement of large language and vision models has created new opportunities for automated text and image analysis in social science research [@schulze_buschoff_visual_2025; @yang_large_2024; @sachdeva_normative_2025]. Researchers increasingly use these tools to code open-ended survey responses, categorize qualitative data, and analyze visual content at scale. Yet challenges persist due to inconsistent output formats, diverse API interfaces, and the lack of standardized workflows for integrating both model outputs and external data sources into traditional statistical analysis pipelines [@rossi_problems_2024]. CatLLM addresses these issues by providing a modular framework with specialized functions that not only ensure consistent data structures across text and image analysis workflows, but also facilitate the automated retrieval of structured data from the web. The package handles different prompting strategies reactively through configurable parameters that allow users to switch between techniques such as Chain-of-Thought (CoT) [@wei_chain_thought_2023], Chain-of-Verification (CoVe) [@dhuliawala_chain_verification_2023], and step-back prompting [@zheng_take_2024], enabling researchers to optimize model reasoning based on task complexity without requiring expertise in prompt engineering. This integration allows researchers to seamlessly combine large model outputs with real-world datasets, maintaining compatibility with standard statistical analysis tools.
 
 # Statement of need
 
-Social scientists increasingly recognize the value of open-ended survey
-input for capturing rich, nuanced responses that closed-ended formats
-cannot provide. However, many researchers avoid incorporating open-ended
-input into their surveys due to the substantial analysis challenges they
-present. The processing of open-ended responses is notoriously
-time-intensive, requiring manual categorization and careful
-interpretation that can quickly become overwhelming with large datasets.
-Even when researchers do include open-ended questions, quantitative
-researchers often fail to fully utilize the resulting qualitative data
-due to limited time, resources, or expertise in analysis techniques.
-This analysis burden not only increases research costs but also creates
-practical barriers that prevent researchers from leveraging the deeper
-insights that open-ended responses can provide.
[…]
