+++ src/catllm/build_web_research.py
@@ -1,355 +1,0 @@
-#build dataset classification
-def build_web_research_dataset(
-    search_question,
-    search_input,
-    api_key,
-    answer_format = "concise",
-    additional_instructions = "",
-    categories = ['Answer'],
-    user_model="claude-sonnet-4-20250514",
-    creativity=None,
-    safety=False,
-    filename="categorized_data.csv",
-    save_directory=None,
-    model_source="Anthropic",
-    start_date=None,
-    end_date=None,
-    output_urls = True,
-    max_retries = 6, #API rate limit error handler retries
-    time_delay=5
-):
-    import os
-    import re
-    import json
-    import pandas as pd
-    import regex
-    from tqdm import tqdm
-    import time
-    from datetime import datetime
-
-    #ensures proper date format
-    def _validate_date(date_str):
-        """Validates YYYY-MM-DD format"""
-        if date_str is None:
-            return True  # None is acceptable (means no date constraint)
-
-        if not isinstance(date_str, str):
-            return False
-
-        # Check pattern: YYYY_MM_DD
-        pattern = r'^\d{4}-\d{2}-\d{2}$'
-        if not re.match(pattern, date_str):
-            return False
-
-        # Validate actual date
-        try:
-            year, month, day = date_str.split('-')
-            datetime(int(year), int(month), int(day))
-            return True
-        except (ValueError, OverflowError):
-            return False
-
-    # Validate dates at the start of the function
-    if not _validate_date(start_date):
-        raise ValueError(f"start_date must be in YYYY-MM-DD format, got: {start_date}")
-
-    if not _validate_date(end_date):
-        raise ValueError(f"end_date must be in YYYY-MM-DD format, got: {end_date}")
-
-    model_source = model_source.lower() # eliminating case sensitivity
-
-    if model_source == "perplexity" and start_date is not None:
-        start_date = datetime.strptime(start_date, "%Y-%m-%d").strftime("%m/%d/%Y")
-    if model_source == "perplexity" and end_date is not None:
-        end_date = datetime.strptime(end_date, "%Y-%m-%d").strftime("%m/%d/%Y")
-
-    # in case user switches to google but doesn't switch model
-    if model_source == "google" and user_model == "claude-sonnet-4-20250514":
-        user_model = "gemini-2.5-flash"
-
-    categories_str = "\n".join(f"{i + 1}. {cat}" for i, cat in enumerate(categories))
-    cat_num = len(categories)
-    category_dict = {str(i+1): "0" for i in range(cat_num)}
-    example_JSON = json.dumps(category_dict, indent=4)
-
-    link1 = []
-    extracted_jsons = []
-    extracted_urls = []
-
-    for idx, item in enumerate(tqdm(search_input, desc="Building dataset")):
-        if idx > 0:  # Skip delay for first item only
-            time.sleep(time_delay)
-        reply = None
-
-        if pd.isna(item):
-            link1.append("Skipped NaN input")
-            extracted_urls.append([])
-            default_json = example_JSON
-            extracted_jsons.append(default_json)
-        else:
-            prompt = f"""<role>You are a research assistant specializing in finding current, factual information.</role>
-
-            <task>Find information about {item}'s {search_question}</task>
-
-            <rules>
-            - Search for the most current and authoritative information available
-            - Provide your answer as {answer_format}
-            - Prioritize official sources when possible
-            - If information is not found, state "Information not found"
-            - Do not include any explanatory text or commentary beyond the JSON
-                {additional_instructions}
-            </rules>
-
-            <format>
-            Return your response as valid JSON with this exact structure:
-            {{
-            "answer": "Your factual answer or 'Information not found'",
-            "second_best_answer": "Your second best factual answer or 'Information not found'",
[…]
