+++ README.md
@@ -21,8 +21,8 @@ CatLLM: A Reproducible LLM Pipeline for Coding Open-Ended Survey Responses
   - [multi_class()](#multi_class)
   - [image_score()](#image_score_drawing)
   - [image_features()](#image_features)
-  - [build_web_research_dataset()](#build_web_research_dataset)
   - [cerad_drawn_score()](#cerad_drawn_score)
+- [Related Projects](#related-projects)
 - [Academic Research](#academic-research)
 - [Contact](#contact)
 - [License](#license)
@@ -35,32 +35,15 @@ pip install cat-llm
 
 -----
 
-## ⚠️ Beta Software Notice
-
-**CatLLM is currently in beta.** While I'm actively working to make this tool robust and reliable, you may encounter bugs or unexpected behavior. 
-
-**I need your help!** This project thrives on community feedback. Please contribute by:
-- 🐛 **Reporting bugs** - Found an issue? Let me know!
-- 💡 **Sharing ideas** - Have suggestions for improvements or new features?
-- 🔧 **Contributing code** - Submit pull requests with fixes or enhancements
-
-**Visit our GitHub**: [github.com/chrissoria/cat-llm](https://github.com/chrissoria/cat-llm)
-
-All feedback helps us build better research software for the community.
-
------
-
 ## Quick Start
 
-CatLLM helps social scientists and researchers automatically categorize open-ended survey responses, images, and web-scraped data using AI models like GPT-5 and Claude. Not to be confused with CAT-LLM for Chinese article‐style transfer ([Tao et al. 2024](https://arxiv.org/html/2401.05707v1)).
+CatLLM helps social scientists and researchers automatically categorize open-ended survey responses and images using AI models like GPT-5 and Claude. Not to be confused with CAT-LLM for Chinese article‐style transfer ([Tao et al. 2024](https://arxiv.org/html/2401.05707v1)).
 
-Text Analysis: Simply provide your survey responses and category list - the package handles the rest and outputs clean data ready for statistical analysis. It works with single or multiple categories per response and automatically skips missing data to save API costs.
+**Text Analysis:** Simply provide your survey responses and category list - the package handles the rest and outputs clean data ready for statistical analysis. It works with single or multiple categories per response and automatically skips missing data to save API costs.
 
-Image Categorization: Uses the same intelligent categorization method to analyze images, extracting specific features, counting objects, identifying colors, or determining the presence of elements based on your research questions.
+**Image Categorization:** Uses the same intelligent categorization method to analyze images, extracting specific features, counting objects, identifying colors, or determining the presence of elements based on your research questions.
 
-Web Data Collection: Builds comprehensive datasets by scraping web data and using Large Language Models to extract exactly the information you need. The function searches across multiple sources, processes the findings through AI models, and structures everything into clean dataframe format ready for export to CSV.
-
-Whether you're working with messy text responses, analyzing visual content, or gathering information from across the web, CatLLM consistently transforms unstructured data into structured categories and datasets you can actually analyze. All outputs are formatted for immediate statistical analysis and can be exported directly to CSV for further research workflows.
+Whether you're working with messy text responses or analyzing visual content, CatLLM consistently transforms unstructured data into structured categories you can actually analyze. All outputs are formatted for immediate statistical analysis and can be exported directly to CSV for further research workflows.
 
 
 
@@ -376,60 +359,6 @@ image_scores = cat.image_features(
     api_key="OPENAI_API_KEY")
 ```
 
-### `build_web_research_dataset()`
-
-Conducts automated web research on specified topics and compiles the findings into a structured dataset, extracting answers and source URLs for comprehensive research workflows. 
[…]
