### 2025-12-22 #9678 @editorialbot
[REVIEW]: CatLLM: A Python package for Generating, Assigning, and ScoringOpen-Ended Survey Data and Images
**Submitting author:** @chrissoria (<a href="http://orcid.org/0009-0007-3558-5896">Chris Soria</a>)
**Repository:** https://github.com/chrissoria/cat-llm
**Branch with paper.md** (empty if default branch):
**Version:** 3.0.0
