### 2025-12-30 511a69a
added pdf processing function
