+++ src/catllm/pdf_functions.py
@@ -1,0 +1,815 @@
+from .text_functions import _detect_model_source
+from .calls.pdf_stepback import get_pdf_stepback_insight
+from .calls.pdf_CoVe import (
+    pdf_chain_of_verification_openai,
+    pdf_chain_of_verification_anthropic,
+    pdf_chain_of_verification_google,
+    pdf_chain_of_verification_mistral
+)
+
+
+def _load_pdf_files(pdf_input):
+    """Load PDF files from directory path or return list as-is."""
+    import os
+    import glob
+
+    if not isinstance(pdf_input, list):
+        pdf_files = glob.glob(os.path.join(pdf_input, '*.pdf'))
+        pdf_files.extend(glob.glob(os.path.join(pdf_input, '*.PDF')))
+        # Remove duplicates (case-insensitive systems)
+        seen = set()
+        unique_files = []
+        for f in pdf_files:
+            if f.lower() not in seen:
+                seen.add(f.lower())
+                unique_files.append(f)
+        pdf_files = unique_files
+        print(f"Found {len(pdf_files)} PDFs.")
+    else:
+        pdf_files = pdf_input
+        print(f"Provided a list of {len(pdf_input)} PDFs.")
+
+    return pdf_files
+
+
+def _get_pdf_pages(pdf_path):
+    """
+    Extract all pages from a PDF as separate page objects.
+    Returns list of tuples: [(page_index, page_label), ...]
+
+    For 'document.pdf' with 3 pages:
+    [(0, "document_p1"), (1, "document_p2"), (2, "document_p3")]
+
+    The actual page data is extracted later based on provider needs.
+    """
+    import os
+    from pathlib import Path
+
+    try:
+        import fitz  # PyMuPDF
+    except ImportError:
+        raise ImportError(
+            "PyMuPDF is required for PDF processing. "
+            "Install it with: pip install PyMuPDF"
+        )
+
+    pdf_name = Path(pdf_path).stem  # filename without extension
+
+    try:
+        doc = fitz.open(pdf_path)
+        page_count = len(doc)
+        doc.close()
+
+        if page_count == 0:
+            print(f"Warning: {pdf_path} has no pages")
+            return []
+
+        pages = []
+        for i in range(page_count):
+            page_label = f"{pdf_name}_p{i+1}"
+            pages.append((pdf_path, i, page_label))
+
+        return pages
+
+    except Exception as e:
+        print(f"Error reading PDF {pdf_path}: {e}")
+        return []
+
+
+def _extract_page_as_pdf_bytes(pdf_path, page_index):
+    """
+    Extract a single page from a PDF as PDF bytes.
+    Used for providers with native PDF support (Anthropic, Google).
+    """
+    import fitz  # PyMuPDF
+
+    try:
+        doc = fitz.open(pdf_path)
+        page = doc[page_index]
+
+        # Create a new PDF with just this page
+        new_doc = fitz.open()
+        new_doc.insert_pdf(doc, from_page=page_index, to_page=page_index)
+
+        pdf_bytes = new_doc.tobytes()
+        new_doc.close()
+        doc.close()
+
+        return pdf_bytes, True
+
+    except Exception as e:
+        print(f"Error extracting page {page_index} from {pdf_path}: {e}")
+        return None, False
+
+
+def _extract_page_as_image_bytes(pdf_path, page_index, dpi=150):
+    """
+    Extract a single page from a PDF as PNG image bytes.
+    Used for providers without native PDF support (OpenAI, Mistral, etc.).
+    """
+    import fitz  # PyMuPDF
+
+    try:
+        doc = fitz.open(pdf_path)
+        page = doc[page_index]
+
+        # Render page to image
[…]
