### 2025-12-30 5daf326
added paramter to switch pdf classification from image to text mode, depending on what the user's goal is
