+++ src/catllm/pdf_functions.py
@@ -134,6 +134,29 @@ def _encode_bytes_to_base64(data_bytes):
     return base64.b64encode(data_bytes).decode("utf-8")
 
 
+def _extract_page_text(pdf_path, page_index):
+    """
+    Extract text content from a single PDF page.
+    Used for text-based processing mode.
+    """
+    import fitz  # PyMuPDF
+
+    try:
+        doc = fitz.open(pdf_path)
+        page = doc[page_index]
+        text = page.get_text("text")
+        doc.close()
+
+        if not text.strip():
+            return None, False, "Page contains no extractable text"
+
+        return text.strip(), True, None
+
+    except Exception as e:
+        print(f"Error extracting text from page {page_index} of {pdf_path}: {e}")
+        return None, False, str(e)
+
+
 # PDF multi-class (binary) function
 def pdf_multi_class(
     pdf_description,
@@ -141,6 +164,7 @@ def pdf_multi_class(
     categories,
     api_key,
     user_model="gpt-4o",
+    mode="image",
     creativity=None,
     safety=False,
     chain_of_verification=False,
@@ -170,6 +194,11 @@ def pdf_multi_class(
         categories (list): List of category names for classification.
         api_key (str): API key for the model provider.
         user_model (str): Model name to use. Default "gpt-4o".
+        mode (str): How to process PDF pages. Options:
+            - "image": Render pages as images (best for visual elements like charts/tables)
+            - "text": Extract text only (best for text-heavy documents, faster/cheaper)
+            - "both": Send both text and image (most comprehensive but slower/costlier)
+            Default is "image".
         creativity (float): Temperature setting. None uses model default.
         safety (bool): If True, saves progress after each page.
         chain_of_verification (bool): Enable Chain of Verification for accuracy.
@@ -193,11 +222,21 @@ def pdf_multi_class(
 
     Example:
         >>> import catllm as cat
+        >>> # Image mode (default) - good for documents with charts/tables
         >>> results = cat.pdf_multi_class(
         ...     pdf_description="financial reports",
         ...     pdf_input="/path/to/pdfs/",
         ...     categories=["has_chart", "has_table", "is_summary"],
-        ...     api_key="your-api-key"
+        ...     api_key="your-api-key",
+        ...     mode="image"
+        ... )
+        >>> # Text mode - good for text-heavy documents, faster and cheaper
+        >>> results = cat.pdf_multi_class(
+        ...     pdf_description="research papers",
+        ...     pdf_input="/path/to/pdfs/",
+        ...     categories=["discusses_methodology", "has_results"],
+        ...     api_key="your-api-key",
+        ...     mode="text"
         ... )
     """
     import os
@@ -210,11 +249,17 @@ def pdf_multi_class(
     if save_directory is not None and not os.path.isdir(save_directory):
         raise FileNotFoundError(f"Directory {save_directory} doesn't exist")
 
+    # Validate mode parameter
+    mode = mode.lower()
+    if mode not in {"image", "text", "both"}:
+        raise ValueError(f"mode must be 'image', 'text', or 'both', got: {mode}")
+
     model_source = _detect_model_source(user_model, model_source)
 
-    # Providers with native PDF support
+    # Providers with native PDF support (only used in image/both modes)
     native_pdf_providers = {"anthropic", "google"}
-    needs_image_conversion = model_source not in native_pdf_providers
+
+    print(f"Processing mode: {mode}")
 
     # Load PDF files
     pdf_files = _load_pdf_files(pdf_input)
@@ -281,12 +326,20 @@ Provide a brief analysis of what content cues to look for when categorizing such
     link1 = []
     extracted_jsons = []
 
-    def _build_base_prompt_text():
[…]
