+++ README.md
@@ -19,6 +19,8 @@ CatLLM: A Reproducible LLM Pipeline for Coding Open-Ended Survey Responses
   - [explore_corpus()](#explore_corpus)
   - [explore_common_categories()](#explore_common_categories)
   - [multi_class()](#multi_class)
+  - [image_multi_class()](#image_multi_class)
+  - [pdf_multi_class()](#pdf_multi_class)
   - [image_score()](#image_score_drawing)
   - [image_features()](#image_features)
   - [cerad_drawn_score()](#cerad_drawn_score)
@@ -82,17 +84,16 @@ Whether you're working with messy text responses or analyzing visual content, Ca
 - **Perplexity**: Sonnar Large, Sonnar Small, etc.
 - **Mistral**: Mistral Large, Mistral Small, etc.
 
-**Fully Tested (Beta):**
-- ✅ OpenAI (GPT-4, GPT-4o, GPT-3.5-turbo, etc.)
-- ✅ Anthropic (Claude 3.5 Sonnet, Haiku)
+**Fully Tested:**
+- ✅ OpenAI (GPT-4, GPT-4o, GPT-5, etc.)
+- ✅ Anthropic (Claude Sonnet 4, Claude 3.5 Sonnet, Haiku)
 - ✅ Perplexity (Sonar models)
 - ✅ Google Gemini - Free tier has severe rate limits (5 RPM). Requires Google AI Studio billing account for large-scale use.
+- ✅ Huggingface - Access to Qwen, Llama 4, DeepSeek, and thousands of user-trained models for specific tasks. API routing can occasionally be unstable.
+- ✅ xAI (Grok models)
+- ✅ Mistral (Mistral Large, Pixtral, etc.)
 
-**Supported but Limited:**
-- 
-- ⚠️ Huggingface - API routing can be unstable
-
-**Note:** For beta testing, I recommend starting with OpenAI or Anthropic.
+**Note:** For best results, I recommend starting with OpenAI or Anthropic.
 
 
 ## API Reference
@@ -276,6 +277,96 @@ image_categories = cat.image_multi_class(
     chain_of_thought=True,
     safety=True,
     api_key="OPENAI_API_KEY")
+```
+
+### `pdf_multi_class()`
+
+Performs multi-label classification of PDF pages into user-defined categories, returning structured results with optional CSV export. Each page of each PDF is processed separately.
+
+**Installation:**
+```console
+pip install cat-llm[pdf]
+```
+Requires PyMuPDF for PDF processing.
+
+**Methodology:**
+Processes each PDF page individually, assigning one or more categories from the provided list. Pages are labeled as `{filename}_p{page_number}` (e.g., "report_p1", "report_p2"). Supports three processing modes for different document types and includes advanced prompting techniques for improved accuracy.
+
+**Parameters:**
+- `pdf_description` (str): A description of what the PDF documents contain
+- `pdf_input` (str or list): Directory path containing PDFs, or list of PDF file paths
+- `categories` (list): List of predefined categories for classification
+- `api_key` (str): API key for the LLM service
+- `user_model` (str, default="gpt-4o"): Specific model to use
+- `mode` (str, default="image"): How to process PDF pages:
+  - `"image"`: Render pages as images (best for visual elements like charts/tables)
+  - `"text"`: Extract text only (faster/cheaper for text-heavy documents)
+  - `"both"`: Send both image and extracted text (most comprehensive)
+- `creativity` (float, optional): Temperature/randomness setting (0.0-1.0)
+- `safety` (bool, default=False): Enable safety checks and save results at each API call step
+- `chain_of_verification` (bool, default=False): Enable Chain-of-Verification prompting - re-examines pages to verify categorization accuracy. **⚠️ Warning: CoVe consumes significantly more tokens (3-5x).**
[…]
