### 2026-01-02 3ac8d7c
Add model.change() handler to update API key status dynamically
