+++ hf_space/app.py
@@ -1,0 +1,283 @@
+"""
+Gradio app - Step 5g: Add actual catllm classification
+"""
+
+import gradio as gr
+import pandas as pd
+import tempfile
+import os
+
+# Import catllm
+try:
+    import catllm
+    CATLLM_AVAILABLE = True
+except ImportError as e:
+    print(f"Warning: Could not import catllm: {e}")
+    CATLLM_AVAILABLE = False
+
+MAX_CATEGORIES = 10
+INITIAL_CATEGORIES = 3
+
+MODEL_CHOICES = [
+    "Qwen/Qwen3-VL-235B-A22B-Instruct:novita (Free)",
+    "deepseek-ai/DeepSeek-V3.1:novita (Free)",
+    "meta-llama/Llama-4-Maverick-17B-128E-Instruct:groq (Free)",
+    "gpt-4o",
+    "claude-sonnet-4-5-20250929",
+    "gemini-2.5-flash",
+]
+
+HF_FREE_MODELS = {
+    "Qwen/Qwen3-VL-235B-A22B-Instruct:novita (Free)": "Qwen/Qwen3-VL-235B-A22B-Instruct:novita",
+    "deepseek-ai/DeepSeek-V3.1:novita (Free)": "deepseek-ai/DeepSeek-V3.1:novita",
+    "meta-llama/Llama-4-Maverick-17B-128E-Instruct:groq (Free)": "meta-llama/Llama-4-Maverick-17B-128E-Instruct:groq",
+}
+
+
+def is_free_model(model):
+    return "(Free)" in model
+
+
+def get_model_source(model):
+    """Auto-detect model source. All HF router models (novita, groq, etc) use 'huggingface'."""
+    model_lower = model.lower()
+    if "gpt" in model_lower:
+        return "openai"
+    elif "claude" in model_lower:
+        return "anthropic"
+    elif "gemini" in model_lower:
+        return "google"
+    elif "mistral" in model_lower and ":novita" not in model_lower:
+        return "mistral"
+    # All models routed through HuggingFace (including novita, groq variants)
+    elif any(x in model_lower for x in [":novita", ":groq", "qwen", "llama", "deepseek"]):
+        return "huggingface"
+    return "huggingface"
+
+
+def load_columns(file):
+    if file is None:
+        return gr.update(choices=[], value=None), "Please upload a file first"
+
+    try:
+        file_path = file if isinstance(file, str) else file.name
+        if file_path.endswith('.csv'):
+            df = pd.read_csv(file_path)
+        else:
+            df = pd.read_excel(file_path)
+
+        columns = df.columns.tolist()
+        return (
+            gr.update(choices=columns, value=columns[0] if columns else None),
+            f"Loaded {len(df)} rows. Select column and click Classify."
+        )
+    except Exception as e:
+        return gr.update(choices=[], value=None), f"**Error:** {str(e)}"
+
+
+def classify_data(input_type, spreadsheet_file, spreadsheet_column,
+                  cat1, cat2, cat3, cat4, cat5, cat6, cat7, cat8, cat9, cat10,
+                  model, model_source_input, api_key_input):
+    """Main classification function."""
+    if not CATLLM_AVAILABLE:
+        return None, None, "**Error:** catllm package not available"
+
+    all_cats = [cat1, cat2, cat3, cat4, cat5, cat6, cat7, cat8, cat9, cat10]
+    categories = [c.strip() for c in all_cats if c and c.strip()]
+
+    if not categories:
+        return None, None, "**Error:** Please enter at least one category"
+
+    # Get API key - priority: user input > environment variable
+    if is_free_model(model):
+        actual_api_key = os.environ.get("HF_API_KEY", "")
[…]
