+++ src/catllm/__init__.py
@@ -15,3 +15,4 @@ from .text_functions import *
 from .CERAD_functions import *
 from .image_functions import *
 from .pdf_functions import *
+from .main import *
