### 2026-01-02 75766f5
Add image classification support with file upload and description
