### 2026-01-02 3c8836f
Simplify app to focus on survey response classification only
