+++ hf_space/app.py
@@ -55,15 +55,6 @@ def get_model_source(model):
     return "huggingface"
 
 
-def toggle_input_type(input_type):
-    """Show/hide input groups based on selected type."""
-    return (
-        gr.update(visible=(input_type == "Spreadsheet")),  # spreadsheet_group
-        gr.update(visible=(input_type == "Image")),        # image_group
-        gr.update(visible=(input_type == "Spreadsheet")),  # column_row (for spreadsheet)
-    )
-
-
 def load_columns(file):
     if file is None:
         return gr.update(choices=[], value=None), "Please upload a file first"
@@ -84,8 +75,7 @@ def load_columns(file):
         return gr.update(choices=[], value=None), f"**Error:** {str(e)}"
 
 
-def classify_data(input_type, spreadsheet_file, spreadsheet_column,
-                  image_files_input, image_desc,
+def classify_data(spreadsheet_file, spreadsheet_column,
                   cat1, cat2, cat3, cat4, cat5, cat6, cat7, cat8, cat9, cat10,
                   model, model_source_input, api_key_input):
     """Main classification function."""
@@ -130,69 +120,36 @@ def classify_data(input_type, spreadsheet_file, spreadsheet_column,
         model_source = model_source_input
 
     try:
-        if input_type == "Spreadsheet":
-            if not spreadsheet_file:
-                return None, None, "**Error:** Please upload a spreadsheet"
-            if not spreadsheet_column:
-                return None, None, "**Error:** Please select a column to classify"
+        if not spreadsheet_file:
+            return None, None, "**Error:** Please upload a file"
+        if not spreadsheet_column:
+            return None, None, "**Error:** Please select a column to classify"
 
-            file_path = spreadsheet_file if isinstance(spreadsheet_file, str) else spreadsheet_file.name
-            if file_path.endswith('.csv'):
-                df = pd.read_csv(file_path)
-            else:
-                df = pd.read_excel(file_path)
-
-            if spreadsheet_column not in df.columns:
-                return None, None, f"**Error:** Column '{spreadsheet_column}' not found"
-
-            input_data = df[spreadsheet_column].tolist()
-
-            result = catllm.multi_class(
-                survey_input=input_data,
-                categories=categories,
-                api_key=actual_api_key,
-                user_model=actual_model,
-                model_source=model_source
-            )
-
-            # Return the result directly from catllm (already properly formatted)
-            # Save for download
-            with tempfile.NamedTemporaryFile(mode='w', suffix='_classified.csv', delete=False) as f:
-                result.to_csv(f.name, index=False)
-                download_path = f.name
-
-            return result, download_path, f"**Success!** Classified {len(input_data)} items"
-
-        elif input_type == "Image":
-            if not image_files_input:
-                return None, None, "**Error:** Please upload at least one image"
-            if not image_desc:
-                return None, None, "**Error:** Please provide an image description"
-
-            # Get file paths from uploaded files
-            image_paths = []
-            for f in image_files_input:
-                file_path = f if isinstance(f, str) else f.name
-                image_paths.append(file_path)
-
-            result = catllm.image_multi_class(
-                image_description=image_desc,
-                image_input=image_paths,
-                categories=categories,
-                api_key=actual_api_key,
-                user_model=actual_model,
-                model_source=model_source
-            )
-
-            # Save for download
-            with tempfile.NamedTemporaryFile(mode='w', suffix='_classified.csv', delete=False) as f:
-                result.to_csv(f.name, index=False)
-                download_path = f.name
-
[…]
