+++ hf_space/app.py
@@ -6,6 +6,7 @@ import gradio as gr
 import pandas as pd
 import tempfile
 import os
+from datetime import datetime
 
 # Import catllm
 try:
@@ -18,24 +19,135 @@ except ImportError as e:
 MAX_CATEGORIES = 10
 INITIAL_CATEGORIES = 3
 
-MODEL_CHOICES = [
-    "Qwen/Qwen3-VL-235B-A22B-Instruct:novita (Free)",
-    "deepseek-ai/DeepSeek-V3.1:novita (Free)",
-    "meta-llama/Llama-4-Maverick-17B-128E-Instruct:groq (Free)",
-    "gpt-4o",
-    "claude-sonnet-4-5-20250929",
+# Free models (uses Space secrets - no user API key needed)
+FREE_MODEL_CHOICES = [
+    "Qwen/Qwen3-VL-235B-A22B-Instruct:novita",
+    "deepseek-ai/DeepSeek-V3.1:novita",
+    "meta-llama/Llama-4.1-8B-Instruct:novita",
     "gemini-2.5-flash",
+    "gpt-4o",
+    "mistral-medium-2505",
+    "claude-3-haiku-20240307",
+    "sonar",
+    "grok-4-fast-non-reasoning",
 ]
 
-HF_FREE_MODELS = {
-    "Qwen/Qwen3-VL-235B-A22B-Instruct:novita (Free)": "Qwen/Qwen3-VL-235B-A22B-Instruct:novita",
-    "deepseek-ai/DeepSeek-V3.1:novita (Free)": "deepseek-ai/DeepSeek-V3.1:novita",
-    "meta-llama/Llama-4-Maverick-17B-128E-Instruct:groq (Free)": "meta-llama/Llama-4-Maverick-17B-128E-Instruct:groq",
-}
+# Paid models (user provides their own API key)
+PAID_MODEL_CHOICES = [
+    "gpt-4.1",
+    "gpt-4o",
+    "gpt-4o-mini",
+    "claude-sonnet-4-5-20250929",
+    "claude-opus-4-20250514",
+    "claude-3-5-haiku-20241022",
+    "gemini-2.5-pro",
+    "gemini-2.5-flash",
+    "mistral-large-latest",
+]
+
+# Models routed through HuggingFace
+HF_ROUTED_MODELS = [
+    "Qwen/Qwen3-VL-235B-A22B-Instruct:novita",
+    "deepseek-ai/DeepSeek-V3.1:novita",
+    "meta-llama/Llama-4.1-8B-Instruct:novita",
+]
 
 
-def is_free_model(model):
-    return "(Free)" in model
+def is_free_model(model, model_tier):
+    """Check if using free tier (Space pays for API)."""
+    return model_tier == "Free Models"
+
+
+def generate_codebook_pdf(categories, model, column_name, num_rows):
+    """Generate a PDF codebook explaining the output columns."""
+    from reportlab.lib.pagesizes import letter
+    from reportlab.lib import colors
+    from reportlab.lib.styles import getSampleStyleSheet, ParagraphStyle
+    from reportlab.platypus import SimpleDocTemplate, Paragraph, Spacer, Table, TableStyle
+
+    # Create temp file for PDF
+    pdf_file = tempfile.NamedTemporaryFile(mode='wb', suffix='_codebook.pdf', delete=False)
+    doc = SimpleDocTemplate(pdf_file.name, pagesize=letter)
+    styles = getSampleStyleSheet()
+
+    # Custom styles
+    title_style = ParagraphStyle('Title', parent=styles['Heading1'], fontSize=18, spaceAfter=20)
[…]
