+++ hf_space/app.py
@@ -6,7 +6,10 @@ import gradio as gr
 import pandas as pd
 import tempfile
 import os
+import time
+import sys
 from datetime import datetime
+import matplotlib.pyplot as plt
 
 # Import catllm
 try:
@@ -58,15 +61,17 @@ def is_free_model(model, model_tier):
     return model_tier == "Free Models"
 
 
-def generate_codebook_pdf(categories, model, column_name, num_rows):
-    """Generate a PDF codebook explaining the output columns."""
+def generate_methodology_report_pdf(categories, model, column_name, num_rows, model_source, filename, success_rate,
+                          result_df=None, processing_time=None, prompt_template=None,
+                          data_quality=None, catllm_version=None, python_version=None):
+    """Generate a PDF methodology report for reproducibility and transparency."""
     from reportlab.lib.pagesizes import letter
     from reportlab.lib import colors
     from reportlab.lib.styles import getSampleStyleSheet, ParagraphStyle
-    from reportlab.platypus import SimpleDocTemplate, Paragraph, Spacer, Table, TableStyle
+    from reportlab.platypus import SimpleDocTemplate, Paragraph, Spacer, Table, TableStyle, PageBreak
 
     # Create temp file for PDF
-    pdf_file = tempfile.NamedTemporaryFile(mode='wb', suffix='_codebook.pdf', delete=False)
+    pdf_file = tempfile.NamedTemporaryFile(mode='wb', suffix='_methodology_report.pdf', delete=False)
     doc = SimpleDocTemplate(pdf_file.name, pagesize=letter)
     styles = getSampleStyleSheet()
 
@@ -74,35 +79,31 @@ def generate_codebook_pdf(categories, model, column_name, num_rows):
     title_style = ParagraphStyle('Title', parent=styles['Heading1'], fontSize=18, spaceAfter=20)
     heading_style = ParagraphStyle('Heading', parent=styles['Heading2'], fontSize=14, spaceAfter=10, spaceBefore=15)
     normal_style = styles['Normal']
+    code_style = ParagraphStyle('Code', parent=styles['Normal'], fontName='Courier', fontSize=9, leftIndent=20, spaceAfter=3)
 
     story = []
 
-    # Title
-    story.append(Paragraph("CatLLM Classification Codebook", title_style))
+    # === PAGE 1: Title, About, Category Mapping ===
+    story.append(Paragraph("CatLLM Methodology Report", title_style))
     story.append(Paragraph(f"Generated: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}", normal_style))
-    story.append(Spacer(1, 20))
+    story.append(Spacer(1, 15))
 
-    # Classification summary
-    story.append(Paragraph("Classification Summary", heading_style))
-    summary_data = [
-        ["Source Column", column_name],
-        ["Model Used", model],
-        ["Rows Classified", str(num_rows)],
-        ["Number of Categories", str(len(categories))],
-    ]
-    summary_table = Table(summary_data, colWidths=[150, 300])
-    summary_table.setStyle(TableStyle([
-        ('BACKGROUND', (0, 0), (0, -1), colors.lightgrey),
-        ('GRID', (0, 0), (-1, -1), 1, colors.black),
-        ('PADDING', (0, 0), (-1, -1), 8),
-    ]))
-    story.append(summary_table)
-    story.append(Spacer(1, 20))
+    # About CatLLM - addressing prompt hacking
+    story.append(Paragraph("About This Report", heading_style))
+    about_text = """This methodology report documents the classification process for reproducibility and transparency. \
+CatLLM addresses an issue identified by researchers in "Prompt-Hacking: The New p-Hacking?" (Kosch &amp; Feger, 2025; \
[…]
