+++ hf_space/requirements.txt
@@ -7,3 +7,4 @@ openpyxl
 requests
 regex
 reportlab
+matplotlib
