+++ hf_space/app.py
@@ -443,7 +443,8 @@ def load_columns(file):
         return gr.update(choices=[], value=None), f"**Error:** {str(e)}"
 
 
-def classify_data(spreadsheet_file, spreadsheet_column,
+def classify_data(input_type, spreadsheet_file, spreadsheet_column,
+                  pdf_file, pdf_description, pdf_mode,
                   cat1, cat2, cat3, cat4, cat5, cat6, cat7, cat8, cat9, cat10,
                   model_tier, model, model_source_input, api_key_input):
     """Main classification function with progress updates. Yields status updates then final results."""
@@ -515,53 +516,132 @@ def classify_data(spreadsheet_file, spreadsheet_column,
         model_source = model_source_input
 
     try:
-        if not spreadsheet_file:
-            yield None, None, None, None, "**Error:** Please upload a file"
-            return
-        if not spreadsheet_column:
-            yield None, None, None, None, "**Error:** Please select a column to classify"
-            return
+        # Determine if we're processing text or PDF
+        is_pdf_mode = input_type == "PDF Documents"
 
-        file_path = spreadsheet_file if isinstance(spreadsheet_file, str) else spreadsheet_file.name
-        if file_path.endswith('.csv'):
-            df = pd.read_csv(file_path)
+        if is_pdf_mode:
+            # PDF validation
+            if not pdf_file:
+                yield None, None, None, None, "**Error:** Please upload a PDF file"
+                return
+
+            pdf_path = pdf_file if isinstance(pdf_file, str) else pdf_file.name
+
+            # Map UI mode to function parameter
+            mode_mapping = {
+                "Image (visual documents)": "image",
+                "Text (text-heavy)": "text",
+                "Both (comprehensive)": "both"
+            }
+            actual_pdf_mode = mode_mapping.get(pdf_mode, "image")
+
+            # Progress update
+            yield None, None, None, None, f"⏳ **Loading PDF...** Processing document."
+
+            # Data quality placeholder for PDFs
+            data_quality = {
+                'null_count': 0,
+                'avg_length': 0,
+                'min_length': 0,
+                'max_length': 0,
+                'error_count': 0
+            }
+
+            # Progress update: starting classification
+            yield None, None, None, None, f"🔄 **Classifying PDF pages...** This may take a moment."
+
+            # Capture timing
+            start_time = time.time()
+
+            result = catllm.pdf_multi_class(
+                pdf_description=pdf_description or "document",
+                pdf_input=pdf_path,
+                categories=categories,
+                api_key=actual_api_key,
+                user_model=actual_model,
+                model_source=model_source,
+                mode=actual_pdf_mode
+            )
+
+            processing_time = time.time() - start_time
+            num_items = len(result)
+            original_filename = pdf_path.split("/")[-1]
+            column_name = "PDF Pages"
+
+            # Build prompt template for PDF
+            prompt_template = f'''Categorize this PDF page from "{pdf_description or 'document'}" into the following categories that apply:
+{{categories}}
+
+Let's think step by step:
+1. First, identify the main themes present in this page
+2. Then, match each theme to the relevant categories
+3. Finally, assign 1 to matching categories and 0 to non-matching categories
+
+Provide your work in JSON format where the number belonging to each category is the key and a 1 if the category is present and a 0 if it is not present as key values.'''
+
         else:
-            df = pd.read_excel(file_path)
+            # Text data validation
+            if not spreadsheet_file:
+                yield None, None, None, None, "**Error:** Please upload a file"
+                return
+            if not spreadsheet_column:
+                yield None, None, None, None, "**Error:** Please select a column to classify"
+                return
 
-        if spreadsheet_column not in df.columns:
[…]
