### 2026-01-03 b8fd572
aligning the pdf function with the text one
