### 2026-01-03 4f8c9c6
fixed glitch that would cause errors in app using image classifcation, same output as other classifiers
