### 2026-01-03 31355fb
adding extract categories to pdf and image functions
