+++ src/catllm/pdf_functions.py
@@ -198,7 +198,8 @@ def pdf_multi_class(
     Args:
         pdf_description (str): Description of the PDF documents being categorized.
         pdf_input (str or list): Directory path containing PDFs, or list of PDF file paths.
-        categories (list): List of category names for classification.
+        categories (list or "auto"): List of category names for classification,
+            or "auto" to automatically extract categories from the PDFs first.
         api_key (str): API key for the model provider.
         user_model (str): Model name to use. Default "gpt-4o".
         mode (str): How to process PDF pages. Options:
@@ -278,6 +279,24 @@ def pdf_multi_class(
         all_pages.extend(pages)
 
     print(f"Total pages to process: {len(all_pages)}")
+
+    # Handle "auto" categories - extract categories first
+    if categories == "auto":
+        if not pdf_description:
+            raise ValueError("pdf_description is required when using categories='auto'")
+
+        print("\nAuto-extracting categories from PDFs...")
+        auto_result = explore_pdf_categories(
+            pdf_input=pdf_input,
+            api_key=api_key,
+            pdf_description=pdf_description,
+            user_model=user_model,
+            mode=mode,
+            model_source=model_source,
+            creativity=creativity
+        )
+        categories = auto_result["top_categories"]
+        print(f"Extracted {len(categories)} categories: {categories}\n")
 
     categories_str = "\n".join(f"{i + 1}. {cat}" for i, cat in enumerate(categories))
     cat_num = len(categories)
@@ -1103,21 +1122,538 @@ Provide the final categorization in the same JSON format:"""
     )
     categorized_data[cat_cols] = categorized_data[cat_cols].astype('Int64')
 
-    # Create categories_id (list of category numbers where value=1)
-    def get_category_ids(row):
-        ids = []
-        for col in cat_cols:
-            val = row[col]
-            if pd.notna(val) and val == 1:
-                # Extract category number from column name (e.g., "category_1" -> "1")
-                cat_num = col.replace('category_', '')
-                ids.append(cat_num)
-        return ','.join(ids)
-
-    categorized_data['categories_id'] = categorized_data.apply(get_category_ids, axis=1)
+    # Create categories_id (comma-separated binary values for each category)
+    categorized_data['categories_id'] = categorized_data[cat_cols].apply(
+        lambda x: ','.join(x.dropna().astype(int).astype(str)), axis=1
+    )
 
     if filename:
         save_path = os.path.join(save_directory, filename) if save_directory else filename
         categorized_data.to_csv(save_path, index=False)
 
     return categorized_data
+
+
+def explore_pdf_categories(
+    pdf_input,
+    api_key,
+    pdf_description="",
+    max_categories=12,
+    categories_per_chunk=10,
+    divisions=5,
+    user_model="gpt-4o",
+    creativity=None,
+    specificity="broad",
+    research_question=None,
+    mode="text",
+    filename=None,
+    model_source="auto",
+    iterations=3,
+    random_state=None
+):
+    """
+    Explore and extract common categories from PDF pages.
+
+    Modes:
+        - "text" (default): Extracts text from pages, concatenates pages within
+          each chunk, and sends combined text to identify categories. Similar to
+          how explore_common_categories works with survey responses. Best for
+          text-heavy documents.
+
+        - "image": Samples random pages from the full pool of all pages across
+          all PDFs and sends them as images to a vision model. Best for visual
+          documents where layout matters.
+
+        - "both": Samples random pages, uses vision model to describe each page's
+          content (text + visual elements), then extracts categories from those
+          descriptions. Best for documents with mixed text and visual content
+          (charts, diagrams, scanned documents).
+
+    Args:
+        pdf_input: Path to PDF file, directory of PDFs, or list of PDF paths
[…]
