### 2026-01-03 ec9d464
Auto-adjust divisions and categories_per_chunk for small datasets
- Replace ValueError with auto-adjustment for all explore functions
- text_functions: at least 3 responses per chunk
- image_functions: at least 2 images per chunk (images have multiple categories)
- pdf_functions: at least 2 pages per chunk
- Print informative messages when auto-adjusting
- Bump version to 0.1.5

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
