+++ src/catllm/text_functions.py
@@ -98,16 +98,24 @@ def explore_corpus(
     print(f"Exploring class for question: '{survey_question}'.\n          {categories_per_chunk * divisions} unique categories to be extracted.")
     print()
 
-    model_source = model_source.lower() # eliminating case sensitivity 
+    model_source = model_source.lower() # eliminating case sensitivity
 
-    chunk_size = round(max(1, len(survey_input) / divisions),0)
+    n = len(survey_input)
+
+    # Auto-adjust divisions for small datasets
+    original_divisions = divisions
+    divisions = min(divisions, max(1, n // 3))  # At least 3 responses per chunk
+    if divisions != original_divisions:
+        print(f"Auto-adjusted divisions from {original_divisions} to {divisions} for {n} responses.")
+
+    chunk_size = round(max(1, n / divisions),0)
     chunk_size = int(chunk_size)
 
     if chunk_size < (categories_per_chunk/2):
-        raise ValueError(f"Cannot extract {categories_per_chunk} {specificity} categories from chunks of only {chunk_size} responses. \n" 
-                    f"Choose one solution: \n"
-                    f"(1) Reduce 'divisions' parameter (currently {divisions}) to create larger chunks, or \n"
-                    f"(2) Reduce 'categories_per_chunk' parameter (currently {categories_per_chunk}) to extract fewer categories per chunk.")
+        # Auto-reduce categories_per_chunk instead of erroring
+        old_categories_per_chunk = categories_per_chunk
+        categories_per_chunk = max(3, chunk_size * 2)
+        print(f"Auto-adjusted categories_per_chunk from {old_categories_per_chunk} to {categories_per_chunk} for chunk size {chunk_size}.")
 
     random_chunks = []
     for i in range(divisions):
@@ -208,15 +216,19 @@ def explore_common_categories(
     if n == 0:
         raise ValueError("survey_input is empty after dropping NA.")
 
+    # --- auto-adjust divisions for small datasets ---
+    original_divisions = divisions
+    divisions = min(divisions, max(1, n // 3))  # At least 3 responses per chunk
+    if divisions != original_divisions:
+        print(f"Auto-adjusted divisions from {original_divisions} to {divisions} for {n} responses.")
+
     # --- chunk sizing ---
     chunk_size = int(round(max(1, n / divisions), 0))
     if chunk_size < (categories_per_chunk / 2):
-        raise ValueError(
-            f"Cannot extract {categories_per_chunk} categories from chunks of only {chunk_size} responses.\n"
-            f"Solutions:\n"
-            f"  (1) Reduce 'divisions' (currently {divisions}) to make larger chunks, or\n"
-            f"  (2) Reduce 'categories_per_chunk' (currently {categories_per_chunk})."
-        )
+        # Auto-reduce categories_per_chunk instead of erroring
+        old_categories_per_chunk = categories_per_chunk
+        categories_per_chunk = max(3, chunk_size * 2)
+        print(f"Auto-adjusted categories_per_chunk from {old_categories_per_chunk} to {categories_per_chunk} for chunk size {chunk_size}.")
 
     print(
         f"Exploring class for question: '{survey_question}'.\n"
