### 2026-01-03 d59c574
PDFs and images can have multiple categories per item
- Updated comments to clarify that PDFs/images yield many categories each
- Less aggressive reduction of categories_per_chunk (only when chunk_size < 2)
- Use max(5, chunk_size * 4) instead of max(3, chunk_size * 2)

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
