+++ src/catllm/pdf_functions.py
@@ -1215,17 +1215,19 @@ def explore_pdf_categories(
         raise ValueError("No pages found in the PDF files.")
 
     # Auto-adjust divisions for small datasets
+    # PDF pages can have multiple categories each, so we can use fewer divisions
     original_divisions = divisions
     divisions = min(divisions, max(1, n // 2))  # At least 2 pages per chunk
     if divisions != original_divisions:
         print(f"Auto-adjusted divisions from {original_divisions} to {divisions} for {n} pages.")
 
-    # Chunk sizing
+    # Chunk sizing - PDF pages often contain multiple categories each
     chunk_size = int(round(max(1, n / divisions), 0))
-    if chunk_size < (categories_per_chunk / 2):
-        # Auto-reduce categories_per_chunk instead of erroring
+    # Don't reduce categories_per_chunk as aggressively for PDFs since each page can yield many categories
+    if chunk_size < 2:
+        # Only reduce if we have very few pages
         old_categories_per_chunk = categories_per_chunk
-        categories_per_chunk = max(3, chunk_size * 2)
+        categories_per_chunk = max(5, chunk_size * 4)
         print(f"Auto-adjusted categories_per_chunk from {old_categories_per_chunk} to {categories_per_chunk} for chunk size {chunk_size}.")
 
     print(
