### 2026-01-03 9fec554
Add mobile-responsive CSS for Hugging Face app
- Add media queries for tablets (max-width: 768px) and phones (max-width: 480px)
- Reduce upload button sizes and padding on mobile devices
- Stack columns vertically on smaller screens
- Compact file upload areas, form inputs, and buttons
- Reduce font sizes and heading scales for mobile
- Shrink logo size on mobile devices
- Make task buttons stack vertically on very small screens
- Add responsive styling for dataframes, plots, and accordions
