### 2026-01-04 80b3146
web app made mobile friendly
