+++ hf_space/app.py
@@ -21,6 +21,167 @@ except ImportError as e:
 
 MAX_CATEGORIES = 10
 INITIAL_CATEGORIES = 3
+MAX_FILE_SIZE_MB = 100  # Warn users if total file size exceeds this
+
+
+def calculate_total_file_size(files):
+    """Calculate total size of uploaded files in MB."""
+    if files is None:
+        return 0
+    if not isinstance(files, list):
+        files = [files]
+
+    total_bytes = 0
+    for f in files:
+        try:
+            file_path = f if isinstance(f, str) else f.name
+            total_bytes += os.path.getsize(file_path)
+        except (OSError, AttributeError):
+            pass
+    return total_bytes / (1024 * 1024)  # Convert to MB
+
+
+def generate_extract_code(input_type, description, model, model_source, max_categories, mode=None):
+    """Generate Python code for category extraction."""
+    if input_type == "text":
+        return f'''import catllm
+import pandas as pd
+
+# Load your data
+df = pd.read_csv("your_data.csv")
+
+# Extract categories from the text column
+result = catllm.extract(
+    input_data=df["{description}"].tolist(),
+    api_key="YOUR_API_KEY",
+    input_type="text",
+    description="{description}",
+    user_model="{model}",
+    model_source="{model_source}",
+    max_categories={max_categories}
+)
+
+# View extracted categories
+print(result["top_categories"])
+print(result["counts_df"])
+'''
+    elif input_type == "pdf":
+        mode_line = f',\n    mode="{mode}"' if mode else ''
+        return f'''import catllm
+
+# Extract categories from PDF documents
+result = catllm.extract(
+    input_data="path/to/your/pdfs/",  # or list of PDF paths
+    api_key="YOUR_API_KEY",
+    input_type="pdf",
+    description="{description}"{mode_line},
+    user_model="{model}",
+    model_source="{model_source}",
+    max_categories={max_categories}
+)
+
+# View extracted categories
+print(result["top_categories"])
+print(result["counts_df"])
+'''
+    else:  # image
+        return f'''import catllm
+
+# Extract categories from images
+result = catllm.extract(
+    input_data="path/to/your/images/",  # or list of image paths
+    api_key="YOUR_API_KEY",
+    input_type="image",
+    description="{description}",
+    user_model="{model}",
+    model_source="{model_source}",
+    max_categories={max_categories}
+)
+
+# View extracted categories
+print(result["top_categories"])
+print(result["counts_df"])
+'''
+
+
+def generate_classify_code(input_type, description, categories, model, model_source, mode=None):
+    """Generate Python code for classification."""
+    categories_str = ",\n    ".join([f'"{cat}"' for cat in categories])
+
+    if input_type == "text":
+        return f'''import catllm
+import pandas as pd
+
+# Load your data
+df = pd.read_csv("your_data.csv")
+
+# Define categories
+categories = [
+    {categories_str}
+]
+
+# Classify the text data
+result = catllm.classify(
+    input_data=df["{description}"].tolist(),
+    categories=categories,
+    api_key="YOUR_API_KEY",
+    input_type="text",
+    description="{description}",
+    user_model="{model}",
+    model_source="{model_source}"
+)
+
+# View results
+print(result)
+result.to_csv("classified_results.csv", index=False)
+'''
+    elif input_type == "pdf":
+        mode_line = f',\n    mode="{mode}"' if mode else ''
+        return f'''import catllm
+
+# Define categories
+categories = [
+    {categories_str}
+]
+
+# Classify PDF documents
+result = catllm.classify(
+    input_data="path/to/your/pdfs/",  # or list of PDF paths
+    categories=categories,
+    api_key="YOUR_API_KEY",
+    input_type="pdf",
+    description="{description}"{mode_line},
[…]
