### 2026-01-04 a3bcfe9
Add unified HTTP-based multi-class text classification
Experimental module that uses raw HTTP requests for all LLM providers
instead of provider-specific SDKs. This reduces dependencies to just
`requests` while maintaining structured output support for each provider.
