### 2026-01-04 eab3f74
Add Ollama support for local model inference
- Added Ollama to provider config (localhost:11434 by default)
- Made api_key optional (not required for Ollama)
- Added set_ollama_endpoint() helper for custom host/port
- Ollama uses json_object mode (doesn't support strict json_schema)
