### 2026-01-04 45e4184
Add Ollama auto-detection and validation
- check_ollama_running(): Verify Ollama service is accessible
- list_ollama_models(): Get all installed Ollama models
- check_ollama_model(): Verify a specific model is available
- Auto-validates when provider="ollama" with helpful error messages
