### 2026-01-04 9b0af27
Auto-download Ollama models when not installed
When using provider="ollama", if the specified model isn't installed locally,
the function will automatically pull it from Ollama's registry with progress
feedback. Users just need Ollama running - no manual model downloads needed.
