### 2026-01-04 fb8e729
Improve Ollama error messages and download warnings
- Clear, formatted error when Ollama isn't running with setup instructions
- Show estimated download size before pulling models (from known sizes table)
- Show actual size during download with human-readable formatting
- Support Ctrl+C to cancel downloads
- Added size estimates for 25+ popular Ollama models
