+++ src/catllm/text_functions_unified.py
@@ -446,6 +446,75 @@ def check_ollama_model(model: str, host: str = "localhost", port: int = 11434) -
     )
 
 
+def _format_bytes(size_bytes: int) -> str:
+    """Format bytes into human-readable string."""
+    if size_bytes < 1024:
+        return f"{size_bytes} B"
+    elif size_bytes < 1024 ** 2:
+        return f"{size_bytes / 1024:.1f} KB"
+    elif size_bytes < 1024 ** 3:
+        return f"{size_bytes / (1024 ** 2):.1f} MB"
+    else:
+        return f"{size_bytes / (1024 ** 3):.2f} GB"
+
+
+# Common model sizes (approximate) for user reference
+OLLAMA_MODEL_SIZES = {
+    "llama3.2": "2.0 GB",
+    "llama3.2:1b": "1.3 GB",
+    "llama3.2:3b": "2.0 GB",
+    "llama3.1": "4.7 GB",
+    "llama3.1:8b": "4.7 GB",
+    "llama3.1:70b": "40 GB",
+    "llama3": "4.7 GB",
+    "llama2": "3.8 GB",
+    "mistral": "4.1 GB",
+    "mixtral": "26 GB",
+    "phi3": "2.2 GB",
+    "phi3:mini": "2.2 GB",
+    "gemma": "5.0 GB",
+    "gemma:2b": "1.7 GB",
+    "gemma:7b": "5.0 GB",
+    "gemma2": "5.4 GB",
+    "gemma2:2b": "1.6 GB",
+    "gemma2:9b": "5.4 GB",
+    "gemma2:27b": "16 GB",
+    "qwen2.5": "4.7 GB",
+    "qwen2.5:0.5b": "397 MB",
+    "qwen2.5:1.5b": "986 MB",
+    "qwen2.5:3b": "1.9 GB",
+    "qwen2.5:7b": "4.7 GB",
+    "deepseek-r1": "4.7 GB",
+    "codellama": "3.8 GB",
+    "codegemma": "5.0 GB",
+    "nomic-embed-text": "274 MB",
+}
+
+
+def get_ollama_model_size_estimate(model: str) -> str:
+    """
+    Get estimated download size for an Ollama model.
+
+    Args:
+        model: Model name
+
+    Returns:
+        Human-readable size estimate or "unknown"
+    """
+    model_lower = model.lower()
+
+    # Check exact match first
+    if model_lower in OLLAMA_MODEL_SIZES:
+        return OLLAMA_MODEL_SIZES[model_lower]
+
+    # Check base model name (without tag)
+    base_model = model_lower.split(":")[0]
+    if base_model in OLLAMA_MODEL_SIZES:
+        return OLLAMA_MODEL_SIZES[base_model]
+
+    return "unknown"
+
+
 def pull_ollama_model(model: str, host: str = "localhost", port: int = 11434) -> bool:
     """
     Pull/download a model in Ollama.
@@ -458,7 +527,15 @@ def pull_ollama_model(model: str, host: str = "localhost", port: int = 11434) ->
     Returns:
         True if model was pulled successfully, False otherwise
     """
-    print(f"Model '{model}' not found locally. Downloading from Ollama...")
+    # Get size estimate and warn user
+    size_estimate = get_ollama_model_size_estimate(model)
+
+    print(f"\n{'='*60}")
+    print(f"  Model '{model}' not found locally")
+    print(f"  Estimated download size: {size_estimate}")
+    print(f"  Downloading from Ollama registry...")
+    print(f"  (Press Ctrl+C to cancel)")
+    print(f"{'='*60}\n")
 
     try:
         # Ollama pull endpoint streams the response
[…]
