### 2026-01-04 53d0669
Add confirmation prompt before downloading Ollama models
- Asks user to confirm download with [y/N] prompt
- Shows model name and estimated size before asking
- Added auto_download parameter to skip confirmation if desired
- Handles Ctrl+C and EOF gracefully during prompt
