### 2026-01-04 15655cd
Add system resource checks before downloading Ollama models
- Check available disk space vs model size (blocks if insufficient)
- Estimate RAM requirements based on model size
- Detect system RAM on Linux/macOS and warn if model too large
- Show detailed system info in download prompt
- Suggest smaller model variants when system can't handle the model
- Block download if disk space is insufficient
- Allow download with warning if RAM might be insufficient
