+++ src/catllm/text_functions_unified.py
@@ -458,6 +458,119 @@ def _format_bytes(size_bytes: int) -> str:
         return f"{size_bytes / (1024 ** 3):.2f} GB"
 
 
+def _parse_size_string(size_str: str) -> int:
+    """Parse a size string like '2.0 GB' into bytes."""
+    if size_str == "unknown":
+        return 0
+
+    size_str = size_str.strip().upper()
+    try:
+        if "GB" in size_str:
+            return int(float(size_str.replace("GB", "").strip()) * 1024 ** 3)
+        elif "MB" in size_str:
+            return int(float(size_str.replace("MB", "").strip()) * 1024 ** 2)
+        elif "KB" in size_str:
+            return int(float(size_str.replace("KB", "").strip()) * 1024)
+        else:
+            return int(float(size_str.replace("B", "").strip()))
+    except ValueError:
+        return 0
+
+
+def check_system_resources(model: str) -> dict:
+    """
+    Check if system has enough resources to download and run a model.
+
+    Args:
+        model: Model name to check
+
+    Returns:
+        dict with 'can_download', 'can_run', 'warnings', and 'details'
+    """
+    import shutil
+    import os
+
+    result = {
+        "can_download": True,
+        "can_run": True,
+        "warnings": [],
+        "details": {}
+    }
+
+    size_estimate = get_ollama_model_size_estimate(model)
+    model_size_bytes = _parse_size_string(size_estimate)
+
+    # Check disk space (Ollama typically stores models in ~/.ollama)
+    ollama_dir = os.path.expanduser("~/.ollama")
+    if not os.path.exists(ollama_dir):
+        ollama_dir = os.path.expanduser("~")
+
+    try:
+        disk_usage = shutil.disk_usage(ollama_dir)
+        free_space = disk_usage.free
+        result["details"]["free_disk_space"] = _format_bytes(free_space)
+        result["details"]["model_size"] = size_estimate
+
+        # Need at least 1.5x model size for download + extraction
+        required_space = int(model_size_bytes * 1.5) if model_size_bytes > 0 else 0
+
+        if required_space > 0 and free_space < required_space:
+            result["can_download"] = False
+            result["warnings"].append(
+                f"Insufficient disk space. Need ~{_format_bytes(required_space)}, "
+                f"but only {_format_bytes(free_space)} available."
+            )
+        elif required_space > 0 and free_space < required_space * 2:
+            result["warnings"].append(
+                f"Low disk space warning: {_format_bytes(free_space)} available."
+            )
+    except Exception:
+        result["details"]["free_disk_space"] = "unknown"
+
+    # Estimate RAM requirements (rough guide: model size * 1.2 for inference)
+    # This is approximate - actual requirements vary by quantization
+    if model_size_bytes > 0:
+        estimated_ram = model_size_bytes * 1.2
+        result["details"]["estimated_ram"] = _format_bytes(int(estimated_ram))
+
+        # Try to get system RAM (works on most systems)
+        try:
+            import subprocess
+            if os.name == 'posix':  # Linux/macOS
+                if os.path.exists('/proc/meminfo'):  # Linux
+                    with open('/proc/meminfo', 'r') as f:
+                        for line in f:
+                            if line.startswith('MemTotal:'):
+                                total_ram = int(line.split()[1]) * 1024  # Convert KB to bytes
+                                break
+                else:  # macOS
+                    output = subprocess.check_output(['sysctl', '-n', 'hw.memsize'], text=True)
+                    total_ram = int(output.strip())
+
+                result["details"]["total_ram"] = _format_bytes(total_ram)
+
+                if estimated_ram > total_ram * 0.8:
+                    result["can_run"] = False
[…]
