### 2026-01-04 90e2a43
Check system resources before running Ollama models
Resource check now happens in two scenarios:
1. Before downloading a new model (existing behavior)
2. Before running an already-installed model that may be too large

For installed models that exceed system RAM, user is warned and asked
to confirm before proceeding, with suggestions for smaller variants.
