### 2026-01-05 4713a18
Fix extract API: use chat.completions for OpenAI-compatible providers
- Fixed bug where huggingface/perplexity/xai providers incorrectly used
  client.messages.create() (Anthropic API) instead of
  client.chat.completions.create() (OpenAI API)
- Bumped version to 0.1.6
