### 2026-01-05 b40c88a
fixed a bug in the extract function, cleaned up huggingface app
