+++ src/catllm/text_functions_unified.py
@@ -789,6 +789,176 @@ def extract_json(reply: str) -> str:
         return '{"1":"e"}'
 
 
+def validate_classification_json(json_str: str, num_categories: int) -> tuple[bool, dict | None]:
+    """
+    Validate that a JSON string contains valid classification output.
+
+    Args:
+        json_str: The JSON string to validate
+        num_categories: Expected number of categories
+
+    Returns:
+        tuple: (is_valid, parsed_dict or None)
+    """
+    try:
+        parsed = json.loads(json_str)
+
+        if not isinstance(parsed, dict):
+            return False, None
+
+        # Check that all expected keys are present and values are "0" or "1"
+        for i in range(1, num_categories + 1):
+            key = str(i)
+            if key not in parsed:
+                return False, None
+            val = str(parsed[key]).strip()
+            if val not in ("0", "1"):
+                return False, None
+
+        # Normalize values to strings
+        normalized = {str(i): str(parsed[str(i)]).strip() for i in range(1, num_categories + 1)}
+        return True, normalized
+
+    except (json.JSONDecodeError, KeyError, TypeError):
+        return False, None
+
+
+def ollama_two_step_classify(
+    client,
+    response_text: str,
+    categories: list,
+    categories_str: str,
+    survey_question: str = "",
+    creativity: float = None,
+    max_retries: int = 5,
+) -> tuple[str, str | None]:
+    """
+    Two-step classification for Ollama models.
+
+    Step 1: Classify the response (natural language output OK)
+    Step 2: Convert classification to strict JSON format
+
+    This approach is more reliable for local models that struggle with
+    simultaneous reasoning and JSON formatting.
+
+    TODO: Fine-tune a small local model specifically for Step 2 (JSON formatting).
+          This would eliminate the need for retries and improve speed/consistency.
+          Could train on examples of classification text -> JSON output pairs.
+
+    Args:
+        client: UnifiedLLMClient instance
+        response_text: The survey response to classify
+        categories: List of category names
+        categories_str: Pre-formatted category string
+        survey_question: Optional context
+        creativity: Temperature setting
+        max_retries: Number of retry attempts for JSON validation
+
+    Returns:
+        tuple: (json_string, error_message or None)
+    """
+    num_categories = len(categories)
+    survey_context = f"A respondent was asked: {survey_question}." if survey_question else ""
+
+    # ==========================================================================
+    # Step 1: Classification (natural language - focus on accuracy)
+    # ==========================================================================
+    step1_messages = [
+        {
+            "role": "system",
+            "content": "You are an expert at categorizing survey responses. Focus on accurate classification."
+        },
+        {
+            "role": "user",
+            "content": f"""{survey_context}
+
+Analyze this survey response and determine which categories apply:
+
+Response: "{response_text}"
+
+Categories:
+{categories_str}
+
+For each category, explain briefly whether it applies (YES) or not (NO) to this response.
+Format your answer as:
+1. [Category name]: YES/NO - [brief reason]
+2. [Category name]: YES/NO - [brief reason]
+...and so on for all categories."""
+        }
+    ]
+
+    step1_reply, step1_error = client.complete(
+        messages=step1_messages,
+        json_schema=None,  # No JSON requirement for step 1
+        creativity=creativity,
+    )
+
+    if step1_error:
+        return '{"1":"e"}', f"Step 1 failed: {step1_error}"
+
+    # ==========================================================================
+    # Step 2: JSON Formatting with validation and retry
+    # ==========================================================================
+    example_json = json.dumps({str(i): "0" for i in range(1, num_categories + 1)})
+
+    for attempt in range(max_retries):
+        step2_messages = [
+            {
+                "role": "system",
[…]
