+++ hf_space/app.py
@@ -187,12 +187,11 @@ result.to_csv("classified_results.csv", index=False)
 FREE_MODEL_CHOICES = [
     "Qwen/Qwen3-VL-235B-A22B-Instruct:novita",
     "deepseek-ai/DeepSeek-V3.1:novita",
-    "meta-llama/Llama-4.1-8B-Instruct:novita",
+    "meta-llama/Llama-3.3-70B-Instruct:groq",
     "gemini-2.5-flash",
     "gpt-4o",
     "mistral-medium-2505",
     "claude-3-haiku-20240307",
-    "sonar",
     "grok-4-fast-non-reasoning",
 ]
 
@@ -213,7 +212,7 @@ PAID_MODEL_CHOICES = [
 HF_ROUTED_MODELS = [
     "Qwen/Qwen3-VL-235B-A22B-Instruct:novita",
     "deepseek-ai/DeepSeek-V3.1:novita",
-    "meta-llama/Llama-4.1-8B-Instruct:novita",
+    "meta-llama/Llama-3.3-70B-Instruct:groq",
 ]
 
 
@@ -293,9 +292,7 @@ def generate_methodology_report_pdf(categories, model, column_name, num_rows, mo
     story = []
 
     # Title based on task type
-    if task_type == "extract":
-        report_title = "CatLLM Category Extraction Report"
-    elif task_type == "extract_and_assign":
+    if task_type == "extract_and_assign":
         report_title = "CatLLM Extraction &amp; Classification Report"
     else:
         report_title = "CatLLM Classification Report"
@@ -306,12 +303,7 @@ def generate_methodology_report_pdf(categories, model, column_name, num_rows, mo
 
     story.append(Paragraph("About This Report", heading_style))
 
-    if task_type == "extract":
-        about_text = """This methodology report documents the category extraction process for reproducibility and transparency. \
-CatLLM uses Large Language Models to automatically discover meaningful categories from your data. The extraction process \
-analyzes your data in chunks, identifies recurring themes, and consolidates them into a final set of categories. \
-This automated approach helps researchers avoid confirmation bias in category selection."""
-    elif task_type == "extract_and_assign":
+    if task_type == "extract_and_assign":
         about_text = """This methodology report documents the automated category extraction and classification process. \
 CatLLM first discovers categories from your data using LLMs, then classifies each item into those categories. \
 This two-phase approach combines exploratory category discovery with systematic classification, ensuring both \
@@ -458,27 +450,14 @@ consistent and reproducible results."""
 
     # Summary section - adjust title based on task type
     story.append(PageBreak())
-    if task_type == "extract":
-        story.append(Paragraph("Extraction Summary", title_style))
-    elif task_type == "extract_and_assign":
+    if task_type == "extract_and_assign":
         story.append(Paragraph("Processing Summary", title_style))
     else:
         story.append(Paragraph("Classification Summary", title_style))
     story.append(Spacer(1, 15))
 
     # Build summary data based on task type
-    if task_type == "extract":
-        story.append(Paragraph("Extraction Details", heading_style))
-        summary_data = [
-            ["Source File", filename],
-            ["Source Column/Description", column_name],
-            ["Input Type", input_type],
-            ["Model Used", model],
-            ["Model Source", model_source],
-            ["Max Categories Requested", str(max_categories or "default")],
-            ["Categories Extracted", str(len(categories)) if categories else "0"],
-        ]
-    else:
+    if task_type == "assign":
         story.append(Paragraph("Classification Details", heading_style))
         summary_data = [
             ["Source File", filename],
@@ -596,28 +575,7 @@ is the key and a 1 if the category is present and a 0 if not.'''
     story.append(PageBreak())
[…]
