+++ hf_space/requirements.txt
@@ -1,5 +1,6 @@
-cat-llm[pdf]>=0.1.4
+cat-llm[pdf]>=0.1.6
 gradio==5.6.0
+mistralai
 pydantic==2.10.6
 huggingface_hub<0.27.0
 pandas
