### 2026-01-06 99d74b9
converted web app to streamlit
