+++ hf_space/requirements.txt
@@ -1,5 +1,5 @@
+streamlit>=1.32.0
 cat-llm[pdf]>=0.1.6
-gradio==5.6.0
 mistralai
 pydantic==2.10.6
 huggingface_hub<0.27.0
@@ -9,3 +9,4 @@ requests
 regex
 reportlab
 matplotlib
+Pillow
