### 2026-01-06 668a382
updating readme, converting explore corpus text function to use http instead of relying on packages
