+++ examples/test_all_models.py
@@ -1,0 +1,404 @@
+#!/usr/bin/env python
+"""
+Test script for all models available in the HuggingFace app.
+
+This script tests the unified text classification function with every model
+that's available in the CatLLM HuggingFace Space.
+
+Run with: python examples/test_all_models.py
+"""
+
+import sys
+import os
+
+# Add the src directory to path so we import from local code, not installed package
+src_path = '/Users/chrissoria/Documents/Research/cat-llm/src'
+sys.path.insert(0, src_path)
+
+# Remove any cached catllm modules to ensure we load from local
+modules_to_remove = [key for key in sys.modules.keys() if key.startswith('catllm')]
+for mod in modules_to_remove:
+    del sys.modules[mod]
+
+import pandas as pd
+import argparse
+import time
+import json
+from datetime import datetime
+from dotenv import load_dotenv, find_dotenv
+from catllm.text_functions import multi_class, detect_provider
+import catllm
+
+# Load API keys from .env file
+os.chdir('/Users/chrissoria/Documents/Research/Categorization_AI_experiments')
+_ = load_dotenv(find_dotenv())
+os.chdir('/Users/chrissoria/Documents/Research/cat-llm')
+
+# Get API keys
+openai_api_key = os.getenv("OPENAI_API_KEY")
+anthropic_api_key = os.getenv("ANTHROPIC_API_KEY")
+google_api_key = os.getenv("GOOGLE_API_KEY")
+mistral_api_key = os.getenv("MISTRAL_API_KEY")
+xai_api_key = os.getenv("XAI_API_KEY")
+huggingface_api_key = os.getenv("HUGGINGFACE_API_KEY")
+perplexity_api_key = os.getenv("PERPLEXITY_API_KEY")
+
+print(f"Testing local catllm version: {catllm.__version__}")
+print(f"Loaded from: {catllm.__file__}")
+print()
+
+# Verify keys loaded
+print("API keys loaded:")
+print(f"  OpenAI: {'✅' if openai_api_key else '❌'}")
+print(f"  Anthropic: {'✅' if anthropic_api_key else '❌'}")
+print(f"  Google: {'✅' if google_api_key else '❌'}")
+print(f"  Mistral: {'✅' if mistral_api_key else '❌'}")
+print(f"  xAI: {'✅' if xai_api_key else '❌'}")
+print(f"  HuggingFace: {'✅' if huggingface_api_key else '❌'}")
+print(f"  Perplexity: {'✅' if perplexity_api_key else '❌'}")
+print()
+
+# Output directory
+output_dir = os.path.join(os.getcwd(), 'examples', 'test_output')
+os.makedirs(output_dir, exist_ok=True)
+
+# =============================================================================
+# Model Configuration (mirrors hf_space/app.py)
+# =============================================================================
+
+FREE_MODEL_CHOICES = [
+    "Qwen/Qwen3-VL-235B-A22B-Instruct:novita",
+    "deepseek-ai/DeepSeek-V3.1:novita",
+    "meta-llama/Llama-3.3-70B-Instruct:groq",
+    "gemini-2.5-flash",
+    "gpt-4o",
+    "mistral-medium-2505",
+    "claude-3-haiku-20240307",
+    "grok-4-fast-non-reasoning",
+]
+
+PAID_MODEL_CHOICES = [
+    "gpt-4.1",
+    "gpt-4o",
+    "gpt-4o-mini",
+    "claude-sonnet-4-5-20250929",
+    "claude-opus-4-20250514",
+    "claude-3-5-haiku-20241022",
+    "gemini-2.5-pro",
+    "gemini-2.5-flash",
+    "mistral-large-latest",
+]
+
+# Models routed through HuggingFace
+HF_ROUTED_MODELS = [
[…]
