+++ src/catllm/text_functions_unified.py
@@ -1,1244 +1,0 @@
-"""
-Unified HTTP-based multi-class text classification.
-
-This module uses raw HTTP requests (via `requests`) for all LLM providers,
-eliminating the need for provider-specific SDKs like openai, anthropic, mistralai, etc.
-"""
-
-import json
-import time
-import requests
-import pandas as pd
-import regex
-from tqdm import tqdm
-
-
-# =============================================================================
-# Provider Configuration
-# =============================================================================
-
-PROVIDER_CONFIG = {
-    "openai": {
-        "endpoint": "https://api.openai.com/v1/chat/completions",
-        "auth_header": "Authorization",
-        "auth_prefix": "Bearer ",
-    },
-    "anthropic": {
-        "endpoint": "https://api.anthropic.com/v1/messages",
-        "auth_header": "x-api-key",
-        "auth_prefix": "",
-    },
-    "google": {
-        "endpoint": "https://generativelanguage.googleapis.com/v1beta/models/{model}:generateContent",
-        "auth_header": "x-goog-api-key",
-        "auth_prefix": "",
-    },
-    "mistral": {
-        "endpoint": "https://api.mistral.ai/v1/chat/completions",
-        "auth_header": "Authorization",
-        "auth_prefix": "Bearer ",
-    },
-    "perplexity": {
-        "endpoint": "https://api.perplexity.ai/chat/completions",
-        "auth_header": "Authorization",
-        "auth_prefix": "Bearer ",
-    },
-    "xai": {
-        "endpoint": "https://api.x.ai/v1/chat/completions",
-        "auth_header": "Authorization",
-        "auth_prefix": "Bearer ",
-    },
-    "huggingface": {
-        "endpoint": "https://router.huggingface.co/v1/chat/completions",
-        "auth_header": "Authorization",
-        "auth_prefix": "Bearer ",
-    },
-    "ollama": {
-        "endpoint": "http://localhost:11434/v1/chat/completions",
-        "auth_header": None,  # No auth required for local Ollama
-        "auth_prefix": "",
-    },
-}
-
-
-# =============================================================================
-# Unified API Client
-# =============================================================================
-
-class UnifiedLLMClient:
-    """A unified client for calling various LLM providers via HTTP."""
-
-    def __init__(self, provider: str, api_key: str, model: str):
-        self.provider = provider.lower()
-        self.api_key = api_key
-        self.model = model
-
-        if self.provider not in PROVIDER_CONFIG:
-            raise ValueError(f"Unsupported provider: {provider}. "
-                           f"Supported: {list(PROVIDER_CONFIG.keys())}")
-
-        self.config = PROVIDER_CONFIG[self.provider]
-
-    def _get_endpoint(self) -> str:
-        """Get the API endpoint, substituting model if needed."""
-        endpoint = self.config["endpoint"]
-        if "{model}" in endpoint:
-            endpoint = endpoint.format(model=self.model)
-        return endpoint
-
-    def _get_headers(self) -> dict:
-        """Build request headers for the provider."""
-        headers = {"Content-Type": "application/json"}
-        auth_header = self.config["auth_header"]
-        auth_prefix = self.config["auth_prefix"]
-
-        # Some providers (like Ollama) don't require auth
-        if auth_header is not None:
-            headers[auth_header] = f"{auth_prefix}{self.api_key}"
-
-        # Anthropic requires additional headers
-        if self.provider == "anthropic":
-            headers["anthropic-version"] = "2023-06-01"
-
-        return headers
-
-    def _build_payload(
-        self,
-        messages: list,
-        json_schema: dict = None,
-        creativity: float = None,
-        max_tokens: int = 4096,
-    ) -> dict:
[…]

[…]
