### 2026-01-06 736cf86
Add progress_callback parameter for PDF page-by-page progress updates
