### 2026-01-06 5c4b62d
made pdf functionality more user friendly in huggingface app
