+++ hf_space/requirements.txt
@@ -1,5 +1,5 @@
 streamlit>=1.32.0
-cat-llm[pdf]>=0.1.6
+cat-llm[pdf] @ git+https://github.com/chrissoria/cat-llm.git@main
 mistralai
 pydantic==2.10.6
 huggingface_hub<0.27.0
