### 2026-01-07 98324d9
Add focus parameter and fix classify() parameter mapping
- Add focus parameter to explore_corpus and explore_common_categories
- Add test script for explore_corpus (test_explore_corpus.py)
- Fix classify() to pass model/provider correctly to multi_class()
- Bump version to 0.1.8

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
