+++ examples/test_explore_corpus.py
@@ -1,0 +1,369 @@
+#!/usr/bin/env python
+"""
+Test script for explore_corpus function across all supported models.
+
+This script tests the explore_corpus function with the new 'focus' parameter
+across all LLM providers supported by CatLLM.
+
+Run with: python examples/test_explore_corpus.py
+"""
+
+import sys
+import os
+
+# Add the src directory to path so we import from local code, not installed package
+src_path = '/Users/chrissoria/Documents/Research/cat-llm/src'
+sys.path.insert(0, src_path)
+
+# Remove any cached catllm modules to ensure we load from local
+modules_to_remove = [key for key in sys.modules.keys() if key.startswith('catllm')]
+for mod in modules_to_remove:
+    del sys.modules[mod]
+
+import pandas as pd
+import argparse
+import time
+import json
+from datetime import datetime
+from dotenv import load_dotenv, find_dotenv
+from catllm.text_functions import explore_corpus, detect_provider
+import catllm
+
+# Load API keys from .env file
+load_dotenv('/Users/chrissoria/Documents/Research/Categorization_AI_experiments/.env')
+
+# Get API keys
+openai_api_key = os.getenv("OPENAI_API_KEY")
+anthropic_api_key = os.getenv("ANTHROPIC_API_KEY")
+google_api_key = os.getenv("GOOGLE_API_KEY")
+mistral_api_key = os.getenv("MISTRAL_API_KEY")
+xai_api_key = os.getenv("XAI_API_KEY")
+huggingface_api_key = os.getenv("HUGGINGFACE_API_KEY")
+perplexity_api_key = os.getenv("PERPLEXITY_API_KEY")
+
+print(f"Testing local catllm version: {catllm.__version__}")
+print(f"Loaded from: {catllm.__file__}")
+print()
+
+# Verify keys loaded
+print("API keys loaded:")
+print(f"  OpenAI: {'Yes' if openai_api_key else 'No'}")
+print(f"  Anthropic: {'Yes' if anthropic_api_key else 'No'}")
+print(f"  Google: {'Yes' if google_api_key else 'No'}")
+print(f"  Mistral: {'Yes' if mistral_api_key else 'No'}")
+print(f"  xAI: {'Yes' if xai_api_key else 'No'}")
+print(f"  HuggingFace: {'Yes' if huggingface_api_key else 'No'}")
+print(f"  Perplexity: {'Yes' if perplexity_api_key else 'No'}")
+print()
+
+# Output directory
+output_dir = os.path.join(os.getcwd(), 'examples', 'test_output')
+os.makedirs(output_dir, exist_ok=True)
+
+# =============================================================================
+# Model Configuration
+# =============================================================================
+
+# One model per provider for testing
+TEST_MODELS = [
+    ("gpt-4o-mini", "openai"),
+    ("claude-3-haiku-20240307", "anthropic"),
+    ("gemini-2.0-flash", "google"),
+    ("mistral-small-latest", "mistral"),
+    ("grok-3-mini-fast", "xai"),
+    ("meta-llama/Llama-3.3-70B-Instruct", "huggingface"),
+]
+
+# =============================================================================
+# Test Data
+# =============================================================================
+
+TEST_SURVEY_QUESTION = "Why did you move to your current residence?"
+
+TEST_RESPONSES = [
+    "I got a new job in the city and needed to be closer to work",
+    "My rent was getting too expensive so I had to find somewhere cheaper",
+    "Wanted to be closer to my aging parents who need help",
+    "Got married and we needed a bigger place together",
+    "The neighborhood was getting unsafe so we moved for safety",
+    "Enrolled in graduate school and moved to be near campus",
+    "Divorced and had to find my own place",
+    "Company relocated me to a new office",
[…]
