+++ examples/test_multi_class_ensemble.py
@@ -1,0 +1,362 @@
+#!/usr/bin/env python
+"""
+Test script for multi_class_ensemble function.
+
+This script tests the multi-model ensemble classification function which
+calls multiple LLM models in parallel and combines results using majority voting.
+
+Run with: python examples/test_multi_class_ensemble.py
+"""
+
+import sys
+import os
+
+# Add the src directory to path so we import from local code, not installed package
+src_path = '/Users/chrissoria/Documents/Research/cat-llm/src'
+sys.path.insert(0, src_path)
+
+# Remove any cached catllm modules to ensure we load from local
+modules_to_remove = [key for key in sys.modules.keys() if key.startswith('catllm')]
+for mod in modules_to_remove:
+    del sys.modules[mod]
+
+import pandas as pd
+import argparse
+import time
+import json
+from datetime import datetime
+from dotenv import load_dotenv, find_dotenv
+from catllm.text_functions_ensemble import multi_class_ensemble
+import catllm
+
+# Load API keys from .env files (try both locations)
+load_dotenv('/Users/chrissoria/Documents/Research/cat-llm/.env')
+load_dotenv('/Users/chrissoria/Documents/Research/Categorization_AI_experiments/.env')
+
+# Get API keys
+openai_api_key = os.getenv("OPENAI_API_KEY")
+anthropic_api_key = os.getenv("ANTHROPIC_API_KEY")
+google_api_key = os.getenv("GOOGLE_API_KEY")
+mistral_api_key = os.getenv("MISTRAL_API_KEY")
+xai_api_key = os.getenv("XAI_API_KEY")
+
+print(f"Testing local catllm version: {catllm.__version__}")
+print(f"Loaded from: {catllm.__file__}")
+print()
+
+# Verify keys loaded
+print("API keys loaded:")
+print(f"  OpenAI: {'Y' if openai_api_key else 'N'}")
+print(f"  Anthropic: {'Y' if anthropic_api_key else 'N'}")
+print(f"  Google: {'Y' if google_api_key else 'N'}")
+print(f"  Mistral: {'Y' if mistral_api_key else 'N'}")
+print(f"  xAI: {'Y' if xai_api_key else 'N'}")
+print()
+
+# Output directory
+output_dir = os.path.join(os.getcwd(), 'examples', 'test_output')
+os.makedirs(output_dir, exist_ok=True)
+
+# =============================================================================
+# Test Data
+# =============================================================================
+
+TEST_RESPONSES = [
+    "I moved because I got a new job in another city",
+    "My family needed to be closer to my elderly parents",
+    "The rent was too expensive so I had to find somewhere cheaper",
+]
+
+TEST_CATEGORIES = [
+    "Employment/Job-related",
+    "Family reasons",
+    "Financial/Cost of living",
+    "Education",
+    "Health reasons",
+]
+
+# =============================================================================
+# Model Configurations
+# =============================================================================
+
+def get_available_models():
+    """Get list of available model configurations based on loaded API keys."""
+    models = []
+
+    if openai_api_key:
+        models.append(("gpt-4o-mini", "openai", openai_api_key))
+
+    if anthropic_api_key:
+        models.append(("claude-3-5-haiku-20241022", "anthropic", anthropic_api_key))
+
+    if google_api_key:
+        models.append(("gemini-2.0-flash", "google", google_api_key))
+
+    if mistral_api_key:
+        models.append(("mistral-small-latest", "mistral", mistral_api_key))
+
+    return models
+
+
+# =============================================================================
+# Test Functions
+# =============================================================================
+
+def test_ensemble(models: list, num_responses: int = 1, verbose: bool = True,
+                  chain_of_verification: bool = False, safety: bool = False,
[…]
