+++ src/catllm/__init__.py
@@ -12,6 +12,7 @@ from .__about__ import (
 )
 
 from .text_functions import *
+from .text_functions_ensemble import *
 from .CERAD_functions import *
 from .image_functions import *
 from .pdf_functions import *
