+++ examples/test_image_ensemble.py
@@ -1,0 +1,285 @@
+#!/usr/bin/env python
+"""
+Test script for image classification with multi_class_ensemble.
+
+Run with: python examples/test_image_ensemble.py
+"""
+
+import sys
+import os
+
+# Add the src directory to path so we import from local code, not installed package
+src_path = '/Users/chrissoria/Documents/Research/cat-llm/src'
+sys.path.insert(0, src_path)
+
+# Remove any cached catllm modules to ensure we load from local
+modules_to_remove = [key for key in sys.modules.keys() if key.startswith('catllm')]
+for mod in modules_to_remove:
+    del sys.modules[mod]
+
+import pandas as pd
+from datetime import datetime
+from dotenv import load_dotenv
+from catllm.text_functions_ensemble import multi_class_ensemble
+import catllm
+
+# Load API keys from .env files
+load_dotenv('/Users/chrissoria/Documents/Research/cat-llm/.env')
+load_dotenv('/Users/chrissoria/Documents/Research/Categorization_AI_experiments/.env')
+
+# Get API keys
+openai_api_key = os.getenv("OPENAI_API_KEY")
+anthropic_api_key = os.getenv("ANTHROPIC_API_KEY")
+
+print(f"Testing local catllm version: {catllm.__version__}")
+print(f"Loaded from: {catllm.__file__}")
+print()
+
+# Verify keys loaded
+print("API keys loaded:")
+print(f"  OpenAI: {'Y' if openai_api_key else 'N'}")
+print(f"  Anthropic: {'Y' if anthropic_api_key else 'N'}")
+print()
+
+# Test images
+TEST_IMAGES = [
+    "/Users/chrissoria/Documents/Research/cat-llm/tests/9924971435_3ac76eed4f_c.jpg",
+    "/Users/chrissoria/Documents/Research/cat-llm/tests/9688423031_065013996f_c.jpg",
+    "/Users/chrissoria/Documents/Research/cat-llm/tests/9688419667_16c304583c_c.jpg",
+]
+
+# Categories for image classification
+TEST_CATEGORIES = [
+    "Contains people",
+    "Outdoor scene",
+    "Contains buildings/architecture",
+    "Contains nature/landscape",
+    "Contains vehicles",
+]
+
+# Output directory
+output_dir = os.path.join(os.getcwd(), 'examples', 'test_output')
+os.makedirs(output_dir, exist_ok=True)
+
+
+def test_single_model_image():
+    """Test image classification with a single model."""
+    print("\n" + "="*70)
+    print("TEST: Single Model Image Classification")
+    print("="*70)
+
+    if not openai_api_key:
+        print("ERROR: No OpenAI API key available")
+        return None
+
+    print(f"Images to classify: {len(TEST_IMAGES)}")
+    for img in TEST_IMAGES:
+        print(f"  - {os.path.basename(img)}")
+
+    print(f"\nCategories ({len(TEST_CATEGORIES)}):")
+    for i, cat in enumerate(TEST_CATEGORIES, 1):
+        print(f"  {i}. {cat}")
+
+    print("\n>>> Running classification with gpt-4o-mini...")
+
+    try:
+        result = multi_class_ensemble(
+            survey_input=TEST_IMAGES,
+            categories=TEST_CATEGORIES,
+            models=[("gpt-4o-mini", "openai", openai_api_key)],
+            input_description="Photographs to classify",
+            creativity=0.1,
+            chain_of_thought=True,
+            use_json_schema=True,
+        )
+
+        print("\nRESULT:")
+        print(f"  Type: {type(result)}")
+
+        if isinstance(result, pd.DataFrame):
+            print(f"  Shape: {result.shape}")
+            print(f"  Columns: {list(result.columns)}")
[…]
