### 2026-01-11 af6d61f
Add summarize() function documentation to README
