### 2026-01-17 1878cee
Support string values for consensus_threshold parameter
Add support for intuitive string aliases alongside numeric values:
- "majority": 50% agreement
- "two-thirds": 67% agreement
- "unanimous": 100% agreement
- Numeric values (0-1) still work for custom thresholds

Also changed threshold comparison from > to >= for more intuitive behavior
(e.g., exactly 50% now meets the "majority" threshold).

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
