### 2026-01-17 cef0355
Add CHANGELOG.md for version tracking
