### 2026-01-17 c417dd7
Release version 2.0.0: Simplified API, ensemble methods, removed SDK dependencies
