+++ CHANGELOG.md
@@ -7,78 +7,268 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+---
+
+## [2.0.0] - 2025-01-17
+
+### Major Release: Simplified API & Ensemble Methods
+
+Version 2.0 represents a major simplification of CatLLM's architecture and API, making it easier to install, use, and extend.
+
 ### Added
-- Ensemble classification: run multiple models in parallel with consensus voting
-- String values for `consensus_threshold` parameter ("majority", "two-thirds", "unanimous")
-- Classification matrix heatmap visualization in web app
-- Category distribution charts in web app
+- **Ensemble classification**: Run multiple models in parallel and combine predictions
+  - Cross-provider ensembles (GPT-4o + Claude + Gemini)
+  - Self-consistency ensembles (same model with temperature variation)
+  - Model comparison mode for side-by-side evaluation
+- **Consensus voting methods**:
+  - `"majority"` - 50%+ agreement required
+  - `"two-thirds"` - 67%+ agreement required
+  - `"unanimous"` - 100% agreement required
+  - Custom numeric thresholds (e.g., `0.75` for 75%)
+- **Visualization tools** in web app:
+  - Classification matrix heatmap
+  - Category distribution charts
+  - Download buttons for all visualizations
+- PDF report generation with methodology documentation
 
 ### Changed
-- Web app: moved "Try Example Dataset" button next to file uploader
-- Web app: improved radio button styling to fill available width
-- Web app: applied Garamond font globally
+- **Simplified to 3 core functions**:
+  - `extract()` - Discover categories in your data
+  - `classify()` - Assign categories to your data
+  - `summarize()` - Generate summaries of your data
+- **Removed SDK dependencies**: All API calls now use pure `requests` library
+  - No more `openai`, `anthropic`, `google-generativeai` package requirements
+  - Lighter installation, fewer dependency conflicts
+  - Unified HTTP interface for all providers
+- **Streamlined parameters**: Consistent parameter names across all functions
+- Web app UI improvements: button alignment, Garamond font, improved layout
+
+### Removed
+- Direct SDK dependencies (openai, anthropic, google-generativeai, mistralai)
+- Legacy function names (old aliases still work but are deprecated)
+
+### Migration from 1.x
+Most code will work without changes. Key differences:
+- SDK-specific features (like streaming) are no longer available
+- All providers now use the same HTTP-based interface
+- New `models` parameter enables ensemble mode
+
+---
 
 ## [0.1.15] - 2025-01-10
 
 ### Added
-- `summarize()` function for text and PDF summarization
-- `focus` parameter for category extraction to prioritize specific themes
+- `summarize()` function for text and PDF summarization with multi-model support
+- `focus` parameter for `extract()` to prioritize specific themes during category discovery
 - `progress_callback` parameter for PDF page-by-page progress updates
-- Multi-model support in `classify()` with `models` parameter
+- Multi-model support in `classify()` via `models` parameter for ensemble classification
+- Documentation for `summarize()` function in README
+
+### Changed
+- Converted web app from Gradio to Streamlit for better mobile support
+- Improved PDF functionality in HuggingFace app
 
 ### Fixed
 - Parameter mapping in `classify()` function
[…]
