+++ src/catllm/calls/all_calls.py
@@ -1,5 +1,8 @@
 # openai stepback prompt
 
+import requests
+
+
 def get_stepback_insight_openai(
     stepback,
     api_key,
@@ -7,27 +10,46 @@ def get_stepback_insight_openai(
     model_source="openai",
     creativity=None
 ):
-    from openai import OpenAI
-    # Conditional base_url setting based on model source
-    base_url = (
-        "https://api.perplexity.ai" if model_source == "perplexity" 
-        else "https://router.huggingface.co/v1" if model_source == "huggingface"
-        else "https://api.x.ai/v1" if model_source == "xai"
-        else None
-    )
-    
-    client = OpenAI(api_key=api_key, base_url=base_url)
-    
+    """
+    Get stepback insight from OpenAI-compatible APIs.
+    Uses direct HTTP requests instead of OpenAI SDK for lighter dependencies.
+    """
+    # Determine the base URL based on model source
+    if model_source == "huggingface":
+        from catllm._providers import _detect_huggingface_endpoint
+        base_url = _detect_huggingface_endpoint(api_key, user_model)
+    elif model_source == "huggingface-together":
+        base_url = "https://router.huggingface.co/together/v1"
+    elif model_source == "perplexity":
+        base_url = "https://api.perplexity.ai"
+    elif model_source == "xai":
+        base_url = "https://api.x.ai/v1"
+    else:
+        base_url = "https://api.openai.com/v1"
+
+    endpoint = f"{base_url}/chat/completions"
+
+    headers = {
+        "Content-Type": "application/json",
+        "Authorization": f"Bearer {api_key}"
+    }
+
+    payload = {
+        "model": user_model,
+        "messages": [{"role": "user", "content": stepback}],
+    }
+
+    if creativity is not None:
+        payload["temperature"] = creativity
+
     try:
-        stepback_response = client.chat.completions.create(
-            model=user_model,
-            messages=[{'role': 'user', 'content': stepback}],
-            **({"temperature": creativity} if creativity is not None else {})
-        )
-        stepback_insight = stepback_response.choices[0].message.content
-        
+        response = requests.post(endpoint, headers=headers, json=payload, timeout=120)
+        response.raise_for_status()
+        result = response.json()
+        stepback_insight = result["choices"][0]["message"]["content"]
+
         return stepback_insight, True
-        
+
     except Exception as e:
         print(f"An error occurred during step-back prompting: {e}")
         return None, False
@@ -42,21 +64,43 @@ def get_stepback_insight_anthropic(
     model_source="anthropic",
     creativity=None
 ):
-    import anthropic
+    """
+    Get stepback insight from Anthropic Claude.
 
-    client = anthropic.Anthropic(api_key=api_key)
-    
+    Uses direct HTTP requests instead of Anthropic SDK for lighter dependencies.
+    """
+    import requests
+
+    endpoint = "https://api.anthropic.com/v1/messages"
+
+    headers = {
+        "Content-Type": "application/json",
+        "x-api-key": api_key,
+        "anthropic-version": "2023-06-01"
+    }
+
+    payload = {
+        "model": user_model,
+        "max_tokens": 4096,
+        "messages": [{"role": "user", "content": stepback}],
+    }
+
+    if creativity is not None:
+        payload["temperature"] = creativity
+
     try:
-        stepback_response = client.messages.create(
-            model=user_model,
-            max_tokens=4096,
-            messages=[{'role': 'user', 'content': stepback}],
-            **({"temperature": creativity} if creativity is not None else {})
-        )
-        stepback_insight = stepback_response.content[0].text
-        
-        return stepback_insight, True
-        
+        response = requests.post(endpoint, headers=headers, json=payload, timeout=120)
+        response.raise_for_status()
+        result = response.json()
+
+        # Parse response - Anthropic returns content as a list
[…]
