### 2026-01-30 39160f7
Add progress_callback support for real-time progress tracking
- Add progress_callback parameter to explore_common_categories(),
  explore_image_categories(), explore_pdf_categories(), and extract()
- Callback signature: progress_callback(current_step, total_steps, step_label)
- Update HuggingFace app to use real progress bars with ETA
- Remove deprecated text_functions_old.py
- Bump version to 2.1.0

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
