+++ src/catllm/text_functions_old.py
@@ -1,1085 +1,0 @@
-from .calls.stepback import (
-    get_stepback_insight_openai,
-    get_stepback_insight_anthropic,
-    get_stepback_insight_google,
-    get_stepback_insight_mistral
-)
-from .calls.CoVe import (
-    chain_of_verification_openai,
-    chain_of_verification_google,
-    chain_of_verification_anthropic,
-    chain_of_verification_mistral
-)
-from .calls.top_n import (
-    get_openai_top_n,
-    get_anthropic_top_n,
-    get_google_top_n,
-    get_mistral_top_n
-)
-
-
-def _detect_model_source(user_model, model_source):
-    """Auto-detect model source from model name if not explicitly provided."""
-    model_source = model_source.lower()
-
-    if model_source is None or model_source == "auto":
-        user_model_lower = user_model.lower()
-
-        if "gpt" in user_model_lower:
-            return "openai"
-        elif "claude" in user_model_lower:
-            return "anthropic"
-        elif "gemini" in user_model_lower or "gemma" in user_model_lower:
-            return "google"
-        elif "llama" in user_model_lower or "meta" in user_model_lower:
-            return "huggingface"
-        elif "mistral" in user_model_lower or "mixtral" in user_model_lower:
-            return "mistral"
-        elif "sonar" in user_model_lower or "pplx" in user_model_lower:
-            return "perplexity"
-        elif "deepseek" in user_model_lower or "qwen" in user_model_lower:
-            return "huggingface"
-        elif "grok" in user_model_lower:
-            return "xai"
-        else:
-            raise ValueError(
-                f"Could not auto-detect model source from '{user_model}'. "
-                "Please specify model_source explicitly: OpenAI, Anthropic, Perplexity, Google, xAI, Huggingface, or Mistral"
-            )
-    return model_source
-
-
-def _get_stepback_insight(model_source, stepback, api_key, user_model, creativity):
-    """Get step-back insight using the appropriate provider."""
-    stepback_functions = {
-        "openai": get_stepback_insight_openai,
-        "perplexity": get_stepback_insight_openai,
-        "huggingface": get_stepback_insight_openai,
-        "xai": get_stepback_insight_openai,
-        "anthropic": get_stepback_insight_anthropic,
-        "google": get_stepback_insight_google,
-        "mistral": get_stepback_insight_mistral,
-    }
-
-    func = stepback_functions.get(model_source)
-    if func is None:
-        return None, False
-
-    return func(
-        stepback=stepback,
-        api_key=api_key,
-        user_model=user_model,
-        model_source=model_source,
-        creativity=creativity
-    )
-
-
-#extract categories from corpus
-def explore_corpus(
-    survey_question, 
-    survey_input,
-    api_key,
-    research_question=None,
-    specificity="broad",
-    categories_per_chunk=10,
-    divisions=5,
-    user_model="gpt-5",
-    creativity=None,
-    filename="corpus_exploration.csv",
-    model_source="OpenAI"
-):
-    import os
-    import pandas as pd
-    import random
-    from openai import OpenAI
-    from openai import OpenAI, BadRequestError
-    from tqdm import tqdm
-
-    print(f"Exploring class for question: '{survey_question}'.\n          {categories_per_chunk * divisions} unique categories to be extracted.")
-    print()
-
-    model_source = model_source.lower() # eliminating case sensitivity
-
-    n = len(survey_input)
-
-    # Auto-adjust divisions for small datasets
-    original_divisions = divisions
-    divisions = min(divisions, max(1, n // 3))  # At least 3 responses per chunk
-    if divisions != original_divisions:
[…]
