### 2026-02-06 chrissoria/cat-llm#1 @BinhL3
JOSS Checklist Review - Unfulfilled Item
Hi @chrissoria,

Sorry for getting to the review a bit late. I noticed a few missing items on the JOSS checklist.

For community guidelines, it should satisfy showing ways others 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support. I know you stated in contact: "Interested in research collaboration? Email: ChrisSoria@Berkeley.edu" but this could be clearer/transparent (e.g. pointing people to GitHub Issues, contribution steps, etc.).

Also nit pick for license:
LICENSE contains full GPL-3.0 text but then pyproject.toml has license = "MIT" while LICENSE and README say GNU/GPL. You should align these.
