### 2026-02-06 #9678 @BinhL3
@ujjwalkarn sorry for the late reply. I opened an [issue](https://github.com/chrissoria/cat-llm/issues/1) with cat-llm here
