### 2026-02-08 ab0034f
Release v2.1.0: Unified classify() API, image auto-categories, ensemble fixes
- Add 9 missing parameters to classify() making it the single entry point
- Add 4-tuple model format for per-model temperature control in ensembles
- Add image/PDF auto-category extraction via routing through extract()
- Add retry logic with exponential backoff for image extraction
- Fix agreement calculation to measure fraction agreeing with consensus
- Fix MIME type normalization for Anthropic image API calls
- Remove dead duplicate classify() from main.py
- Update HuggingFace Space app to use classify() API
- Clea
[…]
