### 2026-02-08 bd35212
Bump version to 2.2.0 for PyPI release
v2.1.0 was already published from an earlier build. Bump to 2.2.0 so
the latest changes (gitignore cleanup, CHANGELOG, code fixes) are
included in the published package.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
