### 2026-02-08 29a994b
Fix expander icon overlap in HuggingFace app
Scope the global font-family CSS override to exclude Streamlit's
Material Icons elements. The `*` selector was replacing the icon font
with EB Garamond, causing "arrow_right" text to render instead of
the chevron icon in the About This App expander.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
