+++ hf_space/app.py
@@ -307,8 +307,17 @@ result = catllm.classify(
     user_model="{cls['model']}"{classify_mode_param}
 )'''
     else:
-        # Multi-model mode
-        models_str = ",\n        ".join([f'("{m}", "auto", "YOUR_API_KEY")' for m in cls['models_list']])
+        # Multi-model mode — include per-model temperatures when set
+        model_temps = cls.get('model_temperatures', {})
+        model_lines = []
+        for m in cls['models_list']:
+            temp = model_temps.get(m) if model_temps else None
+            if temp is not None:
+                model_lines.append(f'("{m}", "auto", "YOUR_API_KEY", {{"creativity": {temp}}})')
+            else:
+                model_lines.append(f'("{m}", "auto", "YOUR_API_KEY")')
+        models_str = ",\n        ".join(model_lines)
+
         classify_mode_param = f',\n    mode="{cls["mode"]}"' if cls.get('mode') and ext['input_type'] == "pdf" else ''
         threshold_str = "majority" if cls['consensus_threshold'] == 0.5 else "two-thirds" if cls['consensus_threshold'] == 0.67 else "unanimous"
         consensus_param = f',\n    consensus_threshold="{threshold_str}"' if cls['classify_mode'] == "Ensemble" else ''
@@ -337,7 +346,9 @@ result.to_csv("classified_results.csv", index=False)
 '''
 
 
-def generate_classify_code(input_type, description, categories, model, model_source, mode=None, classify_mode="Single Model", models_list=None, consensus_threshold=0.5):
+def generate_classify_code(input_type, description, categories, model, model_source, mode=None,
+                           classify_mode="Single Model", models_list=None, consensus_threshold=0.5,
+                           model_temperatures=None):
     """Generate Python code for classification."""
     categories_str = ",\n    ".join([f'"{cat}"' for cat in categories])
 
@@ -382,8 +393,16 @@ result.to_csv("classified_results.csv", index=False)
 '''
     else:
         # Multi-model mode (Comparison or Ensemble)
+        # Build model tuples with per-model temperature when set
         if models_list:
-            models_str = ",\n        ".join([f'("{m}", "auto", "YOUR_API_KEY")' for m in models_list])
+            model_lines = []
+            for m in models_list:
+                temp = model_temperatures.get(m) if model_temperatures else None
+                if temp is not None:
+                    model_lines.append(f'("{m}", "auto", "YOUR_API_KEY", {{"creativity": {temp}}})')
+                else:
+                    model_lines.append(f'("{m}", "auto", "YOUR_API_KEY")')
+            models_str = ",\n        ".join(model_lines)
         else:
             models_str = '("gpt-4o", "auto", "YOUR_API_KEY"),\n        ("claude-sonnet-4-5-20250929", "auto", "YOUR_API_KEY")'
 
@@ -423,7 +442,7 @@ def generate_methodology_report_pdf(categories, model, column_name, num_rows, mo
                           data_quality=None, catllm_version=None, python_version=None,
                           task_type="assign", extracted_categories_df=None, max_categories=None,
                           input_type="text", description=None, classify_mode="Single Model",
-                          models_list=None, code=None):
+                          models_list=None, code=None, consensus_threshold=None):
     """Generate a PDF methodology report."""
     from reportlab.lib.pagesizes import letter
     from reportlab.lib import colors
@@ -465,14 +484,38 @@ consistent and reproducible results."""
 
     if categories:
         story.append(Paragraph("Category Mapping", heading_style))
-        story.append(Paragraph("Each category column contains binary values: 1 = present, 0 = not present", normal_style))
-        story.append(Spacer(1, 8))
 
-        category_data = [["Column Name", "Category Description"]]
-        for i, cat in enumerate(categories, 1):
-            category_data.append([f"category_{i}", cat])
[…]
