### 2026-02-08 eeb95c4
Add website link to About section in HuggingFace app
Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
