### 2026-02-08 230f3c6
Add Acknowledgments section to About in HuggingFace app
Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
