+++ hf_space/app.py
@@ -308,14 +308,19 @@ result = catllm.classify(
 )'''
     else:
         # Multi-model mode — include per-model temperatures when set
-        model_temps = cls.get('model_temperatures', {})
+        ens_runs = cls.get('ensemble_runs')
         model_lines = []
-        for m in cls['models_list']:
-            temp = model_temps.get(m) if model_temps else None
-            if temp is not None:
+        if ens_runs:
+            for m, temp in ens_runs:
                 model_lines.append(f'("{m}", "auto", "YOUR_API_KEY", {{"creativity": {temp}}})')
-            else:
-                model_lines.append(f'("{m}", "auto", "YOUR_API_KEY")')
+        else:
+            model_temps = cls.get('model_temperatures', {})
+            for m in cls['models_list']:
+                temp = model_temps.get(m) if model_temps else None
+                if temp is not None:
+                    model_lines.append(f'("{m}", "auto", "YOUR_API_KEY", {{"creativity": {temp}}})')
+                else:
+                    model_lines.append(f'("{m}", "auto", "YOUR_API_KEY")')
         models_str = ",\n        ".join(model_lines)
 
         classify_mode_param = f',\n    mode="{cls["mode"]}"' if cls.get('mode') and ext['input_type'] == "pdf" else ''
@@ -348,7 +353,7 @@ result.to_csv("classified_results.csv", index=False)
 
 def generate_classify_code(input_type, description, categories, model, model_source, mode=None,
                            classify_mode="Single Model", models_list=None, consensus_threshold=0.5,
-                           model_temperatures=None):
+                           model_temperatures=None, ensemble_runs=None):
     """Generate Python code for classification."""
     categories_str = ",\n    ".join([f'"{cat}"' for cat in categories])
 
@@ -394,7 +399,13 @@ result.to_csv("classified_results.csv", index=False)
     else:
         # Multi-model mode (Comparison or Ensemble)
         # Build model tuples with per-model temperature when set
-        if models_list:
+        if ensemble_runs:
+            # Ensemble with explicit (model, temp) pairs (supports duplicate models)
+            model_lines = []
+            for m, temp in ensemble_runs:
+                model_lines.append(f'("{m}", "auto", "YOUR_API_KEY", {{"creativity": {temp}}})')
+            models_str = ",\n        ".join(model_lines)
+        elif models_list:
             model_lines = []
             for m in models_list:
                 temp = model_temperatures.get(m) if model_temperatures else None
@@ -1339,6 +1350,12 @@ This app is currently in **beta** and **free to use** while CatLLM is under revi
 - Found a bug? Have a feature request? Please open an issue on [GitHub](https://github.com/chrissoria/cat-llm)
 - Reach out directly: [chrissoria@berkeley.edu](mailto:chrissoria@berkeley.edu)
 
+### Acknowledgments
+- **Bashir Ahmed** for his generous fellowship support that makes this free beta possible
+- **Claude Fischer** for his thoughtful feedback and collaboration on research that helped inspire this project
+- **Kevin Collins** from Survey360 for his input
+- **Fendi Tsim** for sharing it widely
+
 ### Links
 - **Website**: [christophersoria.com](https://christophersoria.com)
 - **PyPI**: [pip install cat-llm](https://pypi.org/project/cat-llm/)
@@ -1696,8 +1713,59 @@ with col_input:
         is_multi_model = classify_mode in ["Model Comparison", "Ensemble"]
         min_models = 3 if classify_mode == "Ensemble" else 2
 
-        if model_tier == "Free Models":
-            if is_multi_model:
+        # Track per-run temperatures: list of (model_name, temperature) for ensemble,
+        # or dict {model_name: temperature} for model comparison
[…]
