### 2026-02-08 ebed2f0
Add supported providers to About section in HuggingFace app
Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
