### 2026-02-09 727920e
Release v2.3.0: Add explore() for raw category extraction
New explore() function returns unprocessed category lists for
saturation and frequency analysis. Also removes test_output/,
tests/, and survey-summarizer from tracking.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
