+++ tests/__init__.py
@@ -1,3 +1,0 @@
-# SPDX-FileCopyrightText: 2025-present Christopher Soria <chrissoria@berkeley.edu>
-#
-# SPDX-License-Identifier: MIT
