### 2026-02-11 06d5191
addressing reviewer feedback, aligning license, making contribution steps more clear
