+++ src/catllm/explore.py
@@ -20,14 +20,14 @@ def explore(
     description="",
     max_categories=12,
     categories_per_chunk=10,
-    divisions=5,
+    divisions=12,
     user_model="gpt-4o",
     creativity=None,
     specificity="broad",
     research_question=None,
     filename=None,
     model_source="auto",
-    iterations=3,
+    iterations=8,
     random_state=None,
     focus=None,
     progress_callback=None,
