+++ examples/test_params_8_12.py
@@ -1,0 +1,238 @@
+#!/usr/bin/env python
+"""
+Tests 8-12: step_back_prompt, context_prompt, few-shot examples, safety mode,
+consensus threshold variations. Uses only cheap models (OpenAI, Anthropic, Google).
+"""
+
+import sys, os, time, traceback
+import pandas as pd
+
+src_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), '..', 'src')
+sys.path.insert(0, os.path.abspath(src_path))
+for mod in [k for k in sys.modules if k.startswith('catllm')]:
+    del sys.modules[mod]
+
+from dotenv import load_dotenv
+for p in [
+    os.path.join(os.path.dirname(os.path.abspath(__file__)), '..', '.env'),
+    os.path.expanduser('~/.env'),
+    '/Users/chrissoria/Documents/Research/Categorization_AI_experiments/.env',
+]:
+    if os.path.isfile(p):
+        load_dotenv(p, override=True)
+
+from catllm import classify
+
+api_keys = {
+    "openai": os.getenv("OPENAI_API_KEY"),
+    "anthropic": os.getenv("ANTHROPIC_API_KEY"),
+    "google": os.getenv("GOOGLE_API_KEY"),
+}
+print("API Key Status:")
+for prov, key in api_keys.items():
+    print(f"  {prov:15s} {'Found' if key else 'MISSING'}")
+print()
+
+CHEAP = [
+    {"name": "OpenAI",    "provider": "openai",    "model": "gpt-4o-mini",               "key": "openai"},
+    {"name": "Anthropic", "provider": "anthropic",  "model": "claude-haiku-4-5-20251001", "key": "anthropic"},
+    {"name": "Google",    "provider": "google",     "model": "gemini-2.0-flash",          "key": "google"},
+]
+
+# Load test data
+script_dir = os.path.dirname(os.path.abspath(__file__))
+survey_df = pd.read_csv(os.path.join(script_dir, 'test_data', 'survey_responses.csv'))
+sentiment_df = survey_df[survey_df['survey_question'] == "How do you feel about this product?"]
+moving_df = survey_df[survey_df['survey_question'] == "Why did you move to your current residence?"]
+SENT_TEXTS = sentiment_df['response_text'].tolist()
+MOVE_TEXTS = moving_df['response_text'].tolist()
+CATEGORIES_SENT = ["Positive sentiment", "Negative sentiment", "Neutral sentiment"]
+CATEGORIES_MOVE = ["Employment", "Family", "Financial", "Housing",
+                   "Education", "Safety", "Lifestyle", "Life transition"]
+
+output_dir = os.path.join(os.path.dirname(script_dir), 'test_output')
+os.makedirs(output_dir, exist_ok=True)
+
+all_results = []
+
+def run_test(test_name, cfg, kwargs):
+    key = api_keys.get(cfg["key"])
+    result = {"test": test_name, "provider": cfg["name"], "model": cfg["model"],
+              "status": None, "rows": None, "time_s": None, "error": None}
+    if not key:
+        result["status"] = "SKIPPED"; result["error"] = "No API key"
+        return result
+    print(f"  {cfg['name']:15s} ({cfg['model']}) ... ", end="", flush=True)
+    t0 = time.time()
+    try:
+        df = classify(**kwargs, user_model=cfg["model"], model_source=cfg["provider"], api_key=key)
+        elapsed = round(time.time() - t0, 1)
+        result.update(status="PASS", rows=len(df), time_s=elapsed)
+        print(f"PASS  ({len(df)} rows, {elapsed}s)")
+    except Exception as e:
+        elapsed = round(time.time() - t0, 1)
[…]
