### 2026-02-11 #9678 @chrissoria
Thank you @BinhL3

I'm also attaching here a link to the package working within a huggingface app that requires no API key from users. I hope this is helpful for confirming functionality.

link: https://huggingface.co/spaces/CatLLM/survey-classifier
