### 2026-02-11 b064ff6
adding mention to the web research tool up top
