### 2026-02-11 b77f0af
Fix thinking_budget support for Google, OpenAI, and Anthropic
- Google: Move thinkingConfig inside generationConfig (was rejected at
  top level). Enforce minimum budget of 128 tokens.
- OpenAI: Add reasoning_effort parameter (0→minimal, >0→high). Omit
  temperature when reasoning is enabled to avoid 400 errors.
- Anthropic: Add extended thinking support with min 1024 budget. Set
  temperature=1 when thinking is enabled (Anthropic requirement).
- Pass thinking_budget to all three providers in classify_single.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
