+++ src/catllm/_providers.py
@@ -187,11 +187,13 @@ class UnifiedLLMClient:
         """Build the request payload for the specific provider."""
 
         if self.provider == "anthropic":
-            return self._build_anthropic_payload(messages, json_schema, creativity, max_tokens)
+            return self._build_anthropic_payload(messages, json_schema, creativity, max_tokens, thinking_budget)
         elif self.provider == "google":
             return self._build_google_payload(messages, json_schema, creativity, thinking_budget, force_json)
+        elif self.provider == "openai":
+            return self._build_openai_payload(messages, json_schema, creativity, force_json, thinking_budget)
         else:
-            # OpenAI-compatible providers
+            # Other OpenAI-compatible providers (xai, mistral, huggingface, etc.)
             return self._build_openai_payload(messages, json_schema, creativity, force_json)
 
     def _build_openai_payload(
@@ -200,11 +202,14 @@ class UnifiedLLMClient:
         json_schema: dict = None,
         creativity: float = None,
         force_json: bool = True,
+        thinking_budget: int = None,
     ) -> dict:
         """Build payload for OpenAI-compatible APIs.
 
         Args:
             force_json: If False and no json_schema, don't set response_format (for text responses)
+            thinking_budget: For OpenAI models, maps to reasoning_effort:
+                             0 or None → "minimal", >0 → "high"
         """
         payload = {
             "model": self.model,
@@ -230,7 +235,19 @@ class UnifiedLLMClient:
             payload["response_format"] = {"type": "json_object"}
         # else: no response_format - allow text responses
 
-        if creativity is not None:
+        # OpenAI reasoning_effort: translate thinking_budget to reasoning_effort
+        # thinking_budget=0 → reasoning_effort="minimal" (suppress reasoning)
+        # thinking_budget>0 → reasoning_effort="high" (enable reasoning)
+        if self.provider == "openai" and thinking_budget is not None:
+            if thinking_budget > 0:
+                payload["reasoning_effort"] = "high"
+                # When reasoning is enabled, temperature must NOT be in payload
+            else:
+                payload["reasoning_effort"] = "minimal"
+                # With minimal reasoning, temperature can still be set
+                if creativity is not None:
+                    payload["temperature"] = creativity
+        elif creativity is not None:
             payload["temperature"] = creativity
 
         return payload
@@ -241,8 +258,15 @@ class UnifiedLLMClient:
         json_schema: dict = None,
         creativity: float = None,
         max_tokens: int = 4096,
+        thinking_budget: int = None,
     ) -> dict:
-        """Build payload for Anthropic API."""
+        """Build payload for Anthropic API.
+
+        Args:
+            thinking_budget: Controls extended thinking for Anthropic models.
+                             0 or None → thinking disabled.
+                             >0 → thinking enabled with budget_tokens = max(thinking_budget, 1024).
+        """
         # Extract system message if present
         system_content = None
         user_messages = []
@@ -261,7 +285,20 @@ class UnifiedLLMClient:
         if system_content:
             payload["system"] = system_content
 
-        if creativity is not None:
+        # Extended thinking for Anthropic (minimum 1024 tokens)
+        # When thinking is enabled, temperature must be 1 (Anthropic requirement),
+        # so we skip setting temperature from creativity in that case
+        if thinking_budget and thinking_budget > 0:
+            budget = max(thinking_budget, 1024)
+            payload["thinking"] = {
+                "type": "enabled",
+                "budget_tokens": budget,
+            }
+            payload["temperature"] = 1
+            # When thinking is enabled, max_tokens must be larger than budget_tokens
+            if payload["max_tokens"] <= budget:
+                payload["max_tokens"] = budget + 4096
[…]
