+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to CatLLM will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.2] - 2026-02-10
+
+### Fixed
+- **Google thinking support**: Fixed `thinkingConfig` placement — must be inside `generationConfig`, not at the top level. Added minimum budget of 128 tokens.
+- **OpenAI reasoning support**: Fixed conflict between `reasoning_effort` and `temperature` — temperature is now omitted when reasoning is enabled (`thinking_budget > 0`).
+- **Anthropic thinking support**: Temperature is now set to 1 (Anthropic requirement) when extended thinking is enabled, instead of using the user-specified creativity value.
+
+---
+
 ## [2.3.1] - 2026-02-10
 
 ### Changed
