### 2026-02-11 4c12aad
Fix thinking_budget support in classify pipeline (v2.3.3)
Root cause: text_functions_ensemble.py imports UnifiedLLMClient from
text_functions.py, not _providers.py. The v2.3.2 fixes were applied
only to _providers.py and never reached the actual classify pipeline.

Fixes applied to both text_functions.py and _providers.py:
- Google: thinkingConfig inside generationConfig, min budget 128
- OpenAI: reasoning_effort only for o-series models (o1/o3/o4)
- Anthropic: tool_choice="auto" when thinking enabled (forced
  tool_choice incompatible with extended thinking)

Co-Authored-By: Claude Opus 4.6 <
[…]
