### 2026-02-11 a1d212c
Consolidate UnifiedLLMClient, add HuggingFace thinking support (v2.3.4)
- Remove ~930 lines of duplicated provider code from text_functions.py;
  _providers.py is now the single source of truth, text_functions.py
  re-exports for backward compatibility
- Add HuggingFace thinking control: thinking_budget=0 sends
  chat_template_kwargs to disable thinking on models like Qwen3
- Pass thinking_budget through ensemble pipeline for HuggingFace providers
- Add gpt-5 to OpenAI reasoning model detection
- Add ARCHITECTURE.md with module dependency map and classify() call chain
- Remove tracked __pycach
[…]
