+++ src/catllm/text_functions.py
@@ -31,8 +31,6 @@ __all__ = [
     "_get_stepback_insight",
     "_detect_huggingface_endpoint",
 ]
-import time
-import requests
 import pandas as pd
 import regex
 from tqdm import tqdm
@@ -56,494 +54,26 @@ from .calls.top_n import (
     get_mistral_top_n
 )
 
-
-# =============================================================================
-# HuggingFace Endpoint Auto-Detection
-# =============================================================================
-
-def _detect_huggingface_endpoint(api_key: str, model: str) -> str:
-    """
-    Test which HuggingFace endpoint works for this model.
-    Tries generic router first, then Together.
-
-    Args:
-        api_key: HuggingFace API key
-        model: Model name to test
-
-    Returns:
-        Base URL for the working endpoint (without /chat/completions)
-    """
-    endpoints = [
-        "https://router.huggingface.co/v1/chat/completions",
-        "https://router.huggingface.co/together/v1/chat/completions",
-    ]
-
-    headers = {
-        "Content-Type": "application/json",
-        "Authorization": f"Bearer {api_key}"
-    }
-
-    payload = {
-        "model": model,
-        "messages": [{"role": "user", "content": "hi"}],
-        "max_tokens": 5
-    }
-
-    for endpoint in endpoints:
-        try:
-            response = requests.post(endpoint, headers=headers, json=payload, timeout=30)
-            if response.status_code == 200:
-                # Return the base URL (without /chat/completions)
-                return endpoint.replace("/chat/completions", "")
-        except Exception:
-            continue
-
-    # Default to generic (will fail with informative error)
-    return "https://router.huggingface.co/v1"
-
-
-# =============================================================================
-# Provider Configuration
-# =============================================================================
-
-PROVIDER_CONFIG = {
-    "openai": {
-        "endpoint": "https://api.openai.com/v1/chat/completions",
-        "auth_header": "Authorization",
-        "auth_prefix": "Bearer ",
-    },
-    "anthropic": {
-        "endpoint": "https://api.anthropic.com/v1/messages",
-        "auth_header": "x-api-key",
-        "auth_prefix": "",
-    },
-    "google": {
-        "endpoint": "https://generativelanguage.googleapis.com/v1beta/models/{model}:generateContent",
-        "auth_header": "x-goog-api-key",
-        "auth_prefix": "",
-    },
-    "mistral": {
-        "endpoint": "https://api.mistral.ai/v1/chat/completions",
-        "auth_header": "Authorization",
-        "auth_prefix": "Bearer ",
-    },
-    "perplexity": {
-        "endpoint": "https://api.perplexity.ai/chat/completions",
-        "auth_header": "Authorization",
-        "auth_prefix": "Bearer ",
-    },
-    "xai": {
-        "endpoint": "https://api.x.ai/v1/chat/completions",
-        "auth_header": "Authorization",
-        "auth_prefix": "Bearer ",
-    },
-    "huggingface": {
-        "endpoint": "https://router.huggingface.co/v1/chat/completions",
-        "auth_header": "Authorization",
-        "auth_prefix": "Bearer ",
-    },
-    "huggingface-together": {
-        "endpoint": "https://router.huggingface.co/together/v1/chat/completions",
-        "auth_header": "Authorization",
-        "auth_prefix": "Bearer ",
-    },
-    "ollama": {
-        "endpoint": "http://localhost:11434/v1/chat/completions",
-        "auth_header": None,  # No auth required for local Ollama
-        "auth_prefix": "",
-    },
-}
-
-
-# =============================================================================
-# Unified API Client
-# =============================================================================
-
-class UnifiedLLMClient:
[…]
