### 2026-02-11 e3652d2
Pin HuggingFace Space apps to cat-llm>=2.3.4
Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
