### 2026-02-11 5e86fee
Validate JSON schema in aggregate_results and default missing keys to 0
- aggregate_results() now validates that parsed JSON contains numbered keys
  within the category range (1-N) with valid 0/1 values; wrong key formats
  (e.g. category names) are treated as model failures
- Output builders default missing numbered keys to 0 instead of None,
  matching consensus behavior (omitted key = category not present)

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
