### 2026-02-11 d4c0061
Add schema validation to batch retry failure detection
Batch retries now detect JSON schema failures (wrong key format, out-of-range
keys, invalid values) in addition to API errors and parse failures. The retry
success counter also validates schema, giving accurate reporting.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
