### 2026-02-11 b0b0a7e
Add missing_keys tracking and classification quality summary
Tracks which models had missing JSON keys (filled with 0) and
reports a summary after classification completes, including
per-model breakdown of filled values and schema failures.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
