### 2026-02-11 fab0e1f
Add quality summary, failed-model guard, and revert to lenient validation
- Output builders now check failed_models set: failed models get None/NA,
  valid models get 0-fill for missing keys (prevents false all-zeros)
- Classification Quality Summary prints filled counts and schema failures
  per model after classification completes
- Revert validation to lenient (valid_count > 0): partial numbered keys
  are valid responses with missing keys filled as 0 (omission = not present)
- aggregate_results returns missing_keys metadata for tracking

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.c
[…]
