### 2026-02-11 2a15342
Clean invalid keys from parsed responses, bump to v2.4.0
Responses with at least one valid category key are accepted, but
invalid keys (e.g. "yes", "cats") are now stripped before storing
in successful results. This prevents garbage values from becoming
phantom 0 votes in consensus — missing/invalid keys cleanly default
to 0 via the missing-key path. Failed models produce None/NA in
output CSVs instead of silent zeros.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
