### 2026-02-18 d9dd4b1
Document thinking_budget parameter in classify() API reference
Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
