### 2026-02-19 31e4816
Fix NaN row handling and add row_delay parameter, bump to v2.4.1
Skipped rows (NaN input) no longer falsely list all models as failed.
Add row_delay parameter to classify() for rate limit mitigation when
multiple models share the same API provider.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
