### 2026-02-26 b998adb
Add has_other_category() and add_other parameter in classify()
Detect whether user categories include a catch-all "Other" option and
prompt to add one if missing. Empirical findings show this improves
classification accuracy by preventing forced misclassification of
ambiguous responses.

- New _category_analysis.py with two-tier heuristic (anchored + short
  phrase patterns) and optional LLM fallback for ambiguous cases
- classify() gains add_other param: "prompt" (default), True, or False
- Exported has_other_category from catllm package

Co-Authored-By: Claude Sonnet 4.5 <noreply@anthropic.c
[…]
