+++ src/catllm/_category_analysis.py
@@ -1,0 +1,193 @@
+"""
+Category analysis utilities for CatLLM.
+
+Provides functions for analyzing user-provided category lists,
+such as detecting whether an "Other" catch-all category exists.
+"""
+
+import re
+
+from .text_functions import UnifiedLLMClient, detect_provider
+
+__all__ = ["has_other_category"]
+
+# Max words for a category to be checked against broad phrase patterns.
+# Real catch-all categories are short ("Other", "None of the above", "Does not fit").
+# Longer categories using these words ("Does not fit the clinical profile") are
+# specific descriptive labels, not catch-alls.
+_MAX_HEURISTIC_WORDS = 4
+
+# Tier 1: Anchored patterns — safe at any category length.
+# These only match when the keyword IS the category label itself.
+_ANCHORED_PATTERNS = [
+    re.compile(r"^other\s*$", re.IGNORECASE),         # exact "Other"
+    re.compile(r"^other\s*[:(]", re.IGNORECASE),       # "Other: ...", "Other (..."
+    re.compile(r"^n/?a\s*$", re.IGNORECASE),           # exact "N/A", "NA"
+    re.compile(r"^miscellaneous\s*$", re.IGNORECASE),  # exact "Miscellaneous"
+    re.compile(r"^catch[\s-]?all\s*$", re.IGNORECASE), # exact "catch-all"
+]
+
+# Tier 2: Phrase patterns — only applied to short categories (≤ _MAX_HEURISTIC_WORDS).
+# Multi-word phrases that clearly signal a catch-all when they dominate the category name.
+_SHORT_ONLY_PATTERNS = [
+    re.compile(r"\bnone of the above\b", re.IGNORECASE),
+    re.compile(r"\bdoes not fit\b", re.IGNORECASE),
+    re.compile(r"\bdoesn't fit\b", re.IGNORECASE),
+    re.compile(r"\bnot applicable\b", re.IGNORECASE),
+    re.compile(r"\bnone apply\b", re.IGNORECASE),
+    re.compile(r"\bnone of these\b", re.IGNORECASE),
+]
+
+# Top-tier model per provider for the LLM fallback
+_TOP_TIER_MODELS = {
+    "openai": "gpt-4o",
+    "anthropic": "claude-sonnet-4-5-20250929",
+    "google": "gemini-2.5-flash",
+    "mistral": "mistral-large-latest",
+    "xai": "grok-2",
+    "perplexity": "sonar-pro",
+    "huggingface": "meta-llama/Llama-3.3-70B-Instruct",
+}
+
+
+def _heuristic_check(categories: list) -> bool:
+    """
+    Fast, free check for common "Other" category patterns.
+
+    Uses a two-tier approach to avoid false positives:
+      - Tier 1 (anchored): matches at any length — the pattern is specific enough
+        (e.g. exact "Other", "N/A", or "Other: …" label prefix).
+      - Tier 2 (phrase): only matches short categories (≤ _MAX_HEURISTIC_WORDS words).
+        Phrases like "does not fit" are catch-alls when they ARE the category, but
+        not when embedded in longer descriptions ("Does not fit the clinical profile").
+
+    Returns True if any category matches a known catch-all pattern.
+    """
+    for cat in categories:
+        cat_str = str(cat).strip()
+
+        # Tier 1: anchored patterns — safe at any length
+        for pattern in _ANCHORED_PATTERNS:
+            if pattern.search(cat_str):
+                return True
+
+        # Tier 2: phrase patterns — only for short categories
+        if len(cat_str.split()) <= _MAX_HEURISTIC_WORDS:
+            for pattern in _SHORT_ONLY_PATTERNS:
+                if pattern.search(cat_str):
+                    return True
+
+    return False
+
+
[…]
