### 2026-02-26 99301a5
Add category verbosity check and evidence-based strategy warnings
- check_category_verbosity(): LLM-based assessment of whether each
  category has descriptions and examples (1 API call, per-category flags)
- check_verbosity param in classify(): alerts users when categories
  lack descriptions/examples — verbose definitions improve accuracy ~7 pp
- Evidence-based warnings when users enable strategies shown to be
  ineffective or harmful: CoVe (-2 pp, 4x cost), few-shot (-1 pp),
  thinking (negligible, high failure), CoT (no effect), step-back
  (inconsistent, hurts top-tier)
- Warnings are inf
[…]
