+++ src/catllm/classify.py
@@ -26,7 +26,7 @@ from ._providers import (
 )
 
 # Category analysis
-from ._category_analysis import has_other_category
+from ._category_analysis import has_other_category, check_category_verbosity
 
 # Import the implementation functions from existing modules
 from .text_functions_ensemble import (
@@ -87,6 +87,7 @@ def classify(
     pdf_dpi: int = 150,
     auto_download: bool = False,
     add_other = "prompt",
+    check_verbosity: bool = True,
 ):
     """
     Unified classification function for text, image, and PDF inputs.
@@ -151,6 +152,10 @@ def classify(
             - "prompt" (default): Ask the user to accept or reject the suggestion.
             - True: Silently add "Other" without prompting.
             - False: Never add "Other".
+        check_verbosity (bool): Check whether each category has a description
+            and examples (1 API call). Verbose categories with descriptions and
+            examples significantly improve classification accuracy over bare
+            labels. Default True. Set to False to skip.
 
     Returns:
         pd.DataFrame: Results with classification columns.
@@ -210,6 +215,115 @@ def classify(
                     f"Categories are now: {categories}  "
                     f"(set add_other=False to disable)"
                 )
+
+    # Check category verbosity (1 API call)
+    # TODO: Offer to auto-generate verbose versions of bare categories using
+    # the LLM (description + examples) and let the user accept/edit them.
+    if check_verbosity and categories and categories != "auto":
+        # Extract API key and provider from first model entry
+        first_entry = models[0]
+        check_key = first_entry[2] if len(first_entry) >= 3 else None
+        check_source = first_entry[1] if len(first_entry) >= 2 else "auto"
+
+        if check_key:
+            try:
+                verbosity = check_category_verbosity(
+                    categories,
+                    api_key=check_key,
+                    model_source=check_source,
+                )
+                lacking = [r for r in verbosity if not r["is_verbose"]]
+
+                if lacking:
+                    missing_desc = [r for r in lacking if not r["has_description"]]
+                    missing_ex = [r for r in lacking if not r["has_examples"]]
+
+                    print(
+                        "\n[CatLLM] Category verbosity check (set check_verbosity=False to skip):"
+                    )
+                    for r in lacking:
+                        issues = []
+                        if not r["has_description"]:
+                            issues.append("description")
+                        if not r["has_examples"]:
+                            issues.append("examples")
+                        print(f'  - "{r["category"]}"  (missing: {", ".join(issues)})')
+
+                    print(
+                        "\n  Verbose categories with descriptions and examples significantly\n"
+                        "  improve classification accuracy over bare labels.\n"
+                        "\n"
+                        "  Instead of:\n"
+                        '    "Positive"\n'
+                        "  Consider:\n"
+                        '    "Positive: The response expresses satisfaction, approval, or\n'
+                        "     happiness (e.g., 'I love this product', 'Great experience',\n"
+                        "     'Very pleased with the result')\"\n"
+                    )
+            except Exception:
+                pass  # Non-critical — don't block classification
+
+    # =========================================================================
+    # Evidence-based warnings for prompting strategies
+    # Based on empirical findings from Soria et al. (2026) comparing prompting
+    # strategies across 4 representative models and 4 survey tasks.
+    # =========================================================================
+    _strategy_warnings = []
+
+    if chain_of_verification:
+        _strategy_warnings.append(
+            "[CatLLM] WARNING: chain_of_verification=True is enabled.\n"
[…]
