+++ src/catllm/_category_analysis.py
@@ -5,11 +5,12 @@ Provides functions for analyzing user-provided category lists,
 such as detecting whether an "Other" catch-all category exists.
 """
 
+import json
 import re
 
 from .text_functions import UnifiedLLMClient, detect_provider
 
-__all__ = ["has_other_category"]
+__all__ = ["has_other_category", "check_category_verbosity"]
 
 # Max words for a category to be checked against broad phrase patterns.
 # Real catch-all categories are short ("Other", "None of the above", "Does not fit").
@@ -125,6 +126,20 @@ def _llm_check(categories: list, api_key: str, model: str, provider: str) -> boo
     return answer in ("yes", "true")
 
 
+def _resolve_provider_and_model(user_model, model_source):
+    """Resolve provider and model from user args, falling back to top-tier defaults."""
+    if user_model is not None:
+        provider = detect_provider(user_model, provider=model_source)
+        model = user_model
+    else:
+        if model_source and model_source.lower() != "auto":
+            provider = model_source.lower()
+        else:
+            provider = "openai"
+        model = _TOP_TIER_MODELS.get(provider, "gpt-4o")
+    return provider, model
+
+
 def has_other_category(
     categories: list,
     api_key: str = None,
@@ -178,16 +193,160 @@ def has_other_category(
     if api_key is None:
         return False
 
-    # Resolve provider and model
-    if user_model is not None:
-        provider = detect_provider(user_model, provider=model_source)
-        model = user_model
-    else:
-        # No model specified — pick a default
-        if model_source and model_source.lower() != "auto":
-            provider = model_source.lower()
-        else:
-            provider = "openai"
-        model = _TOP_TIER_MODELS.get(provider, "gpt-4o")
-
+    provider, model = _resolve_provider_and_model(user_model, model_source)
     return _llm_check(categories, api_key, model, provider)
+
+
+# =============================================================================
+# Category Verbosity Check
+# TODO: Add a function that auto-generates verbose category definitions
+# (description + examples) from bare labels, letting the user review/edit.
+# TODO: Consider caching verbosity results per category list to avoid
+# redundant API calls across repeated classify() invocations.
+# =============================================================================
+
+def check_category_verbosity(
+    categories: list,
+    api_key: str,
+    user_model: str = None,
+    model_source: str = "auto",
+) -> list:
+    """
+    Assess whether each category has a clear description and illustrative examples.
+
+    Makes a single LLM call to evaluate all categories at once. Returns per-category
+    flags indicating what's present and what's missing.
+
+    Args:
+        categories: List of category strings to analyze.
+        api_key: API key for the LLM provider (required).
+        user_model: Model name to use. If not provided, a top-tier default is
+                    selected based on the provider.
+        model_source: Provider (e.g. "openai", "anthropic", "google").
+                      Defaults to "auto".
+
+    Returns:
+        A list of dicts, one per category, each containing::
+
+            {
+                "category": str,         # the original category text
+                "has_description": bool,  # has an explanation beyond a bare label
+                "has_examples": bool,     # includes concrete examples
+                "is_verbose": bool,       # True if BOTH description and examples present
+            }
+
+    Examples:
+        >>> check_category_verbosity(
+        ...     ["Positive", "Negative: expresses dissatisfaction (e.g., 'I hate this')"],
+        ...     api_key="sk-...",
+        ... )
+        [
+            {"category": "Positive", "has_description": False, "has_examples": False, "is_verbose": False},
+            {"category": "Negative: ...", "has_description": True, "has_examples": True, "is_verbose": True},
+        ]
+    """
+    if not categories:
+        return []
+
+    provider, model = _resolve_provider_and_model(user_model, model_source)
+
[…]
