### 2026-02-26 b084639
Bump to v2.5.0 for PyPI release
New features: category analysis (has_other_category, check_category_verbosity),
add_other prompt, check_verbosity alert, evidence-based strategy warnings.

Co-Authored-By: Claude Sonnet 4.5 <noreply@anthropic.com>
