### 2026-02-27 bdaf607
Add catllm R package (reticulate wrapper)
R interface to the CatLLM Python package, installable via
devtools::install_github("chrissoria/cat-llm", subdir = "r-package/catllm").
Wraps classify(), extract(), and explore() with native R syntax.

Co-Authored-By: Claude Sonnet 4.5 <noreply@anthropic.com>
