+++ r-package/catllm/tests/testthat.R
@@ -1,0 +1,6 @@
+# This file is the standard testthat entry point generated/maintained by
+# usethis::use_testthat(). Do not edit by hand.
+library(testthat)
+library(catllm)
+
+test_check("catllm")
