### 2026-03-01 6f7041e
Add Claude Code plugin structure for CatLLM
Convert CatLLM into an installable Claude Code plugin with:
- Plugin manifest (.claude-plugin/plugin.json)
- Skills for classify, extract, explore, summarize, and image-score
- Session start hook to auto-install catllm
- CLAUDE.md with project context for Claude Code
- Plugin settings for bash permissions

Users can now install via: claude --plugin-dir ./cat-llm
Or distribute through a Claude Code plugin marketplace.

https://claude.ai/code/session_01NJ2B1RRgo5Vin9LYQS7nxv
