### 2026-03-01 212cd7d
Fix Stata package SFI API calls and Python summarize bugs
Stata package fixes:
- Replace non-existent Data.getNum()/getStr()/setStr() with
  Data.getAt() and Data.storeAt() across all four .ado files
- Fix r() macro storage: use Macro.setGlobal("r(...)") instead of
  setLocal() in catllm_extract and catllm_explore
- Fix category column mapping in catllm_classify: columns returned as
  positional (category_1, category_2, ...) are now mapped back to
  the user-provided category names in order

Python core fixes:
- Fix summarize_ensemble: client.complete() returns (text, error) tuple;
  unpack it
[…]
