### 2026-03-01 d3c0a94
Bump to v2.5.1: Stata fixes, summarize bug fixes, chunk_delay, unanimous default
- Add chunk_delay parameter to explore() and extract() for rate limiting
- Change consensus_threshold default from "majority" to "unanimous"
- Update R package DESCRIPTION with system requirements
- Bump version to 2.5.1

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
