+++ r-package/catllm/DESCRIPTION
@@ -4,17 +4,17 @@ Version: 0.1.0
 Authors@R:
     person("Chris", "Soria", email = "chrissoria@berkeley.edu",
            role = c("aut", "cre"))
-Author: Chris Soria [aut, cre]
-Maintainer: Chris Soria <chrissoria@berkeley.edu>
 Description: R interface to the 'catllm' Python package for LLM-powered text
     classification, category extraction, and corpus exploration. Wraps the
     Python classify(), extract(), and explore() functions via 'reticulate',
     allowing R users to leverage large language models for qualitative data
-    analysis without writing Python code.
+    analysis without writing Python code. Requires Python (>= 3.8) and the
+    'catllm' Python package, which can be installed with install_catllm().
 License: MIT + file LICENSE
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.2
+SystemRequirements: Python (>= 3.8), catllm Python package
 Imports:
     reticulate (>= 1.28)
 Suggests:
