### 2026-03-04 2a79863
Refactor CERAD_functions.py to use classify() instead of image_multi_class
Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
