### 2026-03-04 4b8825e
Update CHANGELOG for CERAD classify() refactor
Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
