### 2026-03-04 #9678 @chrissoria
You can test it here if you want to run without API fees: https://huggingface.co/spaces/CatLLM/survey-classifier
