### 2026-03-04 e5c8616
Add TODO for provider-agnostic batch_mode parameter
Documents batch API support across OpenAI, Anthropic, Google, Mistral,
and xAI — all offering 50% cost savings and JSONL-based job submission.
Notes HuggingFace and Perplexity as unsupported.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
