### 2026-03-04 df389bf
Add batch_mode=True to classify() — 50% cost savings via async batch APIs
Implements provider-agnostic async batch inference for OpenAI, Anthropic,
Google, Mistral, and xAI. All five providers offer 50% cost reduction and
higher rate limits via a JSONL-upload → poll → download pattern.

New files:
- src/catllm/_batch.py: Core batch logic — JSONL building, file upload,
  job creation, polling, result download, result parsing. Pure HTTP,
  no provider SDKs. Exports BatchJobExpiredError, BatchJobFailedError.

Changes:
- classify(): Add batch_mode, batch_poll_interval, batch_timeout params.
  Guar
[…]
