+++ examples/test_batch_mode.py
@@ -1,0 +1,327 @@
+#!/usr/bin/env python
+"""
+Test script for batch_mode=True in catllm.classify().
+
+Tests:
+  1. Smoke test per supported provider (5 rows each)
+  2. Output format parity vs. synchronous path (column names, dtypes, value range)
+  3. ValueError guard conditions (ensemble, image input, unsupported provider)
+  4. Partial failure handling (bad row doesn't crash the job)
+
+Usage:
+  python examples/test_batch_mode.py                  # all providers
+  python examples/test_batch_mode.py openai           # single provider
+  python examples/test_batch_mode.py guards           # guard-condition tests only
+"""
+
+import os
+import sys
+
+# Load from local src
+src_path = os.path.join(os.path.dirname(__file__), '..', 'src')
+sys.path.insert(0, os.path.abspath(src_path))
+
+modules_to_remove = [k for k in sys.modules if k.startswith('catllm')]
+for m in modules_to_remove:
+    del sys.modules[m]
+
+from dotenv import load_dotenv
+load_dotenv('/Users/chrissoria/Documents/Research/Categorization_AI_experiments/.env', override=True)
+
+import pandas as pd
+import catllm
+
+# =============================================================================
+# Config
+# =============================================================================
+
+CATEGORIES = [
+    "Mentions cost or affordability",
+    "Mentions quality or performance",
+    "Mentions customer service",
+    "Other",
+]
+DESCRIPTION = "Survey responses about a consumer product"
+
+SAMPLE_TEXTS = [
+    "The price is too high for what you get.",
+    "Great quality, really durable and well made.",
+    "Customer support was very helpful and responsive.",
+    "I just bought it yesterday, no opinion yet.",
+    "Expensive but the quality makes it worth it.",
+]
+
+API_KEYS = {
+    "openai":    os.getenv("OPENAI_API_KEY"),
+    "anthropic": os.getenv("ANTHROPIC_API_KEY"),
+    "google":    os.getenv("GOOGLE_API_KEY"),
+    "mistral":   os.getenv("MISTRAL_API_KEY"),
+    "xai":       os.getenv("XAI_API_KEY"),
+}
+
+MODELS = {
+    "openai":    "gpt-4o-mini",
+    "anthropic": "claude-haiku-4-5-20251001",
+    "google":    "gemini-2.5-flash",
+    "mistral":   "mistral-small-latest",
+    "xai":       "grok-3-mini",
+}
+
+# =============================================================================
+# Helpers
+# =============================================================================
+
+def print_section(title):
+    print()
+    print("=" * 70)
+    print(f"  {title}")
+    print("=" * 70)
+
+def print_result(df):
+    print(df.to_string())
+    print(f"\nShape: {df.shape}  |  Columns: {list(df.columns)}")
+
+def assert_valid_df(df, label=""):
+    """Assert the DataFrame matches single-model synchronous output format."""
+    prefix = f"[{label}] " if label else ""
+
+    # Must be a DataFrame
+    assert isinstance(df, pd.DataFrame), f"{prefix}Expected DataFrame, got {type(df)}"
+
+    # Must have category columns named category_1, category_2, ... (no model suffix)
+    cat_cols = [c for c in df.columns if c.startswith("category_")]
+    assert len(cat_cols) == len(CATEGORIES), (
+        f"{prefix}Expected {len(CATEGORIES)} category columns, got {len(cat_cols)}: {cat_cols}"
+    )
+    for i in range(1, len(CATEGORIES) + 1):
+        assert f"category_{i}" in df.columns, (
+            f"{prefix}Missing column category_{i}. Columns: {list(df.columns)}"
+        )
+
[…]
