### 2026-03-04 186b773
Add batch_mode parameter to README (experimental)
Documents batch_mode, batch_poll_interval, and batch_timeout in classify()
API reference with supported providers and constraints noted.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
