### 2026-03-04 d3e6be1
Add MASTERPLAN.md: tidyverse-style ecosystem vision
Documents the roadmap for cat-llm as a meta-package pulling in
cat-survey, cat-vader, cat-ademic, and cat-pol sub-packages.
Includes migration phases, shared infrastructure options, and
API consistency principles.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
