+++ MASTERPLAN.md
@@ -1,0 +1,183 @@
+# CatLLM Ecosystem — Master Plan
+
+## Vision
+
+**cat-llm** evolves from a monolithic package into a meta-package — the tidyverse of
+LLM-powered social science text analysis. Installing `cat-llm` pulls in a family of
+domain-specific sub-packages, each with its own PyPI release, versioning, and focus,
+but sharing a common provider infrastructure and a consistent API surface.
+
+```
+pip install cat-llm
+# brings in:
+#   cat-survey    survey response classification & extraction
+#   cat-vader     social media (Reddit, Twitter/X, forums)
+#   cat-ademic    academic papers, PDFs, citations
+#   cat-pol       political text (manifestos, speeches, legislation)
+```
+
+---
+
+## The Sub-Packages
+
+### cat-survey *(extract from current cat-llm core)*
+Survey response classification and category extraction.
+This is the current heart of `cat-llm` — it gets spun out as its own package.
+
+**Scope:**
+- `classify()` for open-ended survey responses (text only)
+- `extract()` / `explore()` for category discovery
+- `summarize()` for response-level summarization
+- All prompt strategies tested in the paper (CoT, CoVe, step-back, few-shot)
+- R and Stata wrappers live here
+
+**Status:** Exists inside `cat-llm/src/catllm/`. Needs to be extracted into its own repo
+(`cat-survey`) and published separately, then re-imported by `cat-llm`.
+
+---
+
+### cat-vader *(already separate)*
+Social media text classification with platform-aware context injection.
+
+**Scope:**
+- `classify()` / `extract()` / `explore()` with social media metadata fields
+  (platform, handle, hashtags, engagement metrics injected into prompt)
+- Reddit thread classification (nested comment structure)
+- HuggingFace Space: `CatLLM/CatVader`
+
+**Status:** Separate repo and PyPI package. Needs to be wired into `cat-llm` as a
+declared dependency and re-exported.
+
+---
+
+### cat-ademic *(planned)*
+Classification and extraction for academic and long-form documents.
+
+**Scope:**
+- PDF-first classify/extract — builds on current `pdf_functions.py` logic
+- Per-page and whole-document classification modes
+- Citation and abstract extraction
+- CERAD scoring (currently in `cat-llm` — natural home here given image + PDF focus)
+- Designed for systematic reviews, coding codebooks, content analysis of papers
+
+**Status:** Not yet started. Core PDF machinery exists in `cat-llm/src/catllm/pdf_functions.py`
+and `image_functions.py` — these become the seed.
+
+---
+
+### cat-pol *(future)*
+Political text analysis — manifestos, speeches, legislation, news.
+
+**Scope:**
+- Domain-tuned prompts for political science categories
+  (ideology, policy area, sentiment toward government, actor identification)
+- Multi-language support (EU parliament, UN documents)
+- Manifesto coding scheme compatibility (MARPOR/CMP)
+- Time-series classification across document corpora
+
+**Status:** Concept only. No code yet.
+
+---
+
+## Shared Infrastructure
+
+All sub-packages depend on a common base layer. Two options:
+
+### Option A — `catcore` (dedicated base package)
+Extract `_providers.py`, `_batch.py`, `text_functions.py`, and `build_json_schema` /
+`extract_json` utilities into a standalone `catcore` package. Each sub-package lists
+`catcore` as a dependency, not `cat-llm`.
+
+**Pros:** Clean separation, `cat-llm` is a true thin meta-package.
+**Cons:** Another repo/package to maintain; versioning coordination overhead.
+
+### Option B — `cat-llm` remains the provider layer
[…]
