### 2026-03-04 08104aa
Update MASTERPLAN: adopt tidyverse-style catcore architecture
Switch from Option B to Option A: catcore becomes the shared
infrastructure layer (like rlang), sub-packages depend on catcore
directly, and cat-llm is a pure meta-package with no domain logic.
Migration path updated to 5 phases.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
