+++ MASTERPLAN.md
@@ -79,52 +79,71 @@ Political text analysis — manifestos, speeches, legislation, news.
 
 ---
 
-## Shared Infrastructure
+## Shared Infrastructure — `catcore`
 
-All sub-packages depend on a common base layer. Two options:
+Modeled on how tidyverse handles shared infrastructure (e.g., `rlang`, `cli`):
+a dedicated base package that sub-packages depend on directly, with the meta-package
+(`cat-llm`) knowing about sub-packages but sub-packages having no knowledge of the
+meta-package.
 
-### Option A — `catcore` (dedicated base package)
-Extract `_providers.py`, `_batch.py`, `text_functions.py`, and `build_json_schema` /
-`extract_json` utilities into a standalone `catcore` package. Each sub-package lists
-`catcore` as a dependency, not `cat-llm`.
+**`catcore`** contains:
+- `_providers.py` — `UnifiedLLMClient`, `PROVIDER_CONFIG`, `detect_provider`, Ollama utils
+- `_batch.py` — async batch API logic for all providers
+- `text_functions.py` — `build_json_schema`, `extract_json`, `validate_classification_json`
 
-**Pros:** Clean separation, `cat-llm` is a true thin meta-package.
-**Cons:** Another repo/package to maintain; versioning coordination overhead.
+Each sub-package lists `catcore` as a dependency. `cat-llm` lists all sub-packages
+as dependencies and is otherwise empty — just an `__init__.py` that imports and
+re-exports everything. This mirrors exactly how `tidyverse` works in R.
 
-### Option B — `cat-llm` remains the provider layer
-`cat-llm` keeps `_providers.py`, `_batch.py`, and shared utilities. Sub-packages list
-`cat-llm` as a dependency. `cat-llm` re-exports sub-package entry points.
+**Dependency graph:**
 
-**Pros:** Fewer repos; provider upgrades propagate automatically.
-**Cons:** Circular dependency risk if sub-packages ever need to import each other.
+```
+catcore                        ← no cat-* deps (leaf)
+    ↑
+cat-survey  cat-vader  cat-ademic  cat-pol    ← each depends only on catcore
+    ↑           ↑          ↑          ↑
+                    cat-llm                    ← pure meta-package
+                                               (depends on all four + catcore)
+```
 
-**Current leaning: Option B** — simpler to execute now. Revisit when there are 3+
-active sub-packages and provider churn becomes painful.
+This means every sub-package is independently installable:
+```
+pip install cat-survey   # works standalone, no cat-llm required
+pip install cat-llm      # installs everything
+```
 
 ---
 
 ## Target Package Structure
 
 ```
-cat-llm/                      ← meta-package + shared provider layer
-├── src/catllm/
-│   ├── _providers.py         ← UnifiedLLMClient, PROVIDER_CONFIG (shared)
-│   ├── _batch.py             ← batch API logic (shared)
-│   ├── text_functions.py     ← shared text utilities
-│   └── __init__.py           ← re-exports from cat-survey, cat-vader, cat-ademic, cat-pol
+catcore/                      ← shared infrastructure (like rlang in R)
+├── src/catcore/
+│   ├── _providers.py         ← UnifiedLLMClient, PROVIDER_CONFIG
+│   ├── _batch.py             ← batch API logic
+│   └── text_functions.py     ← shared text utilities
 
-cat-survey/                   ← survey-specific package
+cat-survey/                   ← survey response package
+│   depends on: catcore
 │   classify(), extract(), explore(), summarize()
 │   R + Stata wrappers
 
[…]
