### 2026-03-04 86f7355
Update MASTERPLAN: cat-content base + cat-llm as meta-package
cat-llm keeps its brand as the meta-package/flagship.
cat-content is the new general-purpose base (like rlang) — both
shared infrastructure and independently useful. Domain packages
(cat-survey, cat-vader, cat-ademic, cat-pol) depend on cat-content.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
