+++ MASTERPLAN.md
@@ -2,41 +2,83 @@
 
 ## Vision
 
-**cat-llm** evolves from a monolithic package into a meta-package — the tidyverse of
-LLM-powered social science text analysis. Installing `cat-llm` pulls in a family of
-domain-specific sub-packages, each with its own PyPI release, versioning, and focus,
-but sharing a common provider infrastructure and a consistent API surface.
+**cat-llm** keeps its brand and identity as the meta-package — the tidyverse of
+LLM-powered text analysis for social science. Installing `cat-llm` pulls in a family
+of domain-specific sub-packages, each with its own PyPI release, versioning, and focus.
+
+The shared infrastructure and general-purpose base lives in **`cat-content`**: a
+standalone package for classifying any text column, with no domain assumptions.
+Domain-specific packages (`cat-survey`, `cat-vader`, etc.) extend `cat-content`
+with domain-tuned prompts, metadata injection, and workflow helpers.
 
 ```
 pip install cat-llm
 # brings in:
+#   cat-content   general-purpose text column classification (the base)
 #   cat-survey    survey response classification & extraction
 #   cat-vader     social media (Reddit, Twitter/X, forums)
 #   cat-ademic    academic papers, PDFs, citations
 #   cat-pol       political text (manifestos, speeches, legislation)
 ```
 
+**Dependency graph** (modeled on tidyverse / rlang):
+
+```
+cat-content                    ← general base + shared infra (like rlang)
+                                  independently useful; no domain assumptions
+    ↑
+cat-survey  cat-vader  cat-ademic  cat-pol    ← domain packages, each depends on cat-content
+    ↑           ↑          ↑          ↑
+                    cat-llm                    ← pure meta-package, keeps the brand
+                                               (depends on all domain packages + cat-content)
+```
+
+Every package is independently installable:
+```
+pip install cat-content   # general text classification, no domain framing
+pip install cat-survey    # survey-specific, pulls in cat-content automatically
+pip install cat-llm       # everything
+```
+
 ---
 
-## The Sub-Packages
+## The Packages
 
-### cat-survey *(extract from current cat-llm core)*
-Survey response classification and category extraction.
-This is the current heart of `cat-llm` — it gets spun out as its own package.
+### `cat-content` *(new — the base)*
+General-purpose text column classification and extraction with no domain assumptions.
+This replaces the need for a separate invisible `catcore` — it is both the shared
+infrastructure layer AND independently useful for researchers who just have a text
+column and don't fit neatly into a domain.
 
 **Scope:**
-- `classify()` for open-ended survey responses (text only)
-- `extract()` / `explore()` for category discovery
-- `summarize()` for response-level summarization
-- All prompt strategies tested in the paper (CoT, CoVe, step-back, few-shot)
-- R and Stata wrappers live here
+- `classify()`, `extract()`, `explore()`, `summarize()` for any text input
+- `UnifiedLLMClient`, `PROVIDER_CONFIG`, `detect_provider` — all provider infrastructure
+- `_batch.py` — async batch API logic
+- Shared text utilities: `build_json_schema`, `extract_json`, `validate_classification_json`
+- No survey framing, no social media framing, no PDF-first assumptions
 
-**Status:** Exists inside `cat-llm/src/catllm/`. Needs to be extracted into its own repo
-(`cat-survey`) and published separately, then re-imported by `cat-llm`.
[…]
