### 2026-03-04 4ed1c24
Rename cat-content → cat-stack in MASTERPLAN
Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
