+++ MASTERPLAN.md
@@ -6,15 +6,15 @@
 LLM-powered text analysis for social science. Installing `cat-llm` pulls in a family
 of domain-specific sub-packages, each with its own PyPI release, versioning, and focus.
 
-The shared infrastructure and general-purpose base lives in **`cat-content`**: a
+The shared infrastructure and general-purpose base lives in **`cat-stack`**: a
 standalone package for classifying any text column, with no domain assumptions.
-Domain-specific packages (`cat-survey`, `cat-vader`, etc.) extend `cat-content`
+Domain-specific packages (`cat-survey`, `cat-vader`, etc.) extend `cat-stack`
 with domain-tuned prompts, metadata injection, and workflow helpers.
 
 ```
 pip install cat-llm
 # brings in:
-#   cat-content   general-purpose text column classification (the base)
+#   cat-stack   general-purpose text column classification (the base)
 #   cat-survey    survey response classification & extraction
 #   cat-vader     social media (Reddit, Twitter/X, forums)
 #   cat-ademic    academic papers, PDFs, citations
@@ -24,19 +24,19 @@ pip install cat-llm
 **Dependency graph** (modeled on tidyverse / rlang):
 
 ```
-cat-content                    ← general base + shared infra (like rlang)
+cat-stack                    ← general base + shared infra (like rlang)
                                   independently useful; no domain assumptions
     ↑
-cat-survey  cat-vader  cat-ademic  cat-pol    ← domain packages, each depends on cat-content
+cat-survey  cat-vader  cat-ademic  cat-pol    ← domain packages, each depends on cat-stack
     ↑           ↑          ↑          ↑
                     cat-llm                    ← pure meta-package, keeps the brand
-                                               (depends on all domain packages + cat-content)
+                                               (depends on all domain packages + cat-stack)
 ```
 
 Every package is independently installable:
 ```
-pip install cat-content   # general text classification, no domain framing
-pip install cat-survey    # survey-specific, pulls in cat-content automatically
+pip install cat-stack   # general text classification, no domain framing
+pip install cat-survey    # survey-specific, pulls in cat-stack automatically
 pip install cat-llm       # everything
 ```
 
@@ -44,7 +44,7 @@ pip install cat-llm       # everything
 
 ## The Packages
 
-### `cat-content` *(new — the base)*
+### `cat-stack` *(new — the base)*
 General-purpose text column classification and extraction with no domain assumptions.
 This replaces the need for a separate invisible `catcore` — it is both the shared
 infrastructure layer AND independently useful for researchers who just have a text
@@ -88,7 +88,7 @@ Social media text classification with platform-aware context injection.
 - HuggingFace Space: `CatLLM/CatVader`
 
 **Status:** Separate repo and PyPI package. Needs to be updated to depend on
-`cat-content` instead of carrying its own provider copy, then wired into `cat-llm`.
+`cat-stack` instead of carrying its own provider copy, then wired into `cat-llm`.
 
 ---
 
@@ -137,8 +137,8 @@ but installing it gives you everything.
 ## Target Package Structure
 
 ```
-cat-content/                  ← general base + shared infra (like rlang)
-├── src/catcontent/
+cat-stack/                  ← general base + shared infra (like rlang)
+├── src/catstack/
 │   ├── _providers.py         ← UnifiedLLMClient, PROVIDER_CONFIG
 │   ├── _batch.py             ← batch API logic
 │   ├── text_functions.py     ← shared text utilities
[…]
