### 2026-03-05 2d67605
Fix Google batch file upload: unwrap response.file.name
Google Files API wraps metadata under a "file" key in the upload
response; we were reading the top-level dict which has no "name".

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
