### 2026-03-05 5eda3c5
Add experimental ensemble batch mode (batch_mode=True with multiple models)
Each model now submits its own batch job concurrently when batch_mode=True
is passed with a multi-model ensemble. Results are merged through the
existing aggregate_results + build_output_dataframes pipeline, returning
the same DataFrame format as synchronous ensemble (per-model columns,
_consensus, _agreement).

Providers without a batch API (HuggingFace, Perplexity, Ollama) fall back
to synchronous per-item calls and are merged in with the batch results.

New internals in _batch.py:
- _run_one_batch_job: extracted fro
[…]
