### 2026-03-06 77f212d
Add json_formatter fallback for malformed LLM classification output
Adds opt-in `json_formatter=True` parameter to classify() that uses a
local fine-tuned Qwen2.5-0.5B model (chrissoria/catllm-json-formatter)
to fix invalid JSON before marking responses as failed. The formatter
runs after extract_json() at all 4 extraction points (text, PDF, image,
Ollama two-step) and before batch retries — zero cost on the happy path.

Install with: pip install cat-llm[formatter]

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
