+++ pyproject.toml
@@ -35,6 +35,7 @@ dependencies = [
 
 [project.optional-dependencies]
 pdf = ["PyMuPDF>=1.23.0"]
+formatter = ["torch>=2.0.0", "transformers>=4.40.0", "accelerate>=0.27.0"]
 
 [project.urls]
 Documentation = "https://github.com/chrissoria/cat-llm#readme"

[…]
