### 2026-03-07 b60a65f
Add single-label classification mode (multi_label=False)
New `multi_label` parameter (default True) switches prompts from
multi-label to single-label mode across text, PDF, and image inputs.
Only prompt text changes — JSON schema, parsing, validation, ensemble
consensus, and DataFrame output are all unchanged. Also adds formatter
fallback after CoVe step 4 output (pre-existing gap).

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
