+++ src/catllm/text_functions_ensemble.py
@@ -797,6 +797,7 @@ def build_text_classification_prompt(
     step_back_prompt: bool = False,
     stepback_insights: dict = None,
     model_name: str = None,
+    multi_label: bool = True,
 ) -> list:
     """
     Build the classification prompt for a text response.
@@ -818,31 +819,41 @@ def build_text_classification_prompt(
     Returns:
         List of message dicts for the LLM
     """
+    if multi_label:
+        categorize_instruction = 'into the following categories that apply'
+        json_instruction = 'Provide your answer in JSON format where the category number is the key and "1" if present, "0" if not.'
+        cot_step3 = 'assign 1 to matching categories and 0 to non-matching categories'
+    else:
+        categorize_instruction = 'into the single most appropriate category'
+        json_instruction = 'Provide your answer in JSON format where the category number is the key. Assign "1" to the single best matching category and "0" to all others.'
+        cot_step3 = 'assign 1 to the single best matching category and 0 to all others'
+
     if chain_of_thought:
         user_prompt = f"""{survey_question_context}
 
-Categorize this survey response "{response_text}" into the following categories that apply:
+Categorize this survey response "{response_text}" {categorize_instruction}:
 {categories_str}
 
 Let's think step by step:
 1. First, identify the main themes mentioned in the response
 2. Then, match each theme to the relevant categories
-3. Finally, assign 1 to matching categories and 0 to non-matching categories
+3. Finally, {cot_step3}
 
 {examples_text}
 
-Provide your answer in JSON format where the category number is the key and "1" if present, "0" if not."""
+{json_instruction}"""
     else:
         user_prompt = f"""{survey_question_context}
-Categorize this survey response "{response_text}" into the following categories that apply:
+Categorize this survey response "{response_text}" {categorize_instruction}:
 {categories_str}
 {examples_text}
-Provide your answer in JSON format where the category number is the key and "1" if present, "0" if not."""
+{json_instruction}"""
 
     # Add context prompt prefix if enabled
     if context_prompt:
-        context = """You are an expert researcher in survey data categorization.
-Apply multi-label classification and base decisions on explicit and implicit meanings.
+        label_type = "multi-label" if multi_label else "single-label"
+        context = f"""You are an expert researcher in survey data categorization.
+Apply {label_type} classification and base decisions on explicit and implicit meanings.
 When uncertain, prioritize precision over recall.
 
 """
@@ -1190,6 +1201,7 @@ def build_pdf_classification_prompt(
     stepback_insights: dict = None,
     model_name: str = None,
     example_json: str = None,
+    multi_label: bool = True,
 ) -> list:
     """
     Build the classification prompt for a PDF page.
@@ -1226,9 +1238,16 @@ def build_pdf_classification_prompt(
     else:  # image mode
         examine_instruction = "Examine the attached PDF page"
 
+    if multi_label:
+        task_instruction = f'{examine_instruction} and decide, **for each category below**, whether it is PRESENT (1) or NOT PRESENT (0).'
+        cot_step3 = 'assign 1 to matching categories and 0 to non-matching categories'
+    else:
+        task_instruction = f'{examine_instruction} and decide which **single category** below best describes this document page. Assign PRESENT (1) to only the best match and NOT PRESENT (0) to all others.'
+        cot_step3 = 'assign 1 to the single best matching category and 0 to all others'
+
     if chain_of_thought:
         base_text = f"""You are a document-tagging assistant.
-Task: {examine_instruction} and decide, **for each category below**, whether it is PRESENT (1) or NOT PRESENT (0).
+Task: {task_instruction}
 
 {f'Document page is expected to contain: {input_description}' if input_description else ''}
 
@@ -1238,14 +1257,14 @@ Categories:
 Let's analyze step by step:
[…]

[…]
