### 2026-03-07 8b3e646
Update README with new classify() parameters and install options
Add multi_label, embeddings, category_descriptions, json_formatter,
add_other, and check_verbosity params. Fix consensus_threshold default
(was 0.5, actually "unanimous"). Update batch_mode to reflect ensemble
support. Add [embeddings] and [formatter] install instructions.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
