### 2026-03-07 dd23938
Add Ollama support for explore() and extract() (v2.7.0)
Local Ollama models now get the same pre-flight validation as classify():
checks server is running, verifies model availability, auto-downloads
missing models, and warns about system resources. New auto_download
parameter on explore(), extract(), and explore_common_categories().

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
