+++ src/catllm/_embeddings.py
@@ -1,0 +1,189 @@
+"""
+Embedding-based similarity scores for CatLLM.
+
+Uses a local sentence-transformer model (BAAI/bge-small-en-v1.5, 33M params,
+~130MB) to compute cosine similarity between each input text and each category.
+Scores are independent per (text, category) pair — no softmax across categories,
+since this is multi-label classification.
+
+The embeddings feature is opt-in via embeddings=True on classify(). It adds
+`_similarity` columns alongside the existing binary 0/1 classification columns.
+
+Requires: pip install cat-llm[embeddings]
+"""
+
+import pandas as pd
+
+_EMBEDDING_MODEL_NAME = "BAAI/bge-small-en-v1.5"
+
+
+def _check_dependencies():
+    """Check that sentence-transformers is installed."""
+    try:
+        import sentence_transformers  # noqa: F401
+    except ImportError:
+        raise ImportError(
+            "The embeddings feature requires sentence-transformers.\n"
+            "Install with: pip install cat-llm[embeddings]\n"
+            "  (requires: sentence-transformers, which pulls in torch and transformers)"
+        )
+
+
+def _is_model_cached() -> bool:
+    """Check if the embedding model is already in the HuggingFace cache."""
+    try:
+        from huggingface_hub import try_to_load_from_cache
+        result = try_to_load_from_cache(_EMBEDDING_MODEL_NAME, "config.json")
+        return result is not None and not isinstance(result, type(None))
+    except Exception:
+        return False
+
+
+def ensure_embeddings_available() -> bool:
+    """
+    Ensure the embedding model is available, prompting to download if needed.
+
+    Returns:
+        True if the model is ready to use, False if user declined download.
+    """
+    _check_dependencies()
+
+    if _is_model_cached():
+        return True
+
+    print(
+        "\n[CatLLM] The embedding model (~130MB) will be downloaded from\n"
+        f"  HuggingFace Hub ({_EMBEDDING_MODEL_NAME}).\n"
+        "  This is a one-time download — the model is cached locally after."
+    )
+    try:
+        answer = input("  Continue? (Y/n): ").strip().lower()
+    except (EOFError, KeyboardInterrupt):
+        answer = "n"
+
+    if answer in ("", "y", "yes"):
+        return True
+    else:
+        print("  -> Embedding scores disabled for this run.\n")
+        return False
+
+
+def load_embedding_model():
+    """
+    Load and return the sentence-transformer embedding model.
+
+    Returns:
+        SentenceTransformer model instance.
+    """
+    _check_dependencies()
+
+    from sentence_transformers import SentenceTransformer
+
+    print(f"[CatLLM] Loading embedding model ({_EMBEDDING_MODEL_NAME})...")
+    model = SentenceTransformer(_EMBEDDING_MODEL_NAME)
+    print("[CatLLM] Embedding model ready.")
+    return model
+
+
+def compute_embedding_scores(texts, categories, model, category_descriptions=None):
+    """
+    Compute cosine similarity scores between texts and categories.
+
+    Each (text, category) score is independent — no softmax across categories.
+    Raw cosine similarity is rescaled from [-1, 1] to [0, 1] via (sim + 1) / 2.
+
+    Args:
+        texts: List of input text strings.
+        categories: List of category name strings.
+        model: Loaded SentenceTransformer model.
+        category_descriptions: Optional dict mapping category names to richer
+            descriptions for embedding (e.g., {"Past_Support": "References to
+            help received from family in the past"}).
+
+    Returns:
+        Dict mapping "category_N_similarity" -> list of float scores, where N
+        is 1-indexed to match the existing classification column naming.
+    """
+    from sentence_transformers import util
+
+    # Convert NaN/None to empty string
[…]
