### 2026-03-07 ce000d2
Remove CI publish workflow (publish manually via twine)
Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
