### 2026-03-09 cf3e23d
Add sequential ensemble mode (parallel parameter)
Auto-detects local models (Ollama) and runs them sequentially by default.
Cloud providers still run in parallel. Users can override with parallel=True/False.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
