### 2026-03-11 9985411
Add categories_per_call parameter for chunked classification (v2.8.0)
Split large category lists into smaller per-call chunks to reduce prompt
complexity and improve accuracy. Each chunk gets a temporary "Other"
catch-all for escape-hatch accuracy; a unified "Other" column is added
when all real categories are 0 but at least one chunk flagged Other.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
