### 2026-03-11 4fd3e4d
Add /catllm:classify-cc skill for Claude Code native classification (v2.8.1)
New slash command lets users classify survey data using Claude Code itself
as the LLM — no API key needed. Supports natural-language input, auto-
discovers categories from data samples, and outputs the same DataFrame
format as the API pipeline. Also adds "Claude Code" as a 3rd model option
in /catllm:classify with clear labeling (empirically validated vs casual).

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
