### 2026-03-11 a85d2ee
Integrate Claude Code classification into /catllm:classify as Path B
Merge the standalone classify-cc.md into classify.md as an inline branch.
When users select "Claude Code (no API key)" in Step 2, the skill follows
Path B (Steps 3B-8B) instead of redirecting to a separate command. Also
adds Step 0 for natural-language argument parsing across all paths.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
