### 2026-03-11 0d07f42
Add claude-code provider backend and redesign /catllm:classify flow (v2.8.2)
- Add claude-code as a provider in _providers.py (shells out to `claude -p`)
- Add check_claude_cli_available() utility, re-exported via text_functions.py
- Skip API key validation for claude-code provider (like Ollama)
- Redesign /catllm:classify with conversational-first flow and smart API key
  auto-detection; hard-cap Path B (native Claude Code) at 200 rows

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
