### 2026-03-15 e6582cb
Add embedding centroid tiebreaker for ensemble consensus ties (v2.9.0)
Resolves true ties (e.g., 2-2 splits with majority vote) by building
per-category centroids from unanimously-agreed rows and comparing tied
texts via cosine similarity. Adds _resolved_by output columns. Fix bug
where resolved_by string columns were wiped by pd.to_numeric conversion.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
