+++ src/catllm/_tiebreaker.py
@@ -1,0 +1,277 @@
+"""
+Embedding-based centroid tiebreaker for ensemble classification.
+
+Resolves true ties in ensemble consensus by building per-category centroids
+from unanimously-agreed rows, then comparing tied texts to those centroids.
+
+This module is called after ensemble classification completes but before
+building output DataFrames. It mutates all_results in place, updating
+consensus values for tied rows and adding tiebreaker_resolved metadata.
+
+How it works:
+1. Identify "confident" rows — where every model unanimously agrees
+2. Embed all texts, compute mean embedding per category from confident rows
+3. For true ties only (positive_rate == threshold exactly), compare text
+   embedding to positive vs negative centroid
+4. Resolve: pick whichever centroid is closer; if only positive centroid
+   exists, use absolute similarity threshold
+
+Requires: pip install cat-llm[embeddings]
+"""
+
+import numpy as np
+import pandas as pd
+
+
+def _compute_centroid(embeddings_matrix):
+    """
+    Compute L2-normalized centroid (mean embedding) from a matrix of embeddings.
+
+    Args:
+        embeddings_matrix: numpy array of shape (N, D) with N embeddings.
+
+    Returns:
+        L2-normalized centroid vector of shape (D,).
+    """
+    mean_vec = np.mean(embeddings_matrix, axis=0)
+    norm = np.linalg.norm(mean_vec)
+    if norm > 0:
+        mean_vec = mean_vec / norm
+    return mean_vec
+
+
+def _find_confident_and_tied_rows(all_results, category_key, threshold):
+    """
+    Bucket rows into confident-positive, confident-negative, and true-tied.
+
+    A row is "confident" when all successful models unanimously agree (all 1s
+    or all 0s). A row is "true-tied" when positive_rate == threshold exactly
+    (e.g., 2-2 split with majority vote threshold=0.5).
+
+    Args:
+        all_results: List of result dicts from classify_ensemble.
+        category_key: String key like "1", "2", etc.
+        threshold: Numeric consensus threshold (0-1).
+
+    Returns:
+        (confident_positive_indices, confident_negative_indices, tied_indices)
+        Each is a list of row indices into all_results.
+    """
+    confident_pos = []
+    confident_neg = []
+    tied = []
+
+    for row_idx, result in enumerate(all_results):
+        if result.get("skipped"):
+            continue
+
+        aggregated = result["aggregated"]
+        if aggregated.get("error"):
+            continue
+
+        per_model = aggregated.get("per_model", {})
+        if not per_model:
+            continue
+
+        # Collect votes for this category from all successful models
+        votes = []
+        for model_name, parsed in per_model.items():
+            vote_str = parsed.get(category_key, "0")
+            try:
+                votes.append(int(vote_str))
+            except (ValueError, TypeError):
+                votes.append(0)
+
+        if not votes:
+            continue
+
+        num_models = len(votes)
+        positive_count = sum(votes)
+        positive_rate = positive_count / num_models
+
+        # Check for true tie: positive_rate == threshold exactly
+        if abs(positive_rate - threshold) < 1e-9:
+            tied.append(row_idx)
+        elif positive_rate == 1.0:
+            # All models agree positive
+            confident_pos.append(row_idx)
+        elif positive_rate == 0.0:
+            # All models agree negative
+            confident_neg.append(row_idx)
+
+    return confident_pos, confident_neg, tied
+
+
+def resolve_ties_with_centroids(
+    all_results,
+    categories,
+    embedding_model,
+    consensus_threshold,
+    min_centroid_size=3,
+):
+    """
[…]

[…]
