### 2026-03-15 a0a52e6
Move test scripts from examples/ to tests/ (gitignored)
examples/ should only contain user-facing examples. Test scripts that
require live API keys are now in tests/ which is gitignored.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
