### 2026-03-15 4c5ddd4
Add summarize() robustness/batch parity, 5 example notebooks, remove logprobs (v2.10.0)
- Add safety saves, row_delay, fail_strategy, batch_mode to summarize()
- Add batch summarization functions in _batch.py (single + ensemble)
- Fix extract() bug: survey_question was ignored for image/PDF input types
- Remove experimental logprobs confidence feature (_confidence.py not included)
- Add example notebooks: summarization, Ollama, ensemble, explore(), extract()
- Update CHANGELOG with entries for v2.9.0 and v2.10.0

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
