+++ src/catllm/summarize.py
@@ -44,6 +44,17 @@ def summarize(
     save_directory: str = None,
     progress_callback=None,
     models: list = None,
+    # Robustness parameters
+    safety: bool = False,
+    max_retries: int = 5,
+    batch_retries: int = 2,
+    retry_delay: float = 1.0,
+    row_delay: float = 0.0,
+    fail_strategy: str = "partial",
+    # Batch mode parameters
+    batch_mode: bool = False,
+    batch_poll_interval: float = 30.0,
+    batch_timeout: float = 86400.0,
 ):
     """
     Summarize text or PDF data using LLMs.
@@ -75,6 +86,17 @@ def summarize(
         save_directory (str): Directory to save results
         progress_callback: Optional callback for progress updates
         models (list): For multi-model mode, list of (model, provider, api_key) tuples
+        safety (bool): If True, saves progress after each item. Requires filename.
+        max_retries (int): Max retries per API call. Default 5.
+        batch_retries (int): Max retries for batch-level failures. Default 2.
+        retry_delay (float): Delay between retries in seconds. Default 1.0.
+        row_delay (float): Delay in seconds between processing each row. Default 0.0.
+        fail_strategy (str): How to handle failures - "partial" (default) or "strict".
+        batch_mode (bool): If True, use async batch API (50% cost savings).
+            Supported providers: openai, anthropic, google, mistral, xai.
+            Not compatible with PDF input.
+        batch_poll_interval (float): Seconds between batch status checks. Default 30.
+        batch_timeout (float): Max seconds to wait for batch completion. Default 86400 (24h).
 
     Returns:
         pd.DataFrame: Results with summary column(s):
@@ -104,13 +126,23 @@ def summarize(
         ...     api_key="your-api-key"
         ... )
         >>>
-        >>> # PDF summarization with list of files
+        >>> # With safety saves and row delay
         >>> results = cat.summarize(
-        ...     input_data=["doc1.pdf", "doc2.pdf"],
-        ...     description="Financial reports",
-        ...     mode="both",
-        ...     focus="key metrics",
-        ...     api_key="your-api-key"
+        ...     input_data=df['responses'],
+        ...     description="Customer feedback",
+        ...     api_key="your-api-key",
+        ...     safety=True,
+        ...     filename="results.csv",
+        ...     row_delay=1.0,
+        ... )
+        >>>
+        >>> # Batch mode (50% cost savings)
+        >>> results = cat.summarize(
+        ...     input_data=df['responses'],
+        ...     description="Customer feedback",
+        ...     api_key="your-api-key",
+        ...     batch_mode=True,
+        ...     filename="batch_results.csv",
         ... )
         >>>
         >>> # Multi-model with synthesis
@@ -124,6 +156,92 @@ def summarize(
     """
     # Map mode to pdf_mode
     pdf_mode = mode if mode in ("image", "text", "both") else "image"
+
+    # =========================================================================
+    # Batch mode — bypass summarize_ensemble entirely
+    # =========================================================================
+    if batch_mode:
+        from ._batch import UNSUPPORTED_BATCH_PROVIDERS, run_batch_summarize
+        from .text_functions_ensemble import _detect_input_type, prepare_model_configs
+
+        # Guard: text input only
+        detected_type = _detect_input_type(input_data)
+        if detected_type == "pdf":
+            raise ValueError(
+                "batch_mode=True only supports text input, but detected input type is 'pdf'. "
+                "Set batch_mode=False for PDF summarization."
+            )
+
+        # Warn if progress_callback was provided (incompatible with batch)
+        if progress_callback is not None:
+            print(
+                "[CatLLM] WARNING: progress_callback is not available in batch_mode "
+                "(no per-item progress until the job completes). Ignoring callback."
+            )
+
[…]
