+++ src/catllm/_batch.py
@@ -555,18 +555,21 @@ def _parse_batch_results(
     raw_results: str,
     custom_id_map: dict,
     client: "UnifiedLLMClient",
+    parse_mode: str = "json",
 ) -> dict:
     """
-    Parse the downloaded JSONL results back into per-item classification strings.
+    Parse the downloaded JSONL results back into per-item strings.
 
     Args:
         provider: Provider name
         raw_results: Raw JSONL text from the batch results download
         custom_id_map: Dict mapping custom_id string → original item index
         client: UnifiedLLMClient instance (used to call _parse_response())
+        parse_mode: "json" (default) runs extract_json() on responses,
+            "text" returns raw text as-is (for summarization)
 
     Returns:
-        Dict mapping item_index → (json_str, error_or_None)
+        Dict mapping item_index → (result_str, error_or_None)
         Missing items (job dropped them) get (None, "Missing from batch results").
     """
     parsed_results = {}
@@ -650,8 +653,11 @@ def _parse_batch_results(
         if idx is None:
             continue
 
-        json_str = extract_json(raw_text)
-        parsed_results[idx] = (json_str, None)
+        if parse_mode == "text":
+            parsed_results[idx] = (raw_text, None)
+        else:
+            json_str = extract_json(raw_text)
+            parsed_results[idx] = (json_str, None)
 
     return parsed_results
 
@@ -1015,3 +1021,368 @@ def run_batch_ensemble_classify(
         })
 
     return build_output_dataframes(all_results, model_configs, categories, filename, save_directory)
+
+
+# =============================================================================
+# Batch summarization
+# =============================================================================
+
+def _run_one_batch_summarize_job(
+    cfg: dict,
+    items: list,
+    prompt_params: dict,
+    batch_poll_interval: float = 30.0,
+    batch_timeout: float = 86400.0,
+) -> dict:
+    """
+    Submit, poll, download, and parse a batch summarization job for one model.
+    Returns {item_index: (summary_text_or_None, error_or_None)}.
+    """
+    from .text_functions_ensemble import build_text_summarization_prompt, build_summary_json_schema
+
+    provider = cfg["provider"]
+    api_key = cfg["api_key"]
+    model = cfg["model"]
+
+    input_description = prompt_params.get("input_description", "")
+    summary_instructions = prompt_params.get("summary_instructions", "")
+    max_length = prompt_params.get("max_length")
+    focus = prompt_params.get("focus")
+    chain_of_thought = prompt_params.get("chain_of_thought", False)
+    context_prompt = prompt_params.get("context_prompt", False)
+    step_back_prompt = prompt_params.get("step_back_prompt", False)
+    stepback_insights = prompt_params.get("stepback_insights", {})
+    creativity = prompt_params.get("creativity")
+
+    include_additional = provider != "google"
+    json_schema = build_summary_json_schema(include_additional)
+
+    client = UnifiedLLMClient(provider=provider, api_key=api_key, model=model)
+
+    print(f"\n[batch] Building {len(items)} summarization request(s) for {model} ({provider})...")
+
+    custom_id_map = {}
+    jsonl_lines = []
+    requests_list = []
+
+    for idx, item in enumerate(items):
+        custom_id = f"item-{idx}"
+        custom_id_map[custom_id] = idx
+
+        text = str(item) if item is not None else ""
+
+        messages = build_text_summarization_prompt(
+            response_text=text,
+            input_description=input_description,
+            summary_instructions=summary_instructions,
+            max_length=max_length,
+            focus=focus,
+            chain_of_thought=chain_of_thought,
[…]
