### 2026-03-15 98c7c02
Add summarize/classify parity, code hygiene, README docs completeness
- Add pdf_dpi, thinking_budget, parallel, auto_download, max_workers to
  summarize() and summarize_ensemble() for feature parity with classify()
- Remove 4 stale TODO comments from classify.py and _category_analysis.py
- Delete dead code main.py (never imported, contained stale duplicates)
- Document 14 missing parameters across all 4 functions in README

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
