+++ src/catllm/main.py
@@ -1,296 +1,0 @@
-"""
-Unified functions for classification and category extraction.
-"""
-
-__all__ = ["extract", "summarize"]
-
-
-def extract(
-    input_data,
-    api_key,
-    input_type="text",
-    description="",
-    max_categories=12,
-    categories_per_chunk=10,
-    divisions=12,
-    user_model="gpt-4o",
-    creativity=None,
-    specificity="broad",
-    research_question=None,
-    mode="text",
-    filename=None,
-    model_source="auto",
-    iterations=8,
-    random_state=None,
-    focus=None,
-):
-    """
-    Unified category extraction function for text, image, and PDF inputs.
-
-    This function dispatches to the appropriate specialized explore function
-    based on the `input_type` parameter, providing a single entry point for
-    discovering categories in your data.
-
-    Args:
-        input_data: The data to explore. Can be:
-            - For text: list of text responses or pandas Series
-            - For image: directory path, single file, or list of image paths
-            - For pdf: directory path, single file, or list of PDF paths
-        api_key (str): API key for the model provider.
-        input_type (str): Type of input data. Options:
-            - "text" (default): Text/survey responses
-            - "image": Image files
-            - "pdf": PDF documents
-        description (str): Description of the input data. Used as:
-            - survey_question for text
-            - image_description for images
-            - pdf_description for PDFs
-        max_categories (int): Maximum number of final categories to return.
-        categories_per_chunk (int): Categories to extract per chunk.
-        divisions (int): Number of chunks to divide data into.
-        user_model (str): Model name to use. Default "gpt-4o".
-        creativity (float): Temperature setting. None uses model default.
-        specificity (str): "broad" or "specific" category granularity.
-        research_question (str): Optional research context.
-        mode (str): Processing mode:
-            - For text: Not used
-            - For image: "image" (default) or "both"
-            - For pdf: "text" (default), "image", or "both"
-        filename (str): Optional CSV filename to save results.
-        model_source (str): Provider - "auto", "openai", "anthropic", "google",
-            "mistral", "huggingface", "xai".
-        iterations (int): Number of passes over the data.
-        random_state (int): Random seed for reproducibility.
-        focus (str): Optional focus instruction for category extraction (e.g.,
-            "decisions to move", "emotional responses"). When provided, the model
-            will prioritize extracting categories related to this focus.
-
-    Returns:
-        dict with keys:
-            - counts_df: DataFrame of categories with counts
-            - top_categories: List of top category names
-            - raw_top_text: Raw model output from final merge step
-
-    Examples:
-        >>> import catllm as cat
-        >>>
-        >>> # Extract categories from survey responses
-        >>> results = cat.extract(
-        ...     input_data=df['responses'],
-        ...     description="Why did you move?",
-        ...     api_key="your-api-key"
-        ... )
-        >>> print(results['top_categories'])
-        >>>
-        >>> # Extract categories from images
-        >>> results = cat.extract(
-        ...     input_data="/path/to/images/",
-        ...     description="Product photos",
-        ...     input_type="image",
-        ...     api_key="your-api-key"
-        ... )
-        >>>
-        >>> # Extract categories from PDFs
-        >>> results = cat.extract(
-        ...     input_data="/path/to/pdfs/",
-        ...     description="Research papers",
-        ...     input_type="pdf",
-        ...     mode="text",
-        ...     api_key="your-api-key"
-        ... )
-    """
-    input_type = input_type.lower().rstrip('s')  # Normalize: "texts" -> "text", "images" -> "image", "pdfs" -> "pdf"
-
-    if input_type == "text":
-        from .text_functions import explore_common_categories
[…]
