### 2026-03-16 c378398
Update master plan: mark Phases 1-3 complete, add domain-suffixed API design
Phases 1-3 (stabilize, create cat-stack, wire domain packages) are done.
Phase 4 now describes the meta-package approach: cat-llm re-exports
domain-suffixed aliases (classify_survey, classify_social, classify_academic,
cerad_drawn_score) from sub-packages, with neutral classify/extract/explore/
summarize pointing to cat-stack.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
