+++ MASTERPLAN.md
@@ -14,30 +14,32 @@ with domain-tuned prompts, metadata injection, and workflow helpers.
 ```
 pip install cat-llm
 # brings in:
-#   cat-stack   general-purpose text column classification (the base)
+#   cat-stack     general-purpose classification engine (the base)
 #   cat-survey    survey response classification & extraction
 #   cat-vader     social media (Reddit, Twitter/X, forums)
 #   cat-ademic    academic papers, PDFs, citations
 #   cat-pol       political text (manifestos, speeches, legislation)
+#   cat-cog       cognitive assessment & visual scoring (CERAD, drawing tests)
 ```
 
 **Dependency graph** (modeled on tidyverse / rlang):
 
 ```
-cat-stack                    ← general base + shared infra (like rlang)
-                                  independently useful; no domain assumptions
+cat-stack                                      ← general base + shared infra (like rlang)
+                                                    independently useful; no domain assumptions
     ↑
-cat-survey  cat-vader  cat-ademic  cat-pol    ← domain packages, each depends on cat-stack
-    ↑           ↑          ↑          ↑
-                    cat-llm                    ← pure meta-package, keeps the brand
-                                               (depends on all domain packages + cat-stack)
+cat-survey  cat-vader  cat-ademic  cat-pol  cat-cog  ← domain packages, each depends on cat-stack
+    ↑           ↑          ↑          ↑        ↑
+                         cat-llm                     ← pure meta-package, keeps the brand
+                                                      (depends on all domain packages + cat-stack)
 ```
 
 Every package is independently installable:
 ```
-pip install cat-stack   # general text classification, no domain framing
-pip install cat-survey    # survey-specific, pulls in cat-stack automatically
-pip install cat-llm       # everything
+pip install cat-stack    # general classification engine, no domain framing
+pip install cat-survey   # survey-specific, pulls in cat-stack automatically
+pip install cat-cog      # cognitive assessment, pulls in cat-stack automatically
+pip install cat-llm      # everything
 ```
 
 ---
@@ -51,11 +53,13 @@ infrastructure layer AND independently useful for researchers who just have a te
 column and don't fit neatly into a domain.
 
 **Scope:**
-- `classify()`, `extract()`, `explore()`, `summarize()` for any text input
+- `classify()`, `extract()`, `explore()`, `summarize()` for text, image, and PDF input
 - `UnifiedLLMClient`, `PROVIDER_CONFIG`, `detect_provider` — all provider infrastructure
 - `_batch.py` — async batch API logic
+- Image and PDF processing: `image_functions.py`, `pdf_functions.py`
 - Shared text utilities: `build_json_schema`, `extract_json`, `validate_classification_json`
-- No survey framing, no social media framing, no PDF-first assumptions
+- No survey framing, no social media framing, no cognitive-test assumptions
+- Namespace: `import cat_stack`
 
 **Status:** Does not exist yet. Code lives inside `cat-llm/src/catllm/` and needs to be
 extracted. This is the prerequisite for all other phases.
@@ -92,18 +96,34 @@ Social media text classification with platform-aware context injection.
 
 ---
 
-### `cat-ademic` *(planned)*
+### `cat-ademic` *(already separate)*
 Classification and extraction for academic and long-form documents.
 
 **Scope:**
-- PDF-first `classify()` / `extract()` — builds on current `pdf_functions.py` logic
[…]
