### 2026-03-19 4b88acb
Slim cat-llm to meta-package, update README for ecosystem (v3.0.0)
cat-llm is now a pure meta-package that pulls in domain sub-packages
(cat-stack, cat-survey, cat-vader, cat-ademic, cat-pol, cat-cog, cat-web).
All source modules removed — __init__.py re-exports from sub-packages.
README rewritten with ecosystem table, dependency diagram, domain package
examples, and dual Quick Start (meta-package vs direct install).

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
