+++ src/catllm/_providers.py
@@ -1,1048 +1,0 @@
-"""
-Unified LLM provider infrastructure for CatLLM.
-
-This module provides a unified HTTP-based approach for calling multiple LLM providers
-(OpenAI, Anthropic, Google, Mistral, xAI, Perplexity, HuggingFace, and Ollama)
-without requiring provider-specific SDKs.
-"""
-
-import json
-import time
-import requests
-
-__all__ = [
-    # Main client
-    "UnifiedLLMClient",
-    "PROVIDER_CONFIG",
-    # Provider detection
-    "detect_provider",
-    "_detect_model_source",
-    "_detect_huggingface_endpoint",
-    # Ollama utilities
-    "set_ollama_endpoint",
-    "check_ollama_running",
-    "list_ollama_models",
-    "check_ollama_model",
-    "check_system_resources",
-    "get_ollama_model_size_estimate",
-    "pull_ollama_model",
-    "OLLAMA_MODEL_SIZES",
-    # Claude Code utilities
-    "check_claude_cli_available",
-]
-
-
-# =============================================================================
-# HuggingFace Endpoint Auto-Detection
-# =============================================================================
-
-def _detect_huggingface_endpoint(api_key: str, model: str) -> str:
-    """
-    Test which HuggingFace endpoint works for this model.
-    Tries generic router first, then Together.
-
-    Args:
-        api_key: HuggingFace API key
-        model: Model name to test
-
-    Returns:
-        Base URL for the working endpoint (without /chat/completions)
-    """
-    endpoints = [
-        "https://router.huggingface.co/v1/chat/completions",
-        "https://router.huggingface.co/together/v1/chat/completions",
-    ]
-
-    headers = {
-        "Content-Type": "application/json",
-        "Authorization": f"Bearer {api_key}"
-    }
-
-    payload = {
-        "model": model,
-        "messages": [{"role": "user", "content": "hi"}],
-        "max_tokens": 5
-    }
-
-    for endpoint in endpoints:
-        try:
-            response = requests.post(endpoint, headers=headers, json=payload, timeout=30)
-            if response.status_code == 200:
-                # Return the base URL (without /chat/completions)
-                return endpoint.replace("/chat/completions", "")
-        except Exception:
-            continue
-
-    # Default to generic (will fail with informative error)
-    return "https://router.huggingface.co/v1"
-
-
-# =============================================================================
-# Provider Configuration
-# =============================================================================
-
-PROVIDER_CONFIG = {
-    "openai": {
-        "endpoint": "https://api.openai.com/v1/chat/completions",
-        "auth_header": "Authorization",
-        "auth_prefix": "Bearer ",
-    },
-    "anthropic": {
-        "endpoint": "https://api.anthropic.com/v1/messages",
-        "auth_header": "x-api-key",
-        "auth_prefix": "",
-    },
-    "google": {
-        "endpoint": "https://generativelanguage.googleapis.com/v1beta/models/{model}:generateContent",
-        "auth_header": "x-goog-api-key",
-        "auth_prefix": "",
-    },
-    "mistral": {
-        "endpoint": "https://api.mistral.ai/v1/chat/completions",
-        "auth_header": "Authorization",
-        "auth_prefix": "Bearer ",
-    },
-    "perplexity": {
-        "endpoint": "https://api.perplexity.ai/chat/completions",
-        "auth_header": "Authorization",
-        "auth_prefix": "Bearer ",
-    },
-    "xai": {
-        "endpoint": "https://api.x.ai/v1/chat/completions",
-        "auth_header": "Authorization",
-        "auth_prefix": "Bearer ",
-    },
-    "huggingface": {
[…]
