+++ src/catllm/__about__.py
@@ -1,10 +1,10 @@
 # SPDX-FileCopyrightText: 2025-present Christopher Soria <chrissoria@berkeley.edu>
 #
 # SPDX-License-Identifier: MIT
-__version__ = "2.10.0"
+__version__ = "3.0.0"
 __author__ = "Chris Soria"
 __email__ = "chrissoria@berkeley.edu"
 __title__ = "cat-llm"
-__description__ = "An AI tool for categorizing and exploring text data and images using LLMs."
+__description__ = "The CatLLM ecosystem — LLM-powered classification for text, images, and PDFs across domains."
 __url__ = "https://github.com/chrissoria/cat-llm"
 __license__ = "GPL-3.0"

[…]
