### 2026-03-23 6809b4c
Update README: add prompt_tune(), cat-pol sources, summarize formats
- Add prompt_tune() to core functions list and TOC
- Update cat-pol section with source= parameter and 15 built-in datasets
- Add summarize format options (paragraph, bullets, one-liner, structured, report)
- Add prompt optimization to meta-package example

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
