### 2026-03-27 8e26585
Add unified Streamlit app with all 7 domains and 5 functions
Merges the separate HF classifier and summarizer apps into a single
modular app with sidebar domain/function routing. Supports General,
Survey, Social Media, Academic, Policy, Web, and Cognitive domains
across Classify, Extract, Explore, Summarize, and CERAD Score
functions. Includes free tier (Qwen 2.5 72B via HuggingFace) and
paid model support with per-provider API key management.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
