### 2026-03-27 b758d05
Update hf_space with unified app for HuggingFace Spaces deployment
Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
