+++ hf_space/app.py
@@ -1,2310 +1,177 @@
 """
-Streamlit app - CatLLM Survey Response Classifier
-Migrated from Gradio for better mobile support
+CatLLM Unified App — HuggingFace Spaces entry point.
 """
 
-import streamlit as st
-import pandas as pd
-import tempfile
-import os
-import time
 import sys
-from datetime import datetime
-import matplotlib.pyplot as plt
+import os
 
-# Import catllm
-try:
-    import catllm
-    CATLLM_AVAILABLE = True
-except ImportError as e:
-    print(f"Warning: Could not import catllm: {e}")
-    CATLLM_AVAILABLE = False
+# Add this directory to path so imports work
+sys.path.insert(0, os.path.dirname(__file__))
 
-MAX_CATEGORIES = 10
-INITIAL_CATEGORIES = 3
-MAX_FILE_SIZE_MB = 100
+import streamlit as st
 
-def count_pdf_pages(pdf_path):
-    """Count the number of pages in a PDF file."""
-    try:
-        import fitz  # PyMuPDF
-        doc = fitz.open(pdf_path)
-        page_count = len(doc)
-        doc.close()
-        return page_count
-    except Exception:
-        return 1  # Default to 1 if can't read
-
-
-def extract_text_from_pdfs(pdf_paths):
-    """Extract text from all pages of all PDFs, returning list of page texts."""
-    import fitz  # PyMuPDF
-    all_texts = []
-    for pdf_path in pdf_paths:
-        try:
-            doc = fitz.open(pdf_path)
-            for page in doc:
-                text = page.get_text().strip()
-                if text:  # Only add non-empty pages
-                    all_texts.append(text)
-            doc.close()
-        except Exception as e:
-            print(f"Error extracting text from {pdf_path}: {e}")
-    return all_texts
-
-
-def extract_pdf_pages(pdf_paths, pdf_name_map, mode="image"):
-    """
-    Extract individual pages from PDFs.
-    Returns list of (page_data, page_label) tuples.
-    For image mode: page_data is path to temp image file
-    For text mode: page_data is extracted text
-    """
-    import fitz  # PyMuPDF
-    pages = []
-
-    for pdf_path in pdf_paths:
-        orig_name = pdf_name_map.get(pdf_path, os.path.basename(pdf_path).replace('.pdf', ''))
-        try:
-            doc = fitz.open(pdf_path)
-            for page_num, page in enumerate(doc, 1):
-                page_label = f"{orig_name}_p{page_num}"
-
-                if mode == "text":
-                    # Extract text
-                    text = page.get_text().strip()
-                    if text:
-                        pages.append((text, page_label, "text"))
-                else:
-                    # Render as image (for image or both mode)
-                    pix = page.get_pixmap(matrix=fitz.Matrix(2, 2))  # 2x zoom for better quality
-                    img_path = tempfile.NamedTemporaryFile(delete=False, suffix='.png').name
-                    pix.save(img_path)
-
-                    if mode == "both":
-                        text = page.get_text().strip()
-                        pages.append((img_path, page_label, "image", text))
-                    else:
-                        pages.append((img_path, page_label, "image"))
-            doc.close()
-        except Exception as e:
-            print(f"Error extracting pages from {pdf_path}: {e}")
-
-    return pages
-
-# Free models - display name -> actual API model name
-FREE_MODELS_MAP = {
-    "GPT-4o Mini": "gpt-4o-mini",
-    "Gemini 2.5 Flash": "gemini-2.5-flash",
-    "Claude 3 Haiku": "claude-3-haiku-20240307",
-    "Llama 3.3 70B": "meta-llama/Llama-3.3-70B-Instruct:groq",
[…]
