### 2026-03-28 e29d2f5
Add missing regex dependency to requirements.txt
Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
