### 2026-03-30 18b524b
Add desktop app, settings, history, R ecosystem, Stata domain commands
Desktop app (macOS):
- PyInstaller bundle with pywebview native window (no browser tabs)
- .dmg installer with custom icon (305MB compressed)
- Launcher with Ollama port detection and clean shutdown

App features:
- Settings dialog (API Keys, Platform Credentials, Defaults, Output, Local Models, About)
- API key persistence (~Library/Application Support/CatLLM/)
- Platform credential storage (Threads, Reddit, Bluesky, Mastodon, YouTube, OpenAlex)
- Run history with auto-save and sidebar browser
- Ollama model detection with
[…]
