+++ r-package/cat.stack/R/classify.R
@@ -1,0 +1,208 @@
+#' Classify text, images, or PDFs using LLMs
+#'
+#' Wraps the Python `cat_stack.classify()` function. Supports both single-model
+#' and multi-model (ensemble) classification.
+#'
+#' @param input_data A character vector, list of text strings, or
+#'   `data.frame` column containing the items to classify. For image or PDF
+#'   classification, a directory path or character vector of file paths.
+#' @param categories A character vector of category names, or `"auto"` to
+#'   infer categories from the data (requires `survey_question`).
+#' @param api_key API key for the model provider (single-model mode).
+#'   Not required when `models` is supplied.
+#' @param description Character. Context description for the classification
+#'   task (e.g., the survey question or image subject). Default `""`.
+#' @param user_model Character. Model name to use in single-model mode.
+#'   Default `"gpt-4o"`.
+#' @param mode Character. PDF processing mode: `"image"` (default), `"text"`,
+#'   or `"both"`.
+#' @param creativity Numeric or `NULL`. Temperature setting (0-2). `NULL` uses
+#'   the provider default. Default `NULL`.
+#' @param safety Logical. If `TRUE`, saves progress after each item. Default
+#'   `FALSE`.
+#' @param chain_of_verification Logical. Enable Chain of Verification.
+#'   Empirically degrades accuracy -- provided for research only. Default
+#'   `FALSE`.
+#' @param chain_of_thought Logical. Enable chain-of-thought reasoning. Default
+#'   `FALSE`.
+#' @param step_back_prompt Logical. Enable step-back prompting. Default
+#'   `FALSE`.
+#' @param context_prompt Logical. Add expert context to prompts. Default
+#'   `FALSE`.
+#' @param thinking_budget Integer. Extended thinking token budget (0 = off).
+#'   Default `0L`.
+#' @param example1,example2,example3,example4,example5,example6 Optional
+#'   few-shot example strings. Empirically degrades accuracy -- provided for
+#'   research only.
+#' @param filename Character or `NULL`. Output CSV filename. Default `NULL`.
+#' @param save_directory Character or `NULL`. Directory to save results.
+#'   Default `NULL`.
+#' @param model_source Character. Provider hint for single-model mode:
+#'   `"auto"`, `"openai"`, `"anthropic"`, `"google"`, `"mistral"`,
+#'   `"huggingface"`, `"xai"`. Default `"auto"`.
+#' @param max_categories Integer. Maximum number of categories when
+#'   `categories = "auto"`. Default `12L`.
+#' @param categories_per_chunk Integer. Categories extracted per chunk when
+#'   `categories = "auto"`. Default `10L`.
+#' @param divisions Integer. Number of data chunks when `categories = "auto"`.
+#'   Default `10L`.
+#' @param research_question Character or `NULL`. Optional research context.
+#'   Default `NULL`.
+#' @param models A list of model specifications for multi-model ensemble mode.
+#'   Each element is either a 3-element character vector
+#'   `c("model", "provider", "api_key")` or a 4-element list
+#'   `list("model", "provider", "api_key", list(creativity = 0.5))`.
+#'   When `models` is supplied, `api_key` and `user_model` are ignored.
+#' @param consensus_threshold Character or numeric. Agreement threshold for
+#'   ensemble mode. Options: `"unanimous"` (default, 100%), `"majority"`
+#'   (50%), `"two-thirds"` (67%), or a numeric value between 0 and 1.
+#' @param survey_question Character. The survey question text (used when
+#'   `categories = "auto"`). Default `""`.
+#' @param use_json_schema Logical. Use JSON schema for structured output.
+#'   Default `TRUE`.
+#' @param max_workers Integer or `NULL`. Max parallel workers. `NULL` = auto.
[…]
