+++ r-package/cat.stack/tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(cat.stack)
+
+test_check("cat.stack")
