+++ app/components/history.py
@@ -1,0 +1,208 @@
+"""
+Run history — auto-saves results and provides a history browser.
+Stores metadata in ~/Library/Application Support/CatLLM/history.json
+and result files in a results/ subdirectory.
+"""
+
+import json
+import os
+import platform
+import shutil
+from datetime import datetime
+
+import pandas as pd
+
+
+def _storage_dir():
+    if platform.system() == "Darwin":
+        return os.path.join(os.path.expanduser("~"), "Library", "Application Support", "CatLLM")
+    return os.path.join(os.path.expanduser("~"), ".catllm")
+
+
+def _results_dir():
+    d = os.path.join(_storage_dir(), "results")
+    os.makedirs(d, exist_ok=True)
+    return d
+
+
+def _history_path():
+    return os.path.join(_storage_dir(), "history.json")
+
+
+def _load_history():
+    path = _history_path()
+    if not os.path.isfile(path):
+        return []
+    try:
+        with open(path, "r") as f:
+            return json.load(f)
+    except Exception:
+        return []
+
+
+def _save_history(entries):
+    d = _storage_dir()
+    os.makedirs(d, exist_ok=True)
+    with open(_history_path(), "w") as f:
+        json.dump(entries, f, indent=2)
+
+
+def save_run(results_dict, settings=None):
+    """Auto-save a run's results to disk and add to history.
+
+    Args:
+        results_dict: The st.session_state.results dict from a function page.
+        settings: The current settings dict (to check auto_save_results).
+    """
+    if settings and not settings.get("auto_save_results", True):
+        return
+
+    task_type = results_dict.get("task_type", "unknown")
+    timestamp = datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
+    filename = f"catllm_{task_type}_{timestamp}.csv"
+    filepath = os.path.join(_results_dir(), filename)
+
+    # Save the dataframe if present
+    df = results_dict.get("df")
+    row_count = 0
+    if df is not None and isinstance(df, pd.DataFrame):
+        df.to_csv(filepath, index=False)
+        row_count = len(df)
+    elif task_type == "extract":
+        # Extract results have counts_df
+        counts_df = results_dict.get("counts_df")
+        if counts_df is not None and isinstance(counts_df, pd.DataFrame):
+            counts_df.to_csv(filepath, index=False)
+            row_count = len(counts_df)
+        else:
+            # Save categories as a simple CSV
+            cats = results_dict.get("categories", [])
+            if cats:
+                pd.DataFrame({"category": cats}).to_csv(filepath, index=False)
+                row_count = len(cats)
+            else:
+                return
+    elif task_type == "explore":
+        cats = results_dict.get("categories", [])
+        if cats:
+            pd.DataFrame({"category": cats}).to_csv(filepath, index=False)
+            row_count = len(cats)
+        else:
+            return
+    else:
+        return
+
+    # Build models string
+    models = results_dict.get("models_list", [])
+    model_str = ", ".join(models) if models else "unknown"
+
+    # Add to history
+    entry = {
+        "timestamp": timestamp,
+        "task_type": task_type,
+        "model": model_str,
+        "row_count": row_count,
+        "categories": results_dict.get("categories", []),
+        "filename": filename,
+        "filepath": filepath,
+        "status": results_dict.get("status", ""),
+    }
+
+    history = _load_history()
+    history.insert(0, entry)
+
+    # Keep last 100 entries
+    if len(history) > 100:
+        # Delete old files
+        for old in history[100:]:
+            try:
+                os.remove(old["filepath"])
+            except OSError:
+                pass
+        history = history[:100]
+
+    _save_history(history)
+
+
+def get_history():
+    """Return list of history entries (newest first)."""
+    return _load_history()
+
+
+def load_run(entry):
[…]
