+++ r-package/cat.vader/tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(cat.vader)
+
+test_check("cat.vader")
