+++ r-package/cat.cog/tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(cat.cog)
+
+test_check("cat.cog")

[…]
