### 2026-04-02 a10a5ea
Add Dependencies section to README
Lists core and optional dependencies with purposes and install commands.
Addresses reviewer question about clearly-stated dependency management.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
