### 2026-04-02 0d05e83
Add Testing section to README with local Ollama example
Describes how to verify functionality locally using Ollama (free,
no API keys) and links to the example notebook.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
