### 2026-04-12 #9678 @Jiang22s
@ujjwalkarn @chrissoria
Sorry for the late reply. I've been overseas in an area with unreliable VPN access, which made it tricky to get to the testing. Thanks for bearing with me!

I've had a chance to go through the functionality and things look great overall. I think we are good to go. Just have a few small observations that might be outside of the scope of the review and I will just comment here.

HuggingFace report output: The plots and category matrices features works correctly and capture responses really well. There is a line in the first section "Each category column contains binary values: 1 = present, 0 = not present" but the exported report only shows category names in that section and they probably understand it better while scrolling to the later sections for those who are not familiar with vision classification models. It may be worth adding a brief note pointing users to the relevant sections later in the report. Please feel free to correct me if I misread this.

Image feature scope: Though I think the flexibility of user customization is great and the current hint is great, it might be helpful adding a brief note or example clarifying the scope of image classification. For instance, whether the feature is intended for classifying among photos (e.g., categorizing a set of images) versus categorizing elements within a single photo could help  users get the most out of it. It might also be worth considering a failure mode.

Minor points overall and do not affect the review on the list. Happy to clarify if helpful :) Solid package, enjoyed going through it.
