### 2026-04-20 #9678 @ujjwalkarn
Thanks @Jiang22s! Just to confirm—would you be comfortable proceeding with the publication if your points are addressed?

@chrissoria - do you have any thoughts on the points raised above?
