### 2026-04-23 9e2042f
Consolidate hf_space to single-file app with reviewer fixes
- Replace modular component/domain/function structure with unified app.py
- Bump cat-llm requirement to >=2.5.0
- Add forward reference in PDF report Category Mapping section pointing to Classification Summary and visualizations pages
- Add image scope and failure modes info box in Images tab

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
