### 2026-05-01 f6dc5b0
Add Zenodo and Citation metadata for v3.0.0
Adds .zenodo.json describing cat-llm as a metapackage bundling cat-stack and
six domain sub-packages, and CITATION.cff referencing the existing concept
DOI 10.5281/zenodo.15620202. Also gitignores the zenodo-upload-*/ staging
folder used to assemble release artifacts.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
