### 2026-05-01 8243562
Add ORCID to Zenodo and Citation metadata
Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
