### 2026-05-01 83a67fe
Fix concept DOI in CITATION.cff
10.5281/zenodo.15620202 was the v0.1.1 version DOI; the concept DOI that
should be cited regardless of version is 10.5281/zenodo.15532316.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
