### 2026-05-01 #9678 @chrissoria
@ujjwalkarn
1. I have verified the author affiliation and ORCID
2. The latest version of the package is up (version 3.0.0)
3. The zenodo URL: https://zenodo.org/records/19960067 and DOI: 10.5281/zenodo.15532316
4. Authors match paper
5. license is the same in the zenodo as in github (the paper doesn't have the license)
