### 2026-05-01 06b2b7c
Update title in Zenodo and Citation metadata
Reframes from "Pipeline for Classifying" to "Ecosystem for Generating,
Assigning, and Scoring" to match the paper's three-verb framing and the
v3.0.0 metapackage scope. "Documents" replaces "PDFs" for broader coverage.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
