### 2026-05-01 bb093fd
Fix typo in paper.md acknowledgements
"this researPartial" → "this research. Partial"

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
