### 2026-05-01 cb24656
Acknowledge Bashir Ahmed and the Bashir Ahmed Fellowship
Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
