### 2026-05-01 0bc2d2e
Trim Bashir Ahmed acknowledgement to fellowship only
Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
