### 2026-05-10 #9678 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
